% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{abs}
\alias{abs}
\alias{abs.fastai.torch_core.TensorMask}
\alias{abs.torch.Tensor}
\title{Abs}
\usage{
\method{abs}{fastai.torch_core.TensorMask}(a)

\method{abs}{torch.Tensor}(a)
}
\description{
Abs

Abs
}
