% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{AccumMetric}
\alias{AccumMetric}
\title{AccumMetric}
\usage{
AccumMetric(
  func,
  dim_argmax = NULL,
  activation = "no",
  thresh = NULL,
  to_np = FALSE,
  invert_arg = FALSE,
  flatten = TRUE,
  ...
)
}
\arguments{
\item{func}{func}

\item{dim_argmax}{dim_argmax}

\item{activation}{activation}

\item{thresh}{thresh}

\item{to_np}{to_np}

\item{invert_arg}{invert_arg}

\item{flatten}{flatten}
}
\description{
Stores predictions and targets on CPU in accumulate to perform final calculations with `func`.
}
