% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy}
\usage{
accuracy(input, targs)
}
\arguments{
\item{input}{input}

\item{targs}{targs}
}
\description{
Computes accuracy with `targs` when `input` is bs * n_classes.
}
