% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{adam_step}
\alias{adam_step}
\title{Adam_step}
\usage{
adam_step(p, lr, mom, step, sqr_mom, grad_avg, sqr_avg, eps, ...)
}
\arguments{
\item{p}{p}

\item{lr}{learning rate}

\item{mom}{momentum}

\item{step}{step}

\item{sqr_mom}{sqr momentum}

\item{grad_avg}{grad average}

\item{sqr_avg}{sqr average}

\item{eps}{epsilon}

\item{...}{additional arguments to pass}
}
\value{
None
}
\description{
Step for Adam with `lr` on `p`
}
