% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{AdaptiveAvgPool}
\alias{AdaptiveAvgPool}
\title{AdaptiveAvgPool}
\usage{
AdaptiveAvgPool(sz = 1, ndim = 2)
}
\arguments{
\item{sz}{size}

\item{ndim}{dimension size}
}
\description{
nn$AdaptiveAvgPool layer for `ndim`
}
