% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{AdaptiveConcatPool1d}
\alias{AdaptiveConcatPool1d}
\title{AdaptiveConcatPool1d}
\usage{
AdaptiveConcatPool1d(size = NULL)
}
\arguments{
\item{size}{output size}
}
\value{
None
}
\description{
Layer that concats `AdaptiveAvgPool1d` and `AdaptiveMaxPool1d`
}
