% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{AdaptiveConcatPool2d}
\alias{AdaptiveConcatPool2d}
\title{AdaptiveConcatPool2d}
\usage{
AdaptiveConcatPool2d(size = NULL)
}
\arguments{
\item{size}{output size}
}
\value{
None
}
\description{
Layer that concats `AdaptiveAvgPool2d` and `AdaptiveMaxPool2d`
}
