% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{AdaptiveGANSwitcher}
\alias{AdaptiveGANSwitcher}
\title{AdaptiveGANSwitcher}
\usage{
AdaptiveGANSwitcher(gen_thresh = NULL, critic_thresh = NULL)

AdaptiveGANSwitcher(gen_thresh = NULL, critic_thresh = NULL)
}
\arguments{
\item{gen_thresh}{gen_thresh}

\item{critic_thresh}{critic_thresh}
}
\description{
Switcher that goes back to generator/critic when the loss goes below `gen_thresh`/`crit_thresh`.

Switcher that goes back to generator/critic when the loss goes below `gen_thresh`/`crit_thresh`.
}
