% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R, R/Loss.R
\name{AdaptiveLoss}
\alias{AdaptiveLoss}
\title{Adaptive Loss}
\usage{
AdaptiveLoss(crit)

AdaptiveLoss(crit)
}
\arguments{
\item{crit}{crit}
}
\value{
None
}
\description{
Expand the `target` to match the `output` size before applying `crit`.

Expand the `target` to match the `output` size before applying `crit`.
}
