% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{add}
\alias{add}
\alias{+.fastai.torch_core.TensorMask}
\alias{floor.fastai.torch_core.TensorMask}
\alias{ceiling.fastai.torch_core.TensorMask}
\alias{cos.fastai.torch_core.TensorMask}
\alias{cosh.fastai.torch_core.TensorMask}
\alias{sin.fastai.torch_core.TensorMask}
\alias{sinh.fastai.torch_core.TensorMask}
\alias{+.torch.Tensor}
\alias{floor.torch.Tensor}
\alias{ceiling.torch.Tensor}
\alias{cos.torch.Tensor}
\alias{cosh.torch.Tensor}
\alias{sin.torch.Tensor}
\alias{sinh.torch.Tensor}
\title{Add}
\usage{
\method{+}{fastai.torch_core.TensorMask}(a, b)

\method{floor}{fastai.torch_core.TensorMask}(a)

\method{ceiling}{fastai.torch_core.TensorMask}(a)

\method{cos}{fastai.torch_core.TensorMask}(a)

\method{cosh}{fastai.torch_core.TensorMask}(a)

\method{sin}{fastai.torch_core.TensorMask}(a)

\method{sinh}{fastai.torch_core.TensorMask}(a)

\method{+}{torch.Tensor}(a, b)

\method{floor}{torch.Tensor}(a)

\method{ceiling}{torch.Tensor}(a)

\method{cos}{torch.Tensor}(a)

\method{cosh}{torch.Tensor}(a)

\method{sin}{torch.Tensor}(a)

\method{sinh}{torch.Tensor}(a)
}
\description{
Add

Floor

Ceil

Cos

Cosh

Sin

Sinh

Add

Floor

Ceil

Cos

Cosh

Sin

Sinh
}
