% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_transform.R
\name{add_cyclic_datepart}
\alias{add_cyclic_datepart}
\title{Add cyclic datepart}
\usage{
add_cyclic_datepart(
  df,
  field_name,
  prefix = NULL,
  drop = TRUE,
  time = FALSE,
  add_linear = FALSE
)
}
\arguments{
\item{df}{df}

\item{field_name}{field_name}

\item{prefix}{prefix}

\item{drop}{drop}

\item{time}{time}

\item{add_linear}{add_linear}
}
\value{
data frame
}
\description{
Helper function that adds trigonometric date/time features to a date in the column `field_name` of `df`.
}
