% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_transform.R
\name{add_datepart}
\alias{add_datepart}
\title{Add datepart}
\usage{
add_datepart(df, field_name, prefix = NULL, drop = TRUE, time = FALSE)
}
\arguments{
\item{df}{df}

\item{field_name}{field_name}

\item{prefix}{prefix}

\item{drop}{drop}

\item{time}{time}
}
\value{
data frame
}
\description{
Helper function that adds columns relevant to a date in the column `field_name` of `df`.
}
