% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{add_test}
\alias{add_test}
\title{Add test}
\usage{
add_test(object, items, label = NULL, tfms = NULL, tfm_y = NULL)
}
\arguments{
\item{items}{items}

\item{label}{label}

\item{tfms}{tfms}

\item{tfm_y}{tfm_y}
}
\description{
Add test set containing `items` with an arbitrary `label`.
}
