% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_signal.R
\name{AddNoise}
\alias{AddNoise}
\title{Add Noise}
\usage{
AddNoise(noise_level = 0.05, color = 0)
}
\arguments{
\item{noise_level}{noise level}

\item{color}{int, color}
}
\value{
None
}
\description{
Adds noise of specified color and level to the audio signal
}
