% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{alexnet}
\alias{alexnet}
\title{alexnet}
\usage{
alexnet(pretrained = FALSE, progress = TRUE)
}
\arguments{
\item{pretrained}{pretrained}

\item{progress}{progress}
}
\description{
AlexNet model architecture from the
}
\details{
`"One weird trick..." <https://arxiv.org/abs/1404.5997>`_ paper.
}
