% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{apply_perspective}
\alias{apply_perspective}
\title{apply_perspective}
\usage{
apply_perspective(coords, coeffs)
}
\arguments{
\item{coords}{coords}

\item{coeffs}{coeffs}
}
\description{
Apply perspective tranfom on `coords` with `coeffs`
}
