% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{APScoreBinary}
\alias{APScoreBinary}
\title{APScoreBinary}
\usage{
APScoreBinary(
  axis = -1,
  average = "macro",
  pos_label = 1,
  sample_weight = NULL
)
}
\arguments{
\item{axis}{axis}

\item{average}{average}

\item{pos_label}{pos_label}

\item{sample_weight}{sample_weight}
}
\value{
None
}
\description{
Average Precision for single-label binary classification problems
}
