% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{APScoreMulti}
\alias{APScoreMulti}
\title{APScoreMulti}
\usage{
APScoreMulti(
  sigmoid = TRUE,
  average = "macro",
  pos_label = 1,
  sample_weight = NULL
)
}
\arguments{
\item{sigmoid}{sigmoid}

\item{average}{average}

\item{pos_label}{pos_label}

\item{sample_weight}{sample_weight}
}
\description{
Average Precision for multi-label classification problems
}
