% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_config.R
\name{AudioBlock}
\alias{AudioBlock}
\title{AudioBlock}
\usage{
AudioBlock(
  cache_folder = NULL,
  sample_rate = 16000,
  force_mono = TRUE,
  crop_signal_to = NULL
)
}
\arguments{
\item{cache_folder}{cache folder}

\item{sample_rate}{sample rate}

\item{force_mono}{force mono or not}

\item{crop_signal_to}{int, crop signal}
}
\value{
None
}
\description{
A `TransformBlock` for audios
}
