% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_config.R
\name{AudioBlock_from_folder}
\alias{AudioBlock_from_folder}
\title{AudioBlock from folder}
\usage{
AudioBlock_from_folder(
  path,
  sample_rate = 16000,
  force_mono = TRUE,
  crop_signal_to = NULL
)
}
\arguments{
\item{path}{path}

\item{sample_rate}{sample_rate}

\item{force_mono}{force_mono}

\item{crop_signal_to}{crop_signal_to}
}
\value{
None
}
\description{
Build a `AudioBlock` from a `path` and caches some intermediary results
}
