% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_spectrogram.R
\name{AudioSpectrogram}
\alias{AudioSpectrogram}
\title{AudioSpectrogram}
\usage{
AudioSpectrogram(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
module
}
\description{
AudioSpectrogram
}
