% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_signal.R
\name{AudioTensor}
\alias{AudioTensor}
\title{Audio Tensor}
\usage{
AudioTensor(x, sr = NULL)
}
\arguments{
\item{x}{tensor}

\item{sr}{sr}
}
\value{
tensor
}
\description{
Semantic torch tensor that represents an audio.
}
