% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_signal.R
\name{AudioTensor_create}
\alias{AudioTensor_create}
\title{AudioTensor create}
\usage{
AudioTensor_create(
  fn,
  cache_folder = NULL,
  out = NULL,
  normalization = TRUE,
  channels_first = TRUE,
  num_frames = 0,
  offset = 0,
  signalinfo = NULL,
  encodinginfo = NULL,
  filetype = NULL
)
}
\arguments{
\item{fn}{fn}

\item{cache_folder}{cache_folder}

\item{out}{out}

\item{normalization}{normalization}

\item{channels_first}{channels_first}

\item{num_frames}{num_frames}

\item{offset}{offset}

\item{signalinfo}{signalinfo}

\item{encodinginfo}{encodinginfo}

\item{filetype}{filetype}
}
\value{
None
}
\description{
Creates audio tensor from file
}
