% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_spectrogram.R
\name{AudioToMFCC}
\alias{AudioToMFCC}
\title{AudioToMFCC}
\usage{
AudioToMFCC(
  sample_rate = 16000,
  n_mfcc = 40,
  dct_type = 2,
  norm = "ortho",
  log_mels = FALSE,
  melkwargs = NULL
)
}
\arguments{
\item{sample_rate}{sample rate}

\item{n_mfcc}{number of mel-frequency cepstral coefficients}

\item{dct_type}{dct type}

\item{norm}{normalization type}

\item{log_mels}{apply log to mels}

\item{melkwargs}{additional arguments for mels}
}
\value{
None
}
\description{
Transform to create MFCC features from audio tensors.
}
