% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_spectrogram.R
\name{AudioToSpec_from_cfg}
\alias{AudioToSpec_from_cfg}
\title{AudioToSpec from cfg}
\usage{
AudioToSpec_from_cfg(audio_cfg)
}
\arguments{
\item{audio_cfg}{audio_cfg}
}
\value{
None
}
\description{
Creates AudioToSpec from configuration file
}
