% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R
\name{aug_transforms}
\alias{aug_transforms}
\title{Aug_transforms}
\usage{
aug_transforms(
  mult = 1,
  do_flip = TRUE,
  flip_vert = FALSE,
  max_rotate = 10,
  min_zoom = 1,
  max_zoom = 1.1,
  max_lighting = 0.2,
  max_warp = 0.2,
  p_affine = 0.75,
  p_lighting = 0.75,
  xtra_tfms = NULL,
  size = NULL,
  mode = "bilinear",
  pad_mode = "reflection",
  align_corners = TRUE,
  batch = FALSE,
  min_scale = 1
)
}
\arguments{
\item{mult}{mult}

\item{do_flip}{do_flip}

\item{flip_vert}{flip_vert}

\item{max_rotate}{max_rotate}

\item{min_zoom}{min_zoom}

\item{max_zoom}{max_zoom}

\item{max_lighting}{max_lighting}

\item{max_warp}{max_warp}

\item{p_affine}{p_affine}

\item{p_lighting}{p_lighting}

\item{xtra_tfms}{xtra_tfms}

\item{size}{size}

\item{mode}{mode}

\item{pad_mode}{pad_mode}

\item{align_corners}{align_corners}

\item{batch}{batch}

\item{min_scale}{min_scale}
}
\description{
Utility func to easily create a list of flip, rotate, zoom, warp, lighting transforms.
}
