% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loss.R
\name{AvgSmoothLoss}
\alias{AvgSmoothLoss}
\title{AvgSmoothLoss}
\usage{
AvgSmoothLoss(beta = 0.98)
}
\arguments{
\item{beta}{beta, defaults to 0.98}
}
\value{
Loss object
}
\description{
Smooth average of the losses (exponentially weighted with `beta`)
}
