% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_models.R
\name{AWD_LSTM}
\alias{AWD_LSTM}
\title{AWD_LSTM}
\usage{
AWD_LSTM(
  vocab_sz,
  emb_sz,
  n_hid,
  n_layers,
  pad_token = 1,
  hidden_p = 0.2,
  input_p = 0.6,
  embed_p = 0.1,
  weight_p = 0.5,
  bidir = FALSE
)
}
\arguments{
\item{vocab_sz}{vocab_sz}

\item{emb_sz}{emb_sz}

\item{n_hid}{n_hid}

\item{n_layers}{n_layers}

\item{pad_token}{pad_token}

\item{hidden_p}{hidden_p}

\item{input_p}{input_p}

\item{embed_p}{embed_p}

\item{weight_p}{weight_p}

\item{bidir}{bidir}
}
\value{
None
}
\description{
AWD-LSTM inspired by https://arxiv.org/abs/1708.02182
}
