% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{BalancedAccuracy}
\alias{BalancedAccuracy}
\title{BalancedAccuracy}
\usage{
BalancedAccuracy(axis = -1, sample_weight = NULL, adjusted = FALSE)
}
\arguments{
\item{axis}{axis}

\item{sample_weight}{sample_weight}

\item{adjusted}{adjusted}
}
\description{
Balanced Accuracy for single-label binary classification problems
}
