% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_core.R
\name{BaseTokenizer}
\alias{BaseTokenizer}
\title{BaseTokenizer}
\usage{
BaseTokenizer(split_char = " ")
}
\arguments{
\item{split_char}{separator}
}
\value{
None
}
\description{
Basic tokenizer that just splits on spaces
}
