% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{basic_critic}
\alias{basic_critic}
\title{Basic critic}
\usage{
basic_critic(in_size, n_channels, ...)
}
\arguments{
\item{in_size}{in_size}

\item{n_channels}{The number of channels}

\item{...}{additional parameters to pass}
}
\value{
None
}
\description{
A basic critic for images `n_channels` x `in_size` x `in_size`.
}
