% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_config.R
\name{BasicMFCC}
\alias{BasicMFCC}
\title{BasicMFCC}
\usage{
BasicMFCC(
  sample_rate = 16000,
  n_mfcc = 40,
  dct_type = 2,
  norm = "ortho",
  log_mels = FALSE,
  melkwargs = NULL
)
}
\arguments{
\item{sample_rate}{sample_rate}

\item{n_mfcc}{n_mfcc}

\item{dct_type}{dct_type}

\item{norm}{norm}

\item{log_mels}{log_mels}

\item{melkwargs}{melkwargs}
}
\value{
None
}
\description{
BasicMFCC
}
