% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_config.R
\name{BasicSpectrogram}
\alias{BasicSpectrogram}
\title{BasicSpectrogram}
\usage{
BasicSpectrogram(
  n_fft = 400,
  win_length = NULL,
  hop_length = NULL,
  pad = 0,
  window_fn = torch$hann_window,
  power = 2,
  normalized = FALSE,
  wkwargs = NULL,
  mel = FALSE,
  to_db = TRUE
)
}
\arguments{
\item{n_fft}{n_fft}

\item{win_length}{win_length}

\item{hop_length}{hop_length}

\item{pad}{pad}

\item{window_fn}{window_fn}

\item{power}{power}

\item{normalized}{normalized}

\item{wkwargs}{wkwargs}

\item{mel}{mel}

\item{to_db}{to_db}
}
\value{
None
}
\description{
BasicSpectrogram
}
