% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{BatchNorm1dFlat}
\alias{BatchNorm1dFlat}
\title{BatchNorm1dFlat}
\usage{
BatchNorm1dFlat(
  num_features,
  eps = 1e-05,
  momentum = 0.1,
  affine = TRUE,
  track_running_stats = TRUE
)
}
\arguments{
\item{num_features}{number of features}

\item{eps}{epsilon}

\item{momentum}{momentum}

\item{affine}{affine}

\item{track_running_stats}{track running statistics}
}
\value{
None
}
\description{
`nn.BatchNorm1d`, but first flattens leading dimensions
}
