% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loss.R
\name{BCELossFlat}
\alias{BCELossFlat}
\title{BCELossFlat}
\usage{
BCELossFlat(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
Loss object
}
\description{
Flattens input and output, same as nn$BCELoss
}
