% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loss.R
\name{BCEWithLogitsLossFlat}
\alias{BCEWithLogitsLossFlat}
\title{BCEWithLogitsLossFlat}
\usage{
BCEWithLogitsLossFlat(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
Loss object
}
\description{
BCEWithLogitsLossFlat
}
