% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{BrierScore}
\alias{BrierScore}
\title{BrierScore}
\usage{
BrierScore(axis = -1, sample_weight = NULL, pos_label = NULL)
}
\arguments{
\item{axis}{axis}

\item{sample_weight}{sample_weight}

\item{pos_label}{pos_label}
}
\value{
None
}
\description{
Brier score for single-label classification problems
}
