% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{BrierScoreMulti}
\alias{BrierScoreMulti}
\title{BrierScoreMulti}
\usage{
BrierScoreMulti(
  thresh = 0.5,
  sigmoid = TRUE,
  sample_weight = NULL,
  pos_label = NULL
)
}
\arguments{
\item{thresh}{thresh}

\item{sigmoid}{sigmoid}

\item{sample_weight}{sample_weight}

\item{pos_label}{pos_label}
}
\description{
Brier score for multi-label classification problems
}
