% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{calculate_rouge}
\alias{calculate_rouge}
\title{Calculate_rouge}
\usage{
calculate_rouge(
  predicted_txts,
  reference_txts,
  rouge_keys = c("rouge1", "rouge2", "rougeL"),
  use_stemmer = TRUE
)
}
\arguments{
\item{predicted_txts}{predicted texts}

\item{reference_txts}{reference texts}

\item{rouge_keys}{rouge keys}

\item{use_stemmer}{use stemmer or not}
}
\value{
None
}
\description{
Calculate_rouge
}
