% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_transform.R
\name{Categorify}
\alias{Categorify}
\title{Categorify}
\usage{
Categorify(cat_names, cont_names)
}
\arguments{
\item{cat_names}{The names of the categorical variables}

\item{cont_names}{The names of the continuous variables}
}
\value{
None
}
\description{
Transform the categorical variables to that type.
}
