% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{ceiling.fastai.torch_core.TensorMask}
\alias{ceiling.fastai.torch_core.TensorMask}
\title{Ceil}
\usage{
\method{ceiling}{fastai.torch_core.TensorMask}(x)
}
\arguments{
\item{x}{tensor}
}
\value{
tensor
}
\description{
Ceil
}
