% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_signal.R
\name{ChangeVolume}
\alias{ChangeVolume}
\title{Change Volume}
\usage{
ChangeVolume(p = 0.5, lower = 0.5, upper = 1.5)
}
\arguments{
\item{p}{p}

\item{lower}{lower}

\item{upper}{upper}
}
\value{
None
}
\description{
Changes the volume of the signal
}
