% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{children_and_parameters}
\alias{children_and_parameters}
\title{Children_and_parameters}
\usage{
children_and_parameters(m)
}
\arguments{
\item{m}{parameters}
}
\value{
None
}
\description{
Return the children of `m` and its direct parameters not registered in modules.
}
