% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{ClassificationInterpretation_from_learner}
\alias{ClassificationInterpretation_from_learner}
\title{ClassificationInterpretation_from_learner}
\usage{
ClassificationInterpretation_from_learner(
  learn,
  ds_idx = 1,
  dl = NULL,
  act = NULL
)
}
\arguments{
\item{learn}{learn}

\item{ds_idx}{ds_idx}

\item{dl}{dl}

\item{act}{act}
}
\value{
interpretation object
}
\description{
Construct interpretation object from a learner
}
