% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{cm}
\alias{cm}
\title{Cmap}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
cm
}
\value{
module
}
\description{
Cmap
}
\keyword{datasets}
