% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_learner.R
\name{cnn_config}
\alias{cnn_config}
\title{Cnn config}
\usage{
cnn_config(
  cut = NULL,
  pretrained = TRUE,
  n_in = 3,
  init = nn()$init$kaiming_normal_,
  custom_head = NULL,
  concat_pool = TRUE,
  lin_ftrs = NULL,
  ps = 0.5,
  bn_final = FALSE,
  lin_first = FALSE,
  y_range = NULL
)
}
\arguments{
\item{cut}{cut}

\item{pretrained}{pre-trained or not}

\item{n_in}{input shape}

\item{init}{initializer}

\item{custom_head}{custom head}

\item{concat_pool}{concatenate pooling}

\item{lin_ftrs}{linear filters}

\item{ps}{parameter server}

\item{bn_final}{batch normalization final}

\item{lin_first}{linear first}

\item{y_range}{y_range}
}
\value{
None
}
\description{
Convenience function to easily create a config for `create_cnn_model`
}
