% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_metrics.R
\name{COCOMetric}
\alias{COCOMetric}
\title{COCOMetric}
\usage{
COCOMetric(
  metric_type = COCOMetricType()$bbox,
  print_summary = FALSE,
  show_pbar = FALSE
)
}
\arguments{
\item{metric_type}{Dependent on the task you're solving.}

\item{print_summary}{If `TRUE`, prints a table with statistics.}

\item{show_pbar}{If `TRUE` shows pbar when preparing the data for evaluation.}
}
\value{
None
}
\description{
Wrapper around [cocoapi evaluator](https://github.com/cocodataset/cocoapi)
}
\details{
Calculates average precision. # Arguments metric_type: Dependent on the task you're solving. print_summary: If `TRUE`, prints a table with statistics. show_pbar: If `TRUE` shows pbar when preparing the data for evaluation.
}
