% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{CohenKappa}
\alias{CohenKappa}
\title{CohenKappa}
\usage{
CohenKappa(axis = -1, labels = NULL, weights = NULL, sample_weight = NULL)
}
\arguments{
\item{axis}{axis}

\item{labels}{labels}

\item{weights}{weights}

\item{sample_weight}{sample_weight}
}
\value{
None
}
\description{
Cohen kappa for single-label classification problems
}
