% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab.R
\name{CollabDataLoaders_from_df}
\alias{CollabDataLoaders_from_df}
\title{CollabDataLoaders_from_df}
\usage{
CollabDataLoaders_from_df(
  ratings,
  valid_pct = 0.2,
  user_name = NULL,
  item_name = NULL,
  rating_name = NULL,
  seed = NULL,
  path = ".",
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  device = NULL
)
}
\arguments{
\item{ratings}{ratings}

\item{valid_pct}{valid_pct}

\item{user_name}{user_name}

\item{item_name}{item_name}

\item{rating_name}{rating_name}

\item{seed}{seed}

\item{path}{path}

\item{bs}{bs}

\item{val_bs}{val_bs}

\item{shuffle_train}{shuffle_train}

\item{device}{device}
}
\description{
Create a `DataLoaders` suitable for collaborative filtering from `ratings`.
}
