% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R, R/callbacks_data.R
\name{CollectDataCallback}
\alias{CollectDataCallback}
\title{CollectDataCallback}
\usage{
CollectDataCallback(...)

CollectDataCallback(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None

None
}
\description{
Collect all batches, along with pred and loss, into self.data. Mainly for testing
}
