% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{colors}
\alias{colors}
\title{Colors}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
colors
}
\value{
module
}
\description{
Colors
}
\keyword{datasets}
