% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{ColReader}
\alias{ColReader}
\title{ColReader}
\usage{
ColReader(cols, pref = "", suff = "", label_delim = NULL)
}
\arguments{
\item{cols}{cols}

\item{pref}{pref}

\item{suff}{suff}

\item{label_delim}{label_delim}
}
\description{
Read `cols` in `row` with potential `pref` and `suff`
}
\details{

}
