% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{combined_flat_anneal}
\alias{combined_flat_anneal}
\title{Combined_flat_anneal}
\usage{
combined_flat_anneal(pct, start_lr, end_lr = 0, curve_type = "linear")
}
\arguments{
\item{pct}{Proportion of training with a constant learning rate.}

\item{start_lr}{Desired starting learning rate, used for beginnning pct of training.}

\item{end_lr}{Desired end learning rate, training will conclude at this learning rate.}

\item{curve_type}{Curve type for learning rate annealing. Options are 'linear', 'cosine', and 'exponential'.}
}
\description{
Create a schedule with constant learning rate `start_lr` for `pct`
proportion of the training, and a `curve_type` learning rate (till `end_lr`) for
remaining portion of training.
}
