% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle_wrappers.R
\name{competition_download_file}
\alias{competition_download_file}
\title{Competition download file}
\usage{
competition_download_file(
  competition,
  file_name,
  path = NULL,
  force = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{competition}{the name of the competition}

\item{file_name}{the configuration file name}

\item{path}{a path to download the file to}

\item{force}{force the download if the file already exists (default FALSE)}

\item{quiet}{suppress verbose output (default is FALSE)}
}
\value{
None
}
\description{
download a competition file to a designated location, or use
}
