% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle_wrappers.R
\name{competition_download_files}
\alias{competition_download_files}
\title{competition_download_files}
\usage{
competition_download_files(
  competition,
  path = NULL,
  force = FALSE,
  quiet = FALSE,
  unzip = FALSE
)
}
\arguments{
\item{competition}{competition}

\item{path}{path}

\item{force}{force}

\item{quiet}{quiet}
}
\description{
downloads all competition files.
}
\details{
Parameters
=========
competition: the name of the competition
path: a path to download the file to
force: force the download if the file already exists (default FALSE)
quiet: suppress verbose output (default is TRUE)
}
\section{competition}{
 the name of the competition:
path: a path to download the file to force: force the download if the file already exists (default FALSE) quiet: suppress verbose output (default is TRUE)
}

\section{path}{
 a path to download the file to:
force: force the download if the file already exists (default FALSE) quiet: suppress verbose output (default is TRUE)
}

\section{force}{
 force the download if the file already exists (default False):
quiet: suppress verbose output (default is TRUE)
}

