% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle_wrappers.R
\name{competition_list_files}
\alias{competition_list_files}
\title{competition_list_files}
\usage{
competition_list_files(competition)
}
\arguments{
\item{competition}{competition the name of the competition}
}
\description{
list files for competition
}
