% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle_wrappers.R
\name{competition_submit}
\alias{competition_submit}
\title{Competition submit}
\usage{
competition_submit(file_name, message, competition, quiet = FALSE)
}
\arguments{
\item{file_name}{the competition metadata file}

\item{message}{the submission description}

\item{competition}{the competition name}

\item{quiet}{suppress verbose output (default is FALSE)}
}
\value{
None
}
\description{
Competition submit
}
