% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle_wrappers.R
\name{competitions_list}
\alias{competitions_list}
\title{competitions_list}
\usage{
competitions_list(
  group = NULL,
  category = NULL,
  sort_by = NULL,
  page = 1,
  search = NULL
)
}
\arguments{
\item{group}{group}

\item{category}{category}

\item{sort_by}{sort_by}

\item{page}{page}

\item{search}{search}
}
\description{
make call to list competitions, format the response, and return
}
\details{
a list of Competition instances Parameters
========== page: the page to return (default is 1)
search: a search term to use (default is empty string)
sort_by: how to sort the result, see valid_competition_sort_by for options
category: category to filter result to
group: group to filter result to
}
\section{page}{
 the page to return (default is 1):
search: a search term to use (default is empty string) sort_by: how to sort the result, see valid_competition_sort_by for options category: category to filter result to group: group to filter result to
}

\section{search}{
 a search term to use (default is empty string):
sort_by: how to sort the result, see valid_competition_sort_by for options category: category to filter result to group: group to filter result to
}

\section{sort_by}{
 how to sort the result, see valid_competition_sort_by for options:
category: category to filter result to group: group to filter result to
}

\section{category}{
 category to filter result to:
group: group to filter result to
}

