% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_transform.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{confusion_matrix}
\usage{
confusion_matrix(model, slice_size = 1)
}
\arguments{
\item{slice_size}{slice_size}
}
\description{
Confusion matrix as an `np.ndarray`.
}
