% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{conv_norm_lr}
\alias{conv_norm_lr}
\title{Conv_norm_lr}
\usage{
conv_norm_lr(
  ch_in,
  ch_out,
  norm_layer = NULL,
  ks = 3,
  bias = TRUE,
  pad = 1,
  stride = 1,
  activ = TRUE,
  slope = 0.2,
  init = nn()$init$normal_,
  init_gain = 0.02
)
}
\arguments{
\item{ch_in}{input}

\item{ch_out}{output}

\item{norm_layer}{normalziation layer}

\item{ks}{kernel size}

\item{bias}{bias}

\item{pad}{pad}

\item{stride}{stride}

\item{activ}{activation}

\item{slope}{slope}

\item{init}{inititializer}

\item{init_gain}{initializer gain}
}
\value{
None
}
\description{
Conv_norm_lr
}
