% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{convT_norm_relu}
\alias{convT_norm_relu}
\title{ConvT_norm_relu}
\usage{
convT_norm_relu(ch_in, ch_out, norm_layer, ks = 3, stride = 2, bias = TRUE)
}
\arguments{
\item{ch_in}{input}

\item{ch_out}{output}

\item{norm_layer}{normalziation layer}

\item{ks}{kernel size}

\item{stride}{stride size}

\item{bias}{bias true or not}
}
\value{
None
}
\description{
ConvT_norm_relu
}
