% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{cos.fastai.torch_core.TensorMask}
\alias{cos.fastai.torch_core.TensorMask}
\title{Cos}
\usage{
\method{cos}{fastai.torch_core.TensorMask}(x)
}
\arguments{
\item{x}{tensor}
}
\value{
tensor
}
\description{
Cos
}
