% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastaibuilt.R
\name{crappifier}
\alias{crappifier}
\title{Crappifier}
\usage{
crappifier(path_lr, path_hr)
}
\arguments{
\item{path_lr}{path from (origin)}

\item{path_hr}{path to (destination)}
}
\value{
None
}
\description{
Crappifier
}
