% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_learner.R
\name{create_body}
\alias{create_body}
\title{Create_body}
\usage{
create_body(...)
}
\arguments{
\item{..}{parameters to pass}
}
\value{
None
}
\description{
Cut off the body of a typically pretrained `arch` as determined by `cut`
}
