% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_learner.R
\name{create_cnn_model}
\alias{create_cnn_model}
\title{Create_cnn_model}
\usage{
create_cnn_model(
  arch,
  n_out,
  cut = NULL,
  pretrained = TRUE,
  n_in = 3,
  init = nn$init$kaiming_normal_,
  custom_head = NULL,
  concat_pool = TRUE,
  lin_ftrs = NULL,
  ps = 0.5,
  bn_final = FALSE,
  lin_first = FALSE,
  y_range = NULL
)
}
\arguments{
\item{arch}{a model architecture}

\item{n_out}{number of out}

\item{cut}{cut}

\item{pretrained}{pretrained model or not}

\item{n_in}{input shape}

\item{init}{initializer}

\item{custom_head}{custom head}

\item{concat_pool}{concatenate pooling}

\item{lin_ftrs}{lin_ftrs}

\item{ps}{ps}

\item{bn_final}{bn_final}

\item{lin_first}{lin_first}

\item{y_range}{y_range}
}
\value{
None
}
\description{
Create custom convnet architecture using `arch`, `n_in` and `n_out`
}
