% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_model.R
\name{create_fcn}
\alias{create_fcn}
\title{Create_fcn}
\usage{
create_fcn(ni, nout, ks = 9, conv_sizes = c(128, 256, 128), stride = 1)
}
\arguments{
\item{ni}{number of input channels}

\item{nout}{output shape}

\item{ks}{kernel size}

\item{conv_sizes}{convolution sizes}

\item{stride}{stride}
}
\value{
model
}
\description{
A bunch of convolutions stacked together.
}
