% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_learner.R
\name{create_head}
\alias{create_head}
\title{Create_head}
\usage{
create_head(
  nf,
  n_out,
  lin_ftrs = NULL,
  ps = 0.5,
  concat_pool = TRUE,
  bn_final = FALSE,
  lin_first = FALSE,
  y_range = NULL
)
}
\arguments{
\item{nf}{number of features}

\item{n_out}{number of out features}

\item{lin_ftrs}{linear features}

\item{ps}{parameter server}

\item{concat_pool}{concatenate pooling}

\item{bn_final}{batch normalization final}

\item{lin_first}{linear first}

\item{y_range}{y_range}
}
\value{
None
}
\description{
Model head that takes `nf` features, runs through `lin_ftrs`, and out `n_out` classes.
}
