% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_model.R
\name{create_mlp}
\alias{create_mlp}
\title{Create_mlp}
\usage{
create_mlp(ni, nout, linear_sizes = c(500, 500, 500))
}
\arguments{
\item{ni}{number of input channels}

\item{nout}{output shape}

\item{linear_sizes}{linear output sizes}
}
\value{
model
}
\description{
A simple model builder to create a bunch of BatchNorm1d, Dropout and
Linear layers, with ```act_fn``` activations.
}
