% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_model.R
\name{create_resnet}
\alias{create_resnet}
\title{Create_resnet}
\usage{
create_resnet(
  ni,
  nout,
  kss = c(9, 5, 3),
  conv_sizes = c(64, 128, 128),
  stride = 1
)
}
\arguments{
\item{ni}{number of input channels}

\item{nout}{output shape}

\item{kss}{kernel size}

\item{conv_sizes}{convolution sizes}

\item{stride}{stride}
}
\value{
model
}
\description{
Basic 11 Layer - 1D resnet builder
}
