% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeze_and_helpers.R
\name{create_slice}
\alias{create_slice}
\title{Create a slice object for fine tuning}
\usage{
create_slice(lr_earliest_layer, lr_last_layer)
}
\arguments{
\item{lr_earliest_layers}{Learning rate in the earliest layer}

\item{lr_last_layers}{Learning rate in the last layer}
}
\value{
A slice object
}
\description{
In transfert learning, the recommended approach to fine_tune a model is to use 
a discriminative learning rate, i.e. having a smaller learning rate for the 
earlier layers of the models, and a bigger one for the last layers. This allow 
to preserve the weight of the layers that recognize basic features.
The fit_one_cycle() method of Fastai accept a slice object as argument to 
the parameter lr_max, to perform fine tuning using a discriminative learning 
rate. This function is meant to create a slice object from R conveniently. 
Please refer to the page 201 of the Fastai book, or the beginner tutorial of 
the text tutorial on the fastai website.
}
\examples{

\dontrun{
sl <- create_slice(1e-6, 1e-4)
learnR \%>\% fit_one_cycle(n_epoch = 4, lr_max=sl)
}
}
