% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{CropPad}
\alias{CropPad}
\title{CropPad}
\usage{
CropPad(size, pad_mode = "zeros", ...)
}
\arguments{
\item{size}{size}

\item{pad_mode}{pad_mode}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Center crop or pad an image to `size`
}
