% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_spectrogram.R
\name{CropTime}
\alias{CropTime}
\title{Crop Time}
\usage{
CropTime(duration, pad_mode = AudioPadType$Zeros)
}
\arguments{
\item{duration}{duration}

\item{pad_mode}{pad_mode}
}
\value{
None
}
\description{
Random crops full spectrogram to be length specified in ms by crop_duration
}
