% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loss.R
\name{CrossEntropyLossFlat}
\alias{CrossEntropyLossFlat}
\title{CrossEntropyLossFlat}
\usage{
CrossEntropyLossFlat(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
Loss object
}
\description{
Same as `nn$Module`, but no need for subclasses to call `super().__init__`
}
