% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{CSVLogger}
\alias{CSVLogger}
\title{CSVLogger}
\usage{
CSVLogger(learn, filename = "history", append = FALSE)
}
\arguments{
\item{learn}{learner object}

\item{filename}{the name of the file "history.csv"}

\item{append}{whether to append the new file or not}
}
\value{
None
}
\description{
A `LearnerCallback` that saves history of metrics while training `learn` into CSV `filename`.
}
