% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{cutout_gaussian}
\alias{cutout_gaussian}
\title{cutout_gaussian}
\usage{
cutout_gaussian(x, areas)
}
\arguments{
\item{x}{x}

\item{areas}{areas}
}
\description{
Replace all `areas` in `x` with N(0,1) noise
}
\details{

}
