% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{cycle_learner}
\alias{cycle_learner}
\title{Cycle_learner}
\usage{
cycle_learner(
  dls,
  m,
  opt_func = Adam(),
  show_imgs = TRUE,
  imgA = TRUE,
  imgB = TRUE,
  show_img_interval = 10,
  ...
)
}
\arguments{
\item{dls}{dataloader}

\item{m}{CycleGAN model}

\item{opt_func}{optimizer}

\item{show_imgs}{show images}

\item{imgA}{image a (from)}

\item{imgB}{image B (to)}

\item{show_img_interval}{show images interval rafe}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Initialize and return a `Learner` object with the data in `dls`, CycleGAN model `m`, optimizer function `opt_func`, metrics `metrics`,
}
\details{
and callbacks `cbs`. Additionally, if `show_imgs` is TRUE, it will show intermediate predictions during training. It will show domain
B-to-A predictions if `imgA` is TRUE and/or domain A-to-B predictions if `imgB` is TRUE. Additionally, it will show images every
`show_img_interval` epochs. ` Other `Learner` arguments can be passed as well.
}
