% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{CycleGAN}
\alias{CycleGAN}
\title{CycleGAN}
\usage{
CycleGAN(
  ch_in = 3,
  ch_out = 3,
  n_features = 64,
  disc_layers = 3,
  gen_blocks = 9,
  lsgan = TRUE,
  drop = 0,
  norm_layer = NULL
)
}
\arguments{
\item{ch_in}{input}

\item{ch_out}{output}

\item{n_features}{number of features}

\item{disc_layers}{discriminator layers}

\item{gen_blocks}{generator blocks}

\item{lsgan}{ls gan}

\item{drop}{dropout rate}

\item{norm_layer}{normalziation layer}
}
\value{
None
}
\description{
CycleGAN model.
}
\details{
When called, takes in input batch of real images from both domains and outputs fake images for the opposite domains (with the generators).
Also outputs identity images after passing the images into generators that outputs its domain type (needed for identity loss). Attributes: `G_A` (`nn.Module`): takes real input B and generates fake input A `G_B` (`nn.Module`): takes real input A and generates fake input B `D_A` (`nn.Module`): trained to make the difference between real input A and fake input A `D_B` (`nn.Module`): trained to make the difference between real input B and fake input B
}
