% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{CycleGANTrainer}
\alias{CycleGANTrainer}
\title{CycleGANTrainer}
\usage{
CycleGANTrainer(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
Learner Callback for training a CycleGAN model.
}
