% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloader.R
\name{Data_Loaders}
\alias{Data_Loaders}
\title{Data Loaders}
\usage{
Data_Loaders(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
loader object
}
\description{
Data Loaders
}
\examples{

\dontrun{

data = Data_Loaders(train_loader, test_loader)

learn = Learner(data, Net(), loss_func = F$nll_loss,
                opt_func = Adam(), metrics = accuracy, cbs = CudaCallback())

learn \%>\% fit_one_cycle(1, 1e-2)

}

}
