% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{DataBlock}
\alias{DataBlock}
\title{DataBlock}
\usage{
DataBlock(
  blocks = NULL,
  dl_type = NULL,
  getters = NULL,
  n_inp = NULL,
  item_tfms = NULL,
  batch_tfms = NULL,
  ...
)
}
\arguments{
\item{blocks}{blocks}

\item{dl_type}{dl_type}

\item{getters}{getters}

\item{n_inp}{n_inp}

\item{item_tfms}{item_tfms}

\item{batch_tfms}{batch_tfms}
}
\description{
Generic container to quickly build `Datasets` and `DataLoaders`
}
