% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{databunch}
\alias{databunch}
\title{Databunch}
\usage{
databunch(
  object,
  path = NULL,
  bs = 64,
  val_bs = NULL,
  num_workers = 6,
  dl_tfms = NULL,
  device = NULL,
  collate_fn = data_collate(),
  no_check = FALSE
)
}
\arguments{
\item{path}{path}

\item{bs}{bs}

\item{val_bs}{val_bs}

\item{num_workers}{num_workers}

\item{dl_tfms}{dl_tfms}

\item{device}{device}

\item{collate_fn}{collate_fn}

\item{no_check}{no_check}
}
\description{
Create an `DataBunch` from self, `path` will override `self.path`,
`kwargs` are passed to `DataBunch.create`.
}
\details{

}
