% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R, R/tabular.R
\name{dataloaders}
\alias{dataloaders}
\title{dataloaders}
\usage{
dataloaders(
  object,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  n = NULL,
  ...
)

dataloaders(
  object,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  n = NULL,
  ...
)
}
\arguments{
\item{object}{object}

\item{bs}{bs}

\item{val_bs}{val_bs}

\item{shuffle_train}{shuffle_train}

\item{n}{n}

\item{...}{parameters to pass}

\item{source}{source}

\item{path}{path}

\item{verbose}{verbose}

\item{shuffle}{shuffle}

\item{num_workers}{num_workers}

\item{do_setup}{do_setup}

\item{pin_memory}{pin_memory}

\item{timeout}{timeout}

\item{batch_size}{batch_size}

\item{drop_last}{drop_last}

\item{indexed}{indexed}

\item{device}{device}
}
\description{
Create a `DataLoaders` object from `source`

Get a `DataLoaders`
}
\details{

}
