% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medical.R
\name{dcmread}
\alias{dcmread}
\title{Read dicom}
\usage{
dcmread(fn, force = FALSE)
}
\arguments{
\item{fn}{file name}

\item{force}{logical, force}
}
\value{
dicom object
}
\description{
Open a `DICOM` file
}
\examples{

\dontrun{

img = dcmread('hemorrhage.dcm')


}

}
