% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastinference.R
\name{decision_plot}
\alias{decision_plot}
\title{Decision_plot}
\usage{
decision_plot(object, class_id = 0, row_idx = -1, dpi = 200, ...)
}
\arguments{
\item{object}{ShapInterpretation object}

\item{class_id}{is used to indicate the class of interest for a classification model.
It can either be an int or str representation for a class of choice. Each colored line in
the plot represents the model's prediction for a single observation.}

\item{row_idx}{If no index is passed in to use from the data, it will default to the first ten samples
on the test set. Note:plotting too many samples at once can make the plot illegible.}

\item{dpi}{dots per inch}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Visualizes a model's decisions using cumulative SHAP values.
}
