% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_learner.R
\name{default_split}
\alias{default_split}
\title{Default_split}
\usage{
default_split(m)
}
\arguments{
\item{m}{parameters}
}
\value{
None
}
\description{
Default split of a model between body and head
}
