% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_utils.R
\name{denormalize_imagenet}
\alias{denormalize_imagenet}
\title{Denormalize_imagenet}
\usage{
denormalize_imagenet(img)
}
\arguments{
\item{img}{img}
}
\value{
None
}
\description{
Denormalize_imagenet
}
