% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{densenet121}
\alias{densenet121}
\title{densenet121}
\usage{
densenet121(pretrained = FALSE, progress)
}
\arguments{
\item{pretrained}{pretrained}

\item{progress}{progress}
}
\description{
Densenet-121 model from
}
\details{
`"Densely Connected Convolutional Networks" <https://arxiv.org/pdf/1608.06993.pdf>`_
}
