% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{densenet161}
\alias{densenet161}
\title{densenet161}
\usage{
densenet161(pretrained = FALSE, progress = TRUE)
}
\arguments{
\item{pretrained}{pretrained}

\item{progress}{progress}
}
\description{
Densenet-161 model from
}
\details{
`"Densely Connected Convolutional Networks" <https://arxiv.org/pdf/1608.06993.pdf>`_
}
