% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{DenseResBlock}
\alias{DenseResBlock}
\title{Dense Res Block}
\usage{
DenseResBlock(
  nf,
  norm_type = 1,
  ks = 3,
  stride = 1,
  padding = NULL,
  bias = NULL,
  ndim = 2,
  bn_1st = TRUE,
  act_cls = nn$ReLU,
  transpose = FALSE,
  init = "auto",
  xtra = NULL,
  bias_std = 0.01,
  dilation = 1,
  groups = 1,
  padding_mode = "zeros"
)
}
\arguments{
\item{nf}{nf}

\item{norm_type}{norm_type}

\item{ks}{ks}

\item{stride}{stride}

\item{padding}{padding}

\item{bias}{bias}

\item{ndim}{ndim}

\item{bn_1st}{bn_1st}

\item{act_cls}{act_cls}

\item{transpose}{transpose}

\item{init}{init}

\item{xtra}{xtra}

\item{bias_std}{bias_std}

\item{dilation}{dilation}

\item{groups}{groups}

\item{padding_mode}{padding_mode}
}
\value{
block
}
\description{
Resnet block of `nf` features. `conv_kwargs` are passed to `conv_layer`.
}
