% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastinference.R
\name{dependence_plot}
\alias{dependence_plot}
\title{Dependence_plot}
\usage{
dependence_plot(object, variable_name = "", class_id = 0, dpi = 200, ...)
}
\arguments{
\item{object}{ShapInterpretation object}

\item{variable_name}{the name of the column}

\item{class_id}{is used to indicate the class of interest for a classification model.
It can either be an int or str representation for a class of choice. This plot shows how the
model depends on the given variable. Vertical dispersion of the datapoints represent
interaction effects. Gray ticks along the y-axis are datapoints where the variable's values were NaN.}

\item{dpi}{dots per inch}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Plots the value of a variable on the x-axis and the SHAP value of the same
variable on the y-axis. Accepts a class_id and variable_name.
}
