% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{DeterministicFlip}
\alias{DeterministicFlip}
\title{DeterministicFlip}
\usage{
DeterministicFlip(
  size = NULL,
  mode = "bilinear",
  pad_mode = "reflection",
  align_corners = TRUE
)
}
\arguments{
\item{size}{size}

\item{mode}{mode}

\item{pad_mode}{pad_mode}

\item{align_corners}{align_corners}
}
\description{
Flip the batch every other call
}
