% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{dicom_windows}
\alias{dicom_windows}
\title{Window effect}
\format{
An object of class \code{types.SimpleNamespace} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
dicom_windows
}
\value{
module
}
\description{
Window effect
}
\keyword{datasets}
