% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R, R/vision_augment.R
\name{Dihedral}
\alias{Dihedral}
\title{Dihedral}
\usage{
Dihedral(
  p = 0.5,
  draw = NULL,
  size = NULL,
  mode = "bilinear",
  pad_mode = "reflection",
  align_corners = NULL,
  batch = FALSE
)

Dihedral(
  p = 0.5,
  draw = NULL,
  size = NULL,
  mode = "bilinear",
  pad_mode = "reflection",
  align_corners = NULL,
  batch = FALSE
)
}
\arguments{
\item{p}{probability}

\item{draw}{draw}

\item{size}{size}

\item{mode}{mode}

\item{pad_mode}{padding mode}

\item{align_corners}{align corners}

\item{batch}{batch}
}
\value{
None

None
}
\description{
Apply a random dihedral transformation to a batch of images with a probability `p`

Apply a random dihedral transformation to a batch of images with a probability `p`
}
