% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{dihedral_mat}
\alias{dihedral_mat}
\title{Dihedral_mat}
\usage{
dihedral_mat(x, p = 0.5, draw = NULL, batch = FALSE)
}
\arguments{
\item{x}{x}

\item{p}{p}

\item{draw}{draw}

\item{batch}{batch}
}
\value{
None
}
\description{
Return a random dihedral matrix
}
