% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{DihedralItem}
\alias{DihedralItem}
\title{DihedralItem}
\usage{
DihedralItem(p = 1, nm = NULL, before_call = NULL)
}
\arguments{
\item{p}{probability}

\item{nm}{nm}

\item{before_call}{before call}
}
\value{
None
}
\description{
Randomly flip with probability `p`
}
