% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{dim}
\alias{dim}
\alias{dim.fastai.torch_core.TensorMask}
\alias{dim.torch.Tensor}
\title{Dim}
\usage{
\method{dim}{fastai.torch_core.TensorMask}(x)

\method{dim}{torch.Tensor}(x)
}
\description{
Dim

Dim
}
