% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{discriminator}
\alias{discriminator}
\title{Discriminator}
\usage{
discriminator(
  ch_in,
  n_ftrs = 64,
  n_layers = 3,
  norm_layer = NULL,
  sigmoid = FALSE
)
}
\arguments{
\item{ch_in}{input}

\item{n_ftrs}{number of filters}

\item{n_layers}{number of layers}

\item{norm_layer}{normalization layer}

\item{sigmoid}{apply sigmoid  function or not}
}
\description{
Discriminator
}
