% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{div}
\alias{div}
\alias{/.torch.Tensor}
\alias{*.torch.Tensor}
\title{Div}
\usage{
\method{/}{torch.Tensor}(a, b)

\method{*}{torch.Tensor}(a, b)
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}
}
\value{
tensor

tensor
}
\description{
Div

Multiply
}
