% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_signal.R
\name{DownmixMono}
\alias{DownmixMono}
\title{Downmix Mono}
\usage{
DownmixMono(enc = NULL, dec = NULL, split_idx = NULL, order = NULL)
}
\arguments{
\item{enc}{encoder}

\item{dec}{decoder}

\item{split_idx}{split by index}

\item{order}{order}
}
\value{
None
}
\description{
Transform multichannel audios into single channel
}
