% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_hook.R
\name{dummy_eval}
\alias{dummy_eval}
\title{Dummy_eval}
\usage{
dummy_eval(m, size = list(64, 64))
}
\arguments{
\item{m}{m parameter}

\item{size}{size}
}
\value{
None
}
\description{
Evaluate `m` on a dummy input of a certain `size`
}
