% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_efficentdet.R
\name{efficientdet_infer_dl}
\alias{efficientdet_infer_dl}
\title{Efficientdet infer dataloader}
\usage{
efficientdet_infer_dl(dataset, batch_tfms = NULL, ...)
}
\arguments{
\item{dataset}{Possibly a `Dataset` object, but more generally, any `Sequence` that returns records.}

\item{batch_tfms}{Transforms to be applied at the batch level. **dataloader_kwargs: Keyword arguments that will be internally passed to a Pytorch `DataLoader`. The parameter `collate_fn` is already defined internally and cannot be passed here.}

\item{...}{additional arguments}
}
\value{
None
}
\description{
A `DataLoader` with a custom `collate_fn` that batches items as required for inferring the model.
}
