% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_models.R
\name{efficientdet_learner}
\alias{efficientdet_learner}
\title{MaskRCNN learner}
\usage{
efficientdet_learner(dls, model, cbs = NULL, ...)
}
\arguments{
\item{dls}{`Sequence` of `DataLoaders` passed to the `Learner`.
The first one will be used for training and the second for validation.}

\item{model}{The model to train.}

\item{cbs}{Optional `Sequence` of callbacks.}

\item{...}{learner_kwargs: Keyword arguments that will be internally passed to `Learner`.}
}
\value{
model
}
\description{
Fastai `Learner` adapted for MaskRCNN.
}
