% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_models.R
\name{efficientdet_model}
\alias{efficientdet_model}
\title{Eficientdet model}
\usage{
efficientdet_model(model_name, num_classes, img_size, pretrained = TRUE)
}
\arguments{
\item{model_name}{Specifies the model to create. For pretrained models, check [this](https://github.com/rwightman/efficientdet-pytorch#models) table.}

\item{num_classes}{Number of classes of your dataset (including background).}

\item{img_size}{Image size that will be fed to the model. Must be squared and divisible by 128.}

\item{pretrained}{If TRUE, use a pretrained backbone (on COCO).}
}
\value{
model
}
\description{
Creates the efficientdet model specified by `model_name`.
}
