% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_efficentdet.R
\name{efficientdet_predict_dl}
\alias{efficientdet_predict_dl}
\title{Efficientdet predict dataloader}
\usage{
efficientdet_predict_dl(model, infer_dl, show_pbar = TRUE)
}
\arguments{
\item{model}{model}

\item{infer_dl}{infer_dl}

\item{show_pbar}{show_pbar}
}
\value{
None
}
\description{
Efficientdet predict dataloader
}
