% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{Embedding}
\alias{Embedding}
\title{Embedding}
\usage{
Embedding(ni, nf)
}
\arguments{
\item{ni}{input}

\item{nf}{output}
}
\value{
None
}
\description{
Embedding layer with truncated normal initialization
}
