% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_models.R
\name{EmbeddingDropout}
\alias{EmbeddingDropout}
\title{EmbeddingDropout}
\usage{
EmbeddingDropout(emb, embed_p)
}
\arguments{
\item{emb}{emb}

\item{embed_p}{embed_p}
}
\value{
None
}
\description{
Apply dropout with probability `embed_p` to an embedding layer `emb`.
}
