% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{==.torch.Tensor}
\alias{==.torch.Tensor}
\title{Compares two tensors if equal}
\usage{
\method{==}{torch.Tensor}(a, b)
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}
}
\value{
A tensor of booleans, where False corresponds to 0, and 1 to True
in a tensor of data type \code{torch$bool}.
}
\description{
This generic is approximately similar to \code{torch$eq(a, b)}, with the
difference that the generic returns a tensor of booleans instead of
a tensor of data type \code{torch$uint8}.
}
