% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{exp}
\alias{exp}
\alias{exp.fastai.torch_core.TensorMask}
\alias{exp.torch.Tensor}
\title{Exp}
\usage{
\method{exp}{fastai.torch_core.TensorMask}(a, b)

\method{exp}{torch.Tensor}(a, b)
}
\description{
Exp

Exp
}
