% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{expm1}
\alias{expm1}
\alias{expm1.fastai.torch_core.TensorMask}
\alias{expm1.torch.Tensor}
\title{Expm1}
\usage{
\method{expm1}{fastai.torch_core.TensorMask}(a, b)

\method{expm1}{torch.Tensor}(a, b)
}
\description{
Expm1

Expm1
}
