% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{export_generator}
\alias{export_generator}
\title{Export_generator}
\usage{
export_generator(
  learn,
  generator_name = "generator",
  path = ".",
  convert_to = "B"
)
}
\arguments{
\item{learn}{learner/model}

\item{generator_name}{generator name}

\item{path}{path (save dir)}

\item{convert_to}{convert to}
}
\value{
None
}
\description{
Export_generator
}
