% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{F}
\alias{F}
\title{Functional interface}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
F
}
\value{
module
}
\description{
Functional interface
}
\keyword{datasets}
