% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{F1Score}
\alias{F1Score}
\title{F1Score}
\usage{
F1Score(
  axis = -1,
  labels = NULL,
  pos_label = 1,
  average = "binary",
  sample_weight = NULL
)
}
\arguments{
\item{axis}{axis}

\item{labels}{labels}

\item{pos_label}{pos_label}

\item{average}{average}

\item{sample_weight}{sample_weight}
}
\value{
None
}
\description{
F1 score for single-label classification problems
}
