% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_faster_rcnn.R
\name{faster_rcnn_predict_dl}
\alias{faster_rcnn_predict_dl}
\title{Faster RCNN predict dataloader}
\usage{
faster_rcnn_predict_dl(model, infer_dl, show_pbar = TRUE)
}
\arguments{
\item{model}{model}

\item{infer_dl}{infer_dl}

\item{show_pbar}{show_pbar}
}
\value{
None
}
\description{
Faster RCNN predict dataloader
}
