% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_dataloaders.R
\name{faster_rcnn_valid_dl}
\alias{faster_rcnn_valid_dl}
\title{Faster RSNN valid dataloader}
\usage{
faster_rcnn_valid_dl(dataset, batch_tfms = NULL, ...)
}
\arguments{
\item{dataset}{Possibly a `Dataset` object, but more generally, any `Sequence` that returns records.}

\item{batch_tfms}{Transforms to be applied at the batch level.}

\item{...}{dataloader_kwargs: Keyword arguments that will be internally passed to a Pytorch `DataLoader`. The parameter `collate_fn` is already defined internally and cannot be passed here.}
}
\value{
None
}
\description{
A `DataLoader` with a custom `collate_fn` that batches items as required for training the model.
}
