% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{FBeta}
\alias{FBeta}
\title{FBeta}
\usage{
FBeta(
  beta,
  axis = -1,
  labels = NULL,
  pos_label = 1,
  average = "binary",
  sample_weight = NULL
)
}
\arguments{
\item{beta}{beta}

\item{axis}{axis}

\item{labels}{labels}

\item{pos_label}{pos_label}

\item{average}{average}

\item{sample_weight}{sample_weight}
}
\value{
None
}
\description{
FBeta score with `beta` for single-label classification problems
}
