% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{FetchPredsCallback}
\alias{FetchPredsCallback}
\title{FetchPredsCallback}
\usage{
FetchPredsCallback(
  ds_idx = 1,
  dl = NULL,
  with_input = FALSE,
  with_decoded = FALSE,
  cbs = NULL,
  reorder = TRUE
)
}
\arguments{
\item{ds_idx}{ds_idx}

\item{dl}{dl}

\item{with_input}{with_input}

\item{with_decoded}{with_decoded}

\item{cbs}{cbs}

\item{reorder}{reorder}
}
\description{
A callback to fetch predictions during the training loop
}
