% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_transform.R
\name{FillMissing}
\alias{FillMissing}
\title{Fill Missing}
\usage{
FillMissing(
  cat_names,
  cont_names,
  fill_strategy = FillStrategy_MEDIAN(),
  add_col = TRUE,
  fill_val = 0
)
}
\arguments{
\item{cat_names}{cat_names}

\item{cont_names}{cont_names}

\item{fill_strategy}{fill_strategy}

\item{add_col}{add_col}

\item{fill_val}{fill_val}
}
\description{
Fill the missing values in continuous columns.
}
