% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner_fits.R
\name{fine_tune}
\alias{fine_tune}
\title{Fine_tune}
\usage{
fine_tune(
  object,
  epochs,
  base_lr = 0.002,
  freeze_epochs = 1,
  lr_mult = 100,
  pct_start = 0.3,
  div = 5,
  ...
)
}
\arguments{
\item{object}{learner/model}

\item{epochs}{epoch number}

\item{base_lr}{base learning rate}

\item{freeze_epochs}{freeze epochs number}

\item{lr_mult}{learning rate multiply}

\item{pct_start}{start percentage}

\item{div}{divide}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Fine tune with `freeze` for `freeze_epochs` then
with `unfreeze` from `epochs` using discriminative LR
}
