% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{fit.fastai.basic_train.Learner}
\alias{fit.fastai.basic_train.Learner}
\title{Fit}
\usage{
\method{fit}{fastai.basic_train.Learner}(object, epochs, lr = 0.01, wd = NULL, callbacks = NULL)
}
\arguments{
\item{epochs}{epochs}

\item{lr}{lr}

\item{wd}{wd}

\item{callbacks}{callbacks}
}
\description{
Fit the model on this learner with `lr` learning rate, `wd` weight decay for `epochs` with `callbacks`.
}
