% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{fit.fastai.vision.gan.GANLearner}
\alias{fit.fastai.vision.gan.GANLearner}
\title{Fit}
\usage{
\method{fit}{fastai.vision.gan.GANLearner}(object, ...)
}
\arguments{
\item{object}{model}

\item{...}{additonal parameters to pass}
}
\value{
train history
}
\description{
Fit the model on this learner with `lr` learning rate, `wd` weight decay for `epochs` with `callbacks`.
}
\examples{
\dontrun{

learn \%>\% fit(1, 2e-4, wd = 0)

}

}
