% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_learner.R
\name{fit.fastai2.learner.Learner}
\alias{fit.fastai2.learner.Learner}
\title{Fit}
\usage{
\method{fit}{fastai2.learner.Learner}(object, n_epoch, lr = 0.01, wd = NULL, callbacks = NULL)
}
\arguments{
\item{lr}{lr}

\item{wd}{wd}

\item{callbacks}{callbacks}

\item{epochs}{epochs}
}
\description{
Fit the model on this learner with `lr` learning rate, `wd` weight decay for `epochs` with `callbacks`.
}
