% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner_fits.R
\name{fit_flat_cos}
\alias{fit_flat_cos}
\title{Fit_flat_cos}
\usage{
fit_flat_cos(
  object,
  n_epoch,
  lr = NULL,
  div_final = 1e+05,
  pct_start = 0.75,
  wd = NULL,
  cbs = NULL,
  reset_opt = FALSE
)
}
\arguments{
\item{object}{learner/model}

\item{n_epoch}{number of epochs}

\item{lr}{learning rate}

\item{div_final}{divide final value}

\item{pct_start}{start percentage}

\item{wd}{weight decay}

\item{cbs}{callbacks}

\item{reset_ops}{reset optimizer}
}
\value{
None
}
\description{
Fit_flat_cos
}
