% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner_fits.R
\name{fit_flat_lin}
\alias{fit_flat_lin}
\title{Fit_flat_lin}
\usage{
fit_flat_lin(
  object,
  n_epochs = 100,
  n_epochs_decay = 100,
  start_lr = NULL,
  end_lr = 0,
  curve_type = "linear",
  wd = NULL,
  cbs = NULL,
  reset_opt = FALSE
)
}
\arguments{
\item{object}{model / learner}

\item{n_epochs}{number of epochs}

\item{n_epochs_decay}{number of epochs with decay}

\item{start_lr}{Desired starting learning rate, used for beginning pct of training.}

\item{end_lr}{Desired end learning rate, training will conclude at this learning rate.}

\item{curve_type}{Curve type for learning rate annealing. Options are 'linear', 'cosine', and 'exponential'.}

\item{wd}{weight decay}

\item{cbs}{callbacks}

\item{reset_opt}{reset optimizer}
}
\value{
None
}
\description{
Fit `self.model` for `n_epoch` at flat `start_lr`
before `curve_type` annealing to `end_lr` with weight decay of `wd` and
callbacks `cbs`.
}
