% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner_fits.R
\name{fit_sgdr}
\alias{fit_sgdr}
\title{Fit_sgdr}
\usage{
fit_sgdr(
  object,
  n_cycles,
  cycle_len,
  lr_max = NULL,
  cycle_mult = 2,
  cbs = NULL,
  reset_opt = FALSE,
  wd = NULL
)
}
\arguments{
\item{object}{learner/model}

\item{n_cycles}{number of cycles}

\item{cycle_len}{length of cycle}

\item{lr_max}{maximum learning rate}

\item{cycle_mult}{cycle mult}

\item{cbs}{callbacks}

\item{reset_opt}{reset optimizer}

\item{wd}{weight decay}
}
\value{
None
}
\description{
Fit_sgdr
}
