% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{Flatten}
\alias{Flatten}
\title{Flatten}
\usage{
Flatten(full = FALSE)
}
\arguments{
\item{full}{bool, full or not}
}
\description{
Flatten `x` to a single dimension, e.g. at end of a model. `full` for rank-1 tensor
}
