% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{flatten_check}
\alias{flatten_check}
\title{Flatten check}
\usage{
flatten_check(inp, targ)
}
\arguments{
\item{inp}{predictions}

\item{targ}{targets}
}
\value{
tensor
}
\description{
Check that `out` and `targ` have the same number of elements and flatten them.
}
