% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{flatten_model}
\alias{flatten_model}
\title{Flatten_model}
\usage{
flatten_model(m)
}
\arguments{
\item{m}{parameters}
}
\value{
None
}
\description{
Return the list of all submodules and parameters of `m`
}
