% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R
\name{Flip}
\alias{Flip}
\title{Flip}
\usage{
Flip(
  p = 0.5,
  draw = NULL,
  size = NULL,
  mode = "bilinear",
  pad_mode = "reflection",
  align_corners = TRUE,
  batch = FALSE
)
}
\arguments{
\item{p}{probability}

\item{draw}{draw}

\item{size}{size of image}

\item{mode}{mode}

\item{pad_mode}{reflection, zeros, border as string parameter}

\item{align_corners}{align corners ot not}

\item{batch}{batch or not}
}
\value{
None
}
\description{
Randomly flip a batch of images with a probability `p`
}
