% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{flip_mat}
\alias{flip_mat}
\title{flip_mat}
\usage{
flip_mat(x, p = 0.5, draw = NULL, batch = FALSE)
}
\arguments{
\item{x}{x}

\item{p}{p}

\item{draw}{draw}

\item{batch}{batch}
}
\description{
Return a random flip matrix
}
