% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{floor_div}
\alias{floor_div}
\alias{\%/\%.fastai.torch_core.TensorMask}
\alias{\%/\%.torch.Tensor}
\title{Floor divide}
\usage{
\method{\%/\%}{fastai.torch_core.TensorMask}(x, y)

\method{\%/\%}{torch.Tensor}(x, y)
}
\description{
Floor divide

Floor divide
}
