% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{fmodule}
\alias{fmodule}
\title{Module}
\usage{
fmodule(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
Module
}
\details{
Decorator to create an nn$Module using f as forward method
}
