% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{FolderDataset}
\alias{FolderDataset}
\title{FolderDataset}
\usage{
FolderDataset(path, transforms = NULL)
}
\arguments{
\item{path}{path to dir}

\item{transforms}{transformations}
}
\value{
None
}
\description{
A PyTorch Dataset class that can be created from a folder `path` of images, for the sole purpose of inference. Optional `transforms`
}
\details{
can be provided. Attributes: `self.files`: A list of the filenames in the folder. `self.totensor`: `torchvision.transforms.ToTensor` transform. `self.transform`: The transforms passed in as `transforms` to the constructor.
}
