% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastinference.R
\name{force_plot}
\alias{force_plot}
\title{Force_plot}
\usage{
force_plot(object, class_id = 0, ...)
}
\arguments{
\item{object}{ShapInterpretation object}

\item{class_id}{Accepts a class_id which is used to indicate the class of interest for a
classification model. It can either be an int or str representation for a class of choice.}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Visualizes the SHAP values with an added force layout. Accepts a class_id which
is used to indicate the class of interest for a classification model.
}
