% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{foreground_acc}
\alias{foreground_acc}
\title{foreground_acc}
\usage{
foreground_acc(inp, targ, bkg_idx = 0, axis = 1)
}
\arguments{
\item{inp}{inp}

\item{targ}{targ}

\item{bkg_idx}{bkg_idx}

\item{axis}{axis}
}
\description{
Computes non-background accuracy for multiclass segmentation
}
\details{

}
