% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_models.R
\name{forget_mult_CPU}
\alias{forget_mult_CPU}
\title{forget_mult_CPU}
\usage{
forget_mult_CPU(x, f, first_h = NULL, batch_first = TRUE, backward = FALSE)
}
\arguments{
\item{x}{x}

\item{f}{f}

\item{first_h}{first_h}

\item{batch_first}{batch_first}

\item{backward}{backward}
}
\description{
ForgetMult gate applied to `x` and `f` on the CPU.
}
