% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_models.R
\name{ForgetMultGPU}
\alias{ForgetMultGPU}
\title{ForgetMultGPU}
\usage{
ForgetMultGPU(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
Wrapper around the CUDA kernels for the ForgetMult gate.
}
