% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeze_and_helpers.R
\name{freeze}
\alias{freeze}
\title{Freeze a model}
\usage{
freeze(object, ...)
}
\arguments{
\item{object}{A model}

\item{...}{Additional parameters}
}
\value{
None
}
\description{
Freeze a model
}
\examples{
\dontrun{
learnR \%>\% freeze()
}
}
