% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{from_csv}
\alias{from_csv}
\title{from_csv}
\usage{
from_csv(
  path,
  csv_fname = "labels.csv",
  header = "infer",
  delimiter = NULL,
  valid_pct = 0.2,
  seed = NULL,
  text_col = 0,
  label_col = 1,
  label_delim = NULL,
  y_block = NULL,
  text_vocab = NULL,
  is_lm = FALSE,
  valid_col = NULL,
  tok_tfm = NULL,
  seq_len = 72,
  backwards = FALSE,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  device = NULL
)
}
\arguments{
\item{path}{path}

\item{csv_fname}{csv_fname}

\item{header}{header}

\item{delimiter}{delimiter}

\item{valid_pct}{valid_pct}

\item{seed}{seed}

\item{text_col}{text_col}

\item{label_col}{label_col}

\item{label_delim}{label_delim}

\item{y_block}{y_block}

\item{text_vocab}{text_vocab}

\item{is_lm}{is_lm}

\item{valid_col}{valid_col}

\item{tok_tfm}{tok_tfm}

\item{seq_len}{seq_len}

\item{backwards}{backwards}

\item{bs}{bs}

\item{val_bs}{val_bs}

\item{shuffle_train}{shuffle_train}

\item{device}{device}

\item{cls}{cls}
}
\description{
Create from `csv` file in `path/csv_fname`
}
\details{

}
