% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{FuncSplitter}
\alias{FuncSplitter}
\title{FuncSplitter}
\usage{
FuncSplitter(func)
}
\arguments{
\item{func}{func}
}
\description{
Split `items` by result of `func` (`TRUE` for validation, `FALSE` for training set).
}
