% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{gan_critic}
\alias{gan_critic}
\title{gan_critic}
\usage{
gan_critic(n_channels = 3, nf = 128, n_blocks = 3, p = 0.15)
}
\arguments{
\item{n_channels}{n_channels}

\item{nf}{nf}

\item{n_blocks}{n_blocks}

\item{p}{p}
}
\description{
Critic to train a `GAN`.
}
