% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{gan_loss_from_func}
\alias{gan_loss_from_func}
\title{gan_loss_from_func}
\usage{
gan_loss_from_func(loss_gen, loss_crit, weights_gen = NULL)
}
\arguments{
\item{loss_gen}{loss_gen}

\item{loss_crit}{loss_crit}

\item{weights_gen}{weights_gen}
}
\description{
Define loss functions for a GAN from `loss_gen` and `loss_crit`.
}
