% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{GANLearner_from_learners}
\alias{GANLearner_from_learners}
\title{GANLearner_from_learners}
\usage{
GANLearner_from_learners(
  gen_learn,
  crit_learn,
  switcher = NULL,
  weights_gen = NULL,
  gen_first = FALSE,
  switch_eval = TRUE,
  show_img = TRUE,
  clip = NULL,
  cbs = NULL,
  metrics = NULL,
  loss_func = NULL,
  opt_func = Adam(),
  lr = 0.001,
  splitter = trainable_params(),
  path = NULL,
  model_dir = "models",
  wd = NULL,
  wd_bn_bias = FALSE,
  train_bn = TRUE,
  moms = list(0.95, 0.85, 0.95)
)
}
\arguments{
\item{gen_learn}{gen_learn}

\item{crit_learn}{crit_learn}

\item{switcher}{switcher}

\item{weights_gen}{weights_gen}

\item{gen_first}{gen_first}

\item{switch_eval}{switch_eval}

\item{show_img}{show_img}

\item{clip}{clip}

\item{cbs}{cbs}

\item{metrics}{metrics}

\item{loss_func}{loss_func}

\item{opt_func}{opt_func}

\item{lr}{lr}

\item{splitter}{splitter}

\item{path}{path}

\item{model_dir}{model_dir}

\item{wd}{wd}

\item{wd_bn_bias}{wd_bn_bias}

\item{train_bn}{train_bn}

\item{moms}{moms}
}
\description{
Create a GAN from `learn_gen` and `learn_crit`.
}
