% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{GANModule}
\alias{GANModule}
\title{GAN Module}
\usage{
GANModule(generator = NULL, critic = NULL, gen_mode = FALSE)
}
\arguments{
\item{generator}{generator}

\item{critic}{critic}

\item{gen_mode}{gen_mode}
}
\value{
None
}
\description{
Wrapper around a `generator` and a `critic` to create a GAN.
}
