% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{GANTrainer}
\alias{GANTrainer}
\title{GAN Trainer}
\usage{
GANTrainer(
  switch_eval = FALSE,
  clip = NULL,
  beta = 0.98,
  gen_first = FALSE,
  show_img = TRUE
)
}
\arguments{
\item{switch_eval}{switch evaluation}

\item{clip}{clip value}

\item{beta}{beta parameter}

\item{gen_first}{generator first}

\item{show_img}{show image or not}
}
\value{
None
}
\description{
Handles GAN Training.
}
