% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{GatherPredsCallback}
\alias{GatherPredsCallback}
\title{GatherPredsCallback}
\usage{
GatherPredsCallback(
  with_input = FALSE,
  with_loss = FALSE,
  save_preds = NULL,
  save_targs = NULL,
  concat_dim = 0
)
}
\arguments{
\item{with_input}{with_input}

\item{with_loss}{with_loss}

\item{save_preds}{save_preds}

\item{save_targs}{save_targs}

\item{concat_dim}{concat_dim}
}
\description{
`Callback` that saves the predictions and targets, optionally `with_loss`
}
\details{

}
