% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_signal.R
\name{get_audio_files}
\alias{get_audio_files}
\title{Get_audio_files}
\usage{
get_audio_files(path, recurse = TRUE, folders = NULL)
}
\arguments{
\item{path}{path}

\item{recurse}{recurse}

\item{folders}{folders}
}
\value{
None
}
\description{
Get audio files in `path` recursively, only in `folders`, if specified.
}
