% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_confusion_matrix}
\alias{get_confusion_matrix}
\title{Extract confusion matrix}
\usage{
get_confusion_matrix(object)
}
\arguments{
\item{object}{model}
}
\value{
matrix
}
\description{
Extract confusion matrix
}
