% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_modules.R
\name{get_data_loaders}
\alias{get_data_loaders}
\title{Get data loaders}
\usage{
get_data_loaders(train_batch_size, val_batch_size)
}
\arguments{
\item{train_batch_size}{train dataset batch size}

\item{val_batch_size}{validation dataset batch size}
}
\value{
None
}
\description{
Get data loaders
}
