% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medical.R
\name{get_dcm_matrix}
\alias{get_dcm_matrix}
\title{Get image matrix}
\usage{
get_dcm_matrix(img, type = "raw", scan = "", size = 50, convert = TRUE)
}
\arguments{
\item{img}{dicom file}

\item{type}{img transformation}
}
\description{
Get image matrix
}
