% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{get_dls}
\alias{get_dls}
\title{Get dls}
\usage{
get_dls(
  pathA,
  pathB,
  num_A = NULL,
  num_B = NULL,
  load_size = 512,
  crop_size = 256,
  bs = 4,
  num_workers = 2
)
}
\arguments{
\item{pathA}{path A (from domain)}

\item{pathB}{path B (to domain)}

\item{num_A}{subset of A data}

\item{num_B}{subset of B data}

\item{load_size}{load size}

\item{crop_size}{crop size}

\item{bs}{bathc size}

\item{num_workers}{number of workers}
}
\value{
None
}
\description{
Given image files from two domains (`pathA`, `pathB`), create `DataLoaders` object.
}
\details{
Loading and randomly cropped sizes of `load_size` and `crop_size` are set to defaults of 512 and 256.
Batch size is specified by `bs` (default=4).
}
