% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_files}
\alias{get_files}
\title{get_files}
\usage{
get_files(
  path,
  extensions = NULL,
  recurse = TRUE,
  folders = NULL,
  followlinks = TRUE
)
}
\arguments{
\item{path}{path}

\item{extensions}{extensions}

\item{recurse}{recurse}

\item{folders}{folders}

\item{followlinks}{followlinks}
}
\description{
Get all the files in `path` with optional `extensions`, optionally with `recurse`, only in `folders`, if specified.
}
