% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_data.R
\name{get_grid}
\alias{get_grid}
\title{Get_grid}
\usage{
get_grid(
  n,
  nrows = NULL,
  ncols = NULL,
  add_vert = 0,
  figsize = NULL,
  double = FALSE,
  title = NULL,
  return_fig = FALSE,
  imsize = 3
)
}
\arguments{
\item{n}{n}

\item{nrows}{number of rows}

\item{ncols}{number of columns}

\item{add_vert}{add vertical}

\item{figsize}{figure size}

\item{double}{double}

\item{title}{title}

\item{return_fig}{return figure or not}

\item{imsize}{image size}
}
\value{
None
}
\description{
Return a grid of `n` axes, `rows` by `cols`
}
