% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{get_hf_objects}
\alias{get_hf_objects}
\title{Get_hf_objects}
\usage{
get_hf_objects(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
Returns the architecture (str), config (obj), tokenizer (obj), and model (obj) given at minimum a
}
\details{
`pre-trained model name or path`. Specify a `task` to ensure the right "AutoModelFor<task>" is used to
create the model. Optionally, you can pass a config (obj), tokenizer (class), and/or model (class) (along with any
related kwargs for each) to get as specific as you want w/r/t what huggingface objects are returned.
}
