% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loaders.R
\name{get_image_files}
\alias{get_image_files}
\title{Get_image_files}
\usage{
get_image_files(path, recurse = TRUE, folders = NULL)
}
\arguments{
\item{path}{path}

\item{recurse}{recurse}

\item{folders}{folders}
}
\description{
Get image files in `path` recursively, only in `folders`, if specified.
}
