% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_models.R
\name{get_language_model}
\alias{get_language_model}
\title{Get_language_model}
\usage{
get_language_model(arch, vocab_sz, config = NULL, drop_mult = 1)
}
\arguments{
\item{arch}{arch}

\item{vocab_sz}{vocab_sz}

\item{config}{config}

\item{drop_mult}{drop_mult}
}
\description{
Create a language model from `arch` and its `config`.
}
