% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{get_preds_cyclegan}
\alias{get_preds_cyclegan}
\title{Get_preds_cyclegan}
\usage{
get_preds_cyclegan(
  learn,
  test_path,
  pred_path,
  bs = 4,
  num_workers = 4,
  suffix = "tif"
)
}
\arguments{
\item{learn}{learner/model}

\item{test_path}{testdat path}

\item{pred_path}{predict data path}

\item{bs}{batch size}

\item{num_workers}{number of workers}

\item{suffix}{suffix}
}
\description{
A prediction function that takes the Learner object `learn` with the trained model, the `test_path` folder with the images to perform
}
\details{
batch inference on, and the output folder `pred_path` where the predictions will be saved, with a batch size `bs`, `num_workers`,
and suffix of the prediction images `suffix` (default='png').
}
