% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_models.R
\name{get_text_classifier}
\alias{get_text_classifier}
\title{get_text_classifier}
\usage{
get_text_classifier(
  arch,
  vocab_sz,
  n_class,
  seq_len = 72,
  config = NULL,
  drop_mult = 1,
  lin_ftrs = NULL,
  ps = NULL,
  pad_idx = 1,
  max_len = 1440,
  y_range = NULL
)
}
\arguments{
\item{arch}{arch}

\item{vocab_sz}{vocab_sz}

\item{n_class}{n_class}

\item{seq_len}{seq_len}

\item{config}{config}

\item{drop_mult}{drop_mult}

\item{lin_ftrs}{lin_ftrs}

\item{ps}{ps}

\item{pad_idx}{pad_idx}

\item{max_len}{max_len}

\item{y_range}{y_range}
}
\description{
Create a text classifier from `arch` and its `config`, maybe `pretrained`
}
