% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_models.R
\name{get_text_files}
\alias{get_text_files}
\title{Get_text_files}
\usage{
get_text_files(path, recurse = TRUE, folders = NULL)
}
\arguments{
\item{path}{path}

\item{recurse}{recurse}

\item{folders}{folders}
}
\value{
None
}
\description{
Get text files in `path` recursively, only in `folders`, if specified.
}
