% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab.R
\name{get_weights}
\alias{get_weights}
\title{Get weights}
\usage{
get_weights(object, arr, is_item = TRUE, convert = FALSE)
}
\arguments{
\item{object}{object}

\item{arr}{arr}

\item{is_item}{is_item}

\item{convert}{to matrix}
}
\description{
Weight for item or user (based on `is_item`) for all in `arr`
}
