% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_training.R
\name{GradientAccumulation}
\alias{GradientAccumulation}
\title{GradientAccumulation}
\usage{
GradientAccumulation(n_acc = 32)
}
\arguments{
\item{n_acc}{number of acc}
}
\value{
None
}
\description{
Accumulate gradients before updating weights
}
