% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitters.R
\name{GrandparentSplitter}
\alias{GrandparentSplitter}
\title{GrandparentSplitter}
\usage{
GrandparentSplitter(train_name = "train", valid_name = "valid")
}
\arguments{
\item{train_name}{train folder name}

\item{valid_name}{validation folder name}
}
\value{
None
}
\description{
Split `items` from the grand parent folder names (`train_name` and `valid_name`).
}
