% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{\%f\%}
\alias{\%f\%}
\title{Fastai assignment}
\usage{
left \%f\% right
}
\arguments{
\item{left}{left side object}

\item{right}{right side object}
}
\value{
None
}
\description{
The assignment has to be used for safe modification of the values inside tensors/layers
}
