% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{\%\%.fastai.torch_core.TensorMask}
\alias{\%\%.fastai.torch_core.TensorMask}
\title{Floor mod}
\usage{
\method{\%\%}{fastai.torch_core.TensorMask}(x, y)
}
\arguments{
\item{x}{tensor}

\item{y}{tensor}
}
\value{
tensor
}
\description{
Floor mod
}
