% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{greater}
\alias{greater}
\alias{>.fastai.torch_core.TensorMask}
\alias{<.fastai.torch_core.TensorMask}
\alias{>.torch.Tensor}
\alias{<.torch.Tensor}
\title{Greater}
\usage{
\method{>}{fastai.torch_core.TensorMask}(a, b)

\method{<}{fastai.torch_core.TensorMask}(a, b)

\method{>}{torch.Tensor}(a, b)

\method{<}{torch.Tensor}(a, b)
}
\description{
Greater

Less

Greater

Less
}
