% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{greater_or_equal}
\alias{greater_or_equal}
\alias{>=.fastai.torch_core.TensorMask}
\alias{>=.torch.Tensor}
\title{Greater or equal}
\usage{
\method{>=}{fastai.torch_core.TensorMask}(a, b)

\method{>=}{torch.Tensor}(a, b)
}
\description{
Greater or equal

Greater or equal
}
