% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{HammingLoss}
\alias{HammingLoss}
\title{HammingLoss}
\usage{
HammingLoss(axis = -1, sample_weight = NULL)
}
\arguments{
\item{axis}{axis}

\item{sample_weight}{sample_weight}
}
\description{
Hamming loss for single-label classification problems
}
