% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{HF_BaseInput}
\alias{HF_BaseInput}
\title{HF_BaseInput}
\usage{
HF_BaseInput(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
A HF_BaseInput object is returned from the decodes method
of HF_BatchTransform as a mean to customize `@typedispatched` functions
like DataLoaders.show_batch and Learner.show_results. It represents the
"input_ids" of a huggingface sequence as a tensor with a show method that
requires a huggingface tokenizer for proper display.
}
