% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{HF_BaseModelCallback}
\alias{HF_BaseModelCallback}
\title{HF_BaseModelCallback}
\usage{
HF_BaseModelCallback(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
HF_BaseModelCallback
}
