% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{HF_BaseModelWrapper}
\alias{HF_BaseModelWrapper}
\title{HF_BaseModelWrapper}
\usage{
HF_BaseModelWrapper(
  hf_model,
  output_hidden_states = FALSE,
  output_attentions = FALSE,
  ...
)
}
\arguments{
\item{hf_model}{model}

\item{output_hidden_states}{output hidden states}

\item{output_attentions}{output attentions}

\item{...}{additional arguments to pass}
}
\value{
None
}
\description{
Same as `nn.Module`, but no need for subclasses to call `super().__init__`
}
