% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{HF_QstAndAnsModelCallback}
\alias{HF_QstAndAnsModelCallback}
\title{HF_QstAndAnsModelCallback}
\usage{
HF_QstAndAnsModelCallback(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
HF_QstAndAnsModelCallback
}
