% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{HF_QuestionAnswerInput}
\alias{HF_QuestionAnswerInput}
\title{HF_QuestionAnswerInput}
\usage{
HF_QuestionAnswerInput(...)
}
\arguments{
\item{...}{parameters to apss}
}
\value{
None
}
\description{
HF_QuestionAnswerInput
}
