% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{hf_splitter}
\alias{hf_splitter}
\title{Hf_splitter}
\usage{
hf_splitter(m)
}
\arguments{
\item{m}{parameters}
}
\value{
None
}
\description{
Splits the huggingface model based on various model architecture conventions
}
