% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{HF_SummarizationModelCallback}
\alias{HF_SummarizationModelCallback}
\title{HF_SummarizationModelCallback}
\usage{
HF_SummarizationModelCallback(
  rouge_metrics = c("rouge1", "rouge2", "rougeL"),
  ignore_token_id = -100,
  ...
)
}
\arguments{
\item{rouge_metrics}{rouge metrics}

\item{ignore_token_id}{integer, ignore token id}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Basic class handling tweaks of the training loop by changing a `Learner` in various events
}
