% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{HF_Text2TextAfterBatchTransform}
\alias{HF_Text2TextAfterBatchTransform}
\title{HF_Text2TextAfterBatchTransform}
\usage{
HF_Text2TextAfterBatchTransform(
  hf_tokenizer,
  input_return_type = HF_BaseInput()
)
}
\arguments{
\item{hf_tokenizer}{tokenizer}

\item{input_return_type}{input return type}
}
\value{
None
}
\description{
Delegates (`__call__`,`decode`,`setup`) to (<code>encodes</code>,<code>decodes</code>,<code>setups</code>) if `split_idx` matches
}
