% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{HF_Text2TextBlock}
\alias{HF_Text2TextBlock}
\title{HF_Text2TextBlock}
\usage{
HF_Text2TextBlock(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
A basic wrapper that links defaults transforms for the data block API
}
