% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{HF_TokenCategorize}
\alias{HF_TokenCategorize}
\title{HF_TokenCategorize}
\usage{
HF_TokenCategorize(vocab = NULL, ignore_token = NULL, ignore_token_id = NULL)
}
\arguments{
\item{vocab}{vocabulary}

\item{ignore_token}{ignore token}

\item{ignore_token_id}{ignore token id}
}
\value{
None
}
\description{
Reversible transform of a list of category string to `vocab` id
}
