% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{HF_TokenCategoryBlock}
\alias{HF_TokenCategoryBlock}
\title{HF_TokenCategoryBlock}
\usage{
HF_TokenCategoryBlock(
  vocab = NULL,
  ignore_token = NULL,
  ignore_token_id = NULL
)
}
\arguments{
\item{vocab}{vocabulary}

\item{ignore_token}{ignore token}

\item{ignore_token_id}{ignore token id}
}
\value{
None
}
\description{
`TransformBlock` for single-label categorical targets
}
