% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{HF_TokenClassBeforeBatchTransform}
\alias{HF_TokenClassBeforeBatchTransform}
\title{HF_TokenClassBeforeBatchTransform}
\usage{
HF_TokenClassBeforeBatchTransform(
  hf_arch,
  hf_tokenizer,
  ignore_token_id = -100,
  max_length = NULL,
  padding = TRUE,
  truncation = TRUE,
  is_split_into_words = TRUE,
  n_tok_inps = 1,
  ...
)
}
\arguments{
\item{hf_arch}{architecture}

\item{hf_tokenizer}{tokenizer}

\item{ignore_token_id}{ignore token id}

\item{max_length}{maximum length}

\item{padding}{padding or not}

\item{truncation}{truncation or not}

\item{is_split_into_words}{to split into_words}

\item{n_tok_inps}{number tok inputs}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Handles everything you need to assemble a mini-batch of inputs and targets,
as well as decode the dictionary produced
}
\details{
as a byproduct of the tokenization process in the `encodes` method.
}
