% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_hook.R
\name{Hook}
\alias{Hook}
\title{Hook}
\usage{
Hook(
  m,
  hook_func,
  is_forward = TRUE,
  detach = TRUE,
  cpu = FALSE,
  gather = FALSE
)
}
\arguments{
\item{m}{m aprameter}

\item{hook_func}{hook function}

\item{is_forward}{is_forward or not}

\item{detach}{detach or not}

\item{cpu}{cpu or not}

\item{gather}{gather or not}
}
\value{
None
}
\description{
Create a hook on `m` with `hook_func`.
}
\details{
Hooks are functions you can attach to a particular layer in your
model and that will be executed in the forward pass (for forward hooks)
or backward pass (for backward hooks).
}
