% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_hook.R
\name{hook_output}
\alias{hook_output}
\title{Hook_output}
\usage{
hook_output(module, detach = TRUE, cpu = FALSE, grad = FALSE)
}
\arguments{
\item{module}{module}

\item{detach}{detach or not}

\item{cpu}{cpu or not}

\item{grad}{grad or not}
}
\value{
None
}
\description{
Return a `Hook` that stores activations of `module` in `self$stored`
}
