% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_hook.R
\name{hook_outputs}
\alias{hook_outputs}
\title{Hook_outputs}
\usage{
hook_outputs(modules, detach = TRUE, cpu = FALSE, grad = FALSE)
}
\arguments{
\item{modules}{modules}

\item{detach}{detach or not}

\item{cpu}{cpu or not}

\item{grad}{grad or not}
}
\value{
None
}
\description{
Return `Hooks` that store activations of all `modules` in `self.stored`
}
