% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{HookCallback}
\alias{HookCallback}
\title{HookCallback}
\usage{
HookCallback(
  modules = NULL,
  every = NULL,
  remove_end = TRUE,
  is_forward = TRUE,
  detach = TRUE,
  cpu = TRUE
)
}
\arguments{
\item{modules}{modules}

\item{every}{every}

\item{remove_end}{remove_end}

\item{is_forward}{is_forward}

\item{detach}{detach}

\item{cpu}{cpu}
}
\description{
`Callback` that can be used to register hooks on `modules`
}
\details{

}
