% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_hook.R
\name{Hooks}
\alias{Hooks}
\title{Hooks}
\usage{
Hooks(ms, hook_func, is_forward = TRUE, detach = TRUE, cpu = FALSE)
}
\arguments{
\item{ms}{ms parameter}

\item{hook_func}{hook function}

\item{is_forward}{is_forward or not}

\item{detach}{detach or not}

\item{cpu}{cpu or not}
}
\value{
None
}
\description{
Create several hooks on the modules in `ms` with `hook_func`.
}
