% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{Hue}
\alias{Hue}
\title{Hue}
\usage{
Hue(max_hue = 0.1, p = 0.75, draw = NULL, batch = FALSE)
}
\arguments{
\item{max_hue}{max_hue}

\item{p}{p}

\item{draw}{draw}

\item{batch}{batch}
}
\description{
Apply change in hue of `max_hue` to batch of images with probability `p`.
}
