% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_Adapter}
\alias{icevision_Adapter}
\title{Adapter}
\usage{
icevision_Adapter(tfms)
}
\arguments{
\item{tfms}{`Sequence` of albumentation transforms.}
}
\value{
None
}
\description{
Adapter that enables the use of albumentations transforms.
}
