% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_BasicIAATransform}
\alias{icevision_BasicIAATransform}
\title{BasicIAATransform}
\usage{
icevision_BasicIAATransform(always_apply = FALSE, p = 0.5)
}
\arguments{
\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
BasicIAATransform
}
