% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_BasicTransform}
\alias{icevision_BasicTransform}
\title{BasicTransform}
\usage{
icevision_BasicTransform(always_apply = FALSE, p = 0.5)
}
\arguments{
\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
BasicTransform
}
