% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_ChannelDropout}
\alias{icevision_ChannelDropout}
\title{ChannelDropout}
\usage{
icevision_ChannelDropout(
  channel_drop_range = list(1, 1),
  fill_value = 0,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{channel_drop_range}{channel_drop_range}

\item{fill_value}{fill_value}

\item{always_apply}{always_apply}

\item{p}{p}
}
\description{
Randomly Drop Channels in the input Image.
}
\section{Targets}{

image
}

\section{Image types}{

uint8, uint16, unit32, float32
}

