% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_ChannelShuffle}
\alias{icevision_ChannelShuffle}
\title{ChannelShuffle}
\usage{
icevision_ChannelShuffle(always_apply = FALSE, p = 0.5)
}
\arguments{
\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Randomly rearrange channels of the input RGB image.
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

