% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_CLAHE}
\alias{icevision_CLAHE}
\title{CLAHE}
\usage{
icevision_CLAHE(
  clip_limit = 4,
  tile_grid_size = list(8, 8),
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{clip_limit}{clip_limit}

\item{tile_grid_size}{tile_grid_size}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Apply Contrast Limited Adaptive Histogram Equalization to the input image.
}
\details{

}
\section{Targets}{

image
}

\section{Image types}{

uint8
}

