% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_parsers.R
\name{icevision_ClassMap}
\alias{icevision_ClassMap}
\title{ClassMap}
\usage{
icevision_ClassMap(classes, background = 0)
}
\arguments{
\item{classes}{classes}

\item{background}{background}
}
\value{
Python dictionary
}
\description{
Utility class for mapping between class name and id.
}
