% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_ColorJitter}
\alias{icevision_ColorJitter}
\title{ColorJitter}
\usage{
icevision_ColorJitter(
  brightness = 0.2,
  contrast = 0.2,
  saturation = 0.2,
  hue = 0.2,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{brightness}{brightness}

\item{contrast}{contrast}

\item{saturation}{saturation}

\item{hue}{hue}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Randomly changes the brightness, contrast, and saturation of an image. Compared to ColorJitter from torchvision,
}
\details{
this transform gives a little bit different results because Pillow (used in torchvision) and OpenCV (used in
Albumentations) transform an image to HSV format by different formulas. Another difference - Pillow uses uint8
overflow, but we use value saturation.
}
