% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_Compose}
\alias{icevision_Compose}
\title{Compose}
\usage{
icevision_Compose(
  transforms,
  bbox_params = NULL,
  keypoint_params = NULL,
  additional_targets = NULL,
  p = 1
)
}
\arguments{
\item{transforms}{transforms}

\item{bbox_params}{bbox_params}

\item{keypoint_params}{keypoint_params}

\item{additional_targets}{additional_targets}

\item{p}{p}
}
\value{
None
}
\description{
Compose transforms and handle all transformations regrading bounding boxes
}
