% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_Crop}
\alias{icevision_Crop}
\title{Crop}
\usage{
icevision_Crop(
  x_min = 0,
  y_min = 0,
  x_max = 1024,
  y_max = 1024,
  always_apply = FALSE,
  p = 1
)
}
\arguments{
\item{x_min}{x_min}

\item{y_min}{y_min}

\item{x_max}{x_max}

\item{y_max}{y_max}

\item{always_apply}{always_apply}

\item{p}{p}
}
\description{
Crop region from image.
}
\details{

}
\section{Targets}{

image, mask, bboxes, keypoints
}

\section{Image types}{

uint8, float32
}

