% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_CropNonEmptyMaskIfExists}
\alias{icevision_CropNonEmptyMaskIfExists}
\title{CropNonEmptyMaskIfExists}
\usage{
icevision_CropNonEmptyMaskIfExists(
  height,
  width,
  ignore_values = NULL,
  ignore_channels = NULL,
  always_apply = FALSE,
  p = 1
)
}
\arguments{
\item{height}{height}

\item{width}{width}

\item{ignore_values}{ignore_values}

\item{ignore_channels}{ignore_channels}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Crop area with mask if mask is non-empty, else make random crop.
}
\section{Targets}{

image, mask, bboxes, keypoints
}

\section{Image types}{

uint8, float32
}

