% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_Cutout}
\alias{icevision_Cutout}
\title{Cutout}
\usage{
icevision_Cutout(
  num_holes = 8,
  max_h_size = 8,
  max_w_size = 8,
  fill_value = 0,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{num_holes}{num_holes}

\item{max_h_size}{max_h_size}

\item{max_w_size}{max_w_size}

\item{fill_value}{fill_value}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
CoarseDropout of the square regions in the image.
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

\section{Reference}{

| https://arxiv.org/abs/1708.04552 | https://github.com/uoguelph-mlrg/Cutout/blob/master/util/cutout.py | https://github.com/aleju/imgaug/blob/master/imgaug/augmenters/arithmetic.py
}

