% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_dataset.R
\name{icevision_Dataset}
\alias{icevision_Dataset}
\title{Dataset}
\usage{
icevision_Dataset(records, tfm = NULL)
}
\arguments{
\item{records}{A list of records.}

\item{tfm}{Transforms to be applied to each item.}
}
\value{
None
}
\description{
Container for a list of records and transforms.
}
\details{
Steps each time an item is requested (normally via directly indexing the `Dataset`):
Grab a record from the internal list of records.
Prepare the record (open the image, open the mask, add metadata).
Apply transforms to the record.
}
