% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_dataset.R
\name{icevision_Dataset_from_images}
\alias{icevision_Dataset_from_images}
\title{Icevision Dataset from images}
\usage{
icevision_Dataset_from_images(images, tfm = NULL, ...)
}
\arguments{
\item{images}{`Sequence` of images in memory (numpy arrays).}

\item{tfm}{Transforms to be applied to each item.}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Creates a `Dataset` from a list of images.
}
