% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_Downscale}
\alias{icevision_Downscale}
\title{Downscale}
\usage{
icevision_Downscale(
  scale_min = 0.25,
  scale_max = 0.25,
  interpolation = 0,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{scale_min}{scale_min}

\item{scale_max}{scale_max}

\item{interpolation}{cv2 interpolation method. cv2.INTER_NEAREST by default}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Decreases image quality by downscaling and upscaling back.
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

