% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_ElasticTransform}
\alias{icevision_ElasticTransform}
\title{ElasticTransform}
\usage{
icevision_ElasticTransform(
  alpha = 1,
  sigma = 50,
  alpha_affine = 50,
  interpolation = 1,
  border_mode = 4,
  value = NULL,
  mask_value = NULL,
  always_apply = FALSE,
  approximate = FALSE,
  p = 0.5
)
}
\arguments{
\item{alpha}{alpha}

\item{sigma}{sigma}

\item{alpha_affine}{alpha_affine}

\item{interpolation}{interpolation}

\item{border_mode}{border_mode}

\item{value}{value}

\item{mask_value}{mask_value}

\item{always_apply}{always_apply}

\item{approximate}{approximate}

\item{p}{p}
}
\value{
None
}
\description{
Elastic deformation of images as described in [Simard2003]_ (with modifications).
}
\details{
Based on https://gist.github.com/erniejunior/601cdf56d2b424757de5 .. [Simard2003] Simard, Steinkraus and Platt, "Best Practices for Convolutional Neural Networks applied to Visual Document Analysis", in Proc. of the International Conference on Document Analysis and Recognition, 2003.
}
\section{Targets}{

image, mask
}

\section{Image types}{

uint8, float32
}

