% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_Equalize}
\alias{icevision_Equalize}
\title{Equalize}
\usage{
icevision_Equalize(mode = "cv", by_channels = TRUE, mask = NULL, ...)
}
\arguments{
\item{mode}{mode}

\item{by_channels}{by_channels}

\item{mask}{mask}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Equalize the image histogram.
}
\section{Targets}{

image
}

\section{Image types}{

uint8
}

