% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_FancyPCA}
\alias{icevision_FancyPCA}
\title{FancyPCA}
\usage{
icevision_FancyPCA(alpha = 0.1, always_apply = FALSE, p = 0.5)
}
\arguments{
\item{alpha}{alpha}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Augment RGB image using FancyPCA from Krizhevsky's paper
}
\details{
"ImageNet Classification with Deep Convolutional Neural Networks"
}
\section{Targets}{

image
}

\section{Image types}{

3-channel uint8 images only
}

\section{Credit}{

http://papers.nips.cc/paper/4824-imagenet-classification-with-deep-convolutional-neural-networks.pdf https://deshanadesai.github.io/notes/Fancy-PCA-with-Scikit-Image https://pixelatedbrian.github.io/2018-04-29-fancy_pca/
}

