% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_datasplitters.R
\name{icevision_FixedSplitter}
\alias{icevision_FixedSplitter}
\title{FixedSplitter}
\usage{
icevision_FixedSplitter(splits)
}
\arguments{
\item{splits}{The predefined splits.}
}
\value{
None
}
\description{
Split `ids` based on predefined splits.
}
