% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_FromFloat}
\alias{icevision_FromFloat}
\title{FromFloat}
\usage{
icevision_FromFloat(
  dtype = "uint16",
  max_value = NULL,
  always_apply = FALSE,
  p = 1
)
}
\arguments{
\item{dtype}{dtype}

\item{max_value}{max_value}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Take an input array where all values should lie in the range [0, 1.0], multiply them by `max_value` and then
}
\details{
cast the resulted value to a type specified by `dtype`. If `max_value` is NULL the transform will try to infer
the maximum value for the data type from the `dtype` argument. This is the inverse transform for :class:`~albumentations.augmentations.transforms.ToFloat`.
}
\section{Targets}{

image
}

\section{Image types}{

float32
}

