% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_GaussianBlur}
\alias{icevision_GaussianBlur}
\title{GaussianBlur}
\usage{
icevision_GaussianBlur(
  blur_limit = list(3, 7),
  sigma_limit = 0,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{blur_limit}{blur_limit}

\item{sigma_limit}{sigma_limit}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Blur the input image using a Gaussian filter with a random kernel size.
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

