% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_GlassBlur}
\alias{icevision_GlassBlur}
\title{GlassBlur}
\usage{
icevision_GlassBlur(
  sigma = 0.7,
  max_delta = 4,
  iterations = 2,
  always_apply = FALSE,
  mode = "fast",
  p = 0.5
)
}
\arguments{
\item{sigma}{sigma}

\item{max_delta}{max_delta}

\item{iterations}{iterations}

\item{always_apply}{always_apply}

\item{mode}{mode}

\item{p}{p}
}
\value{
None
}
\description{
Apply glass noise to the input image.
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

\section{Reference}{

| https://arxiv.org/abs/1903.12261 | https://github.com/hendrycks/robustness/blob/master/ImageNet-C/create_c/make_imagenet_c.py
}

