% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_GridDropout}
\alias{icevision_GridDropout}
\title{GridDropout}
\usage{
icevision_GridDropout(
  ratio = 0.5,
  unit_size_min = NULL,
  unit_size_max = NULL,
  holes_number_x = NULL,
  holes_number_y = NULL,
  shift_x = 0,
  shift_y = 0,
  random_offset = FALSE,
  fill_value = 0,
  mask_fill_value = NULL,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{ratio}{ratio}

\item{unit_size_min}{unit_size_min}

\item{unit_size_max}{unit_size_max}

\item{holes_number_x}{holes_number_x}

\item{holes_number_y}{holes_number_y}

\item{shift_x}{shift_x}

\item{shift_y}{shift_y}

\item{random_offset}{random_offset}

\item{fill_value}{fill_value}

\item{mask_fill_value}{mask_fill_value}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
GridDropout, drops out rectangular regions of an image and the corresponding mask in a grid fashion.
}
\section{Targets}{

image, mask
}

\section{Image types}{

uint8, float32
}

\section{References}{

https://arxiv.org/abs/2001.04086
}

