% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_HorizontalFlip}
\alias{icevision_HorizontalFlip}
\title{HorizontalFlip}
\usage{
icevision_HorizontalFlip(always_apply = FALSE, p = 0.5)
}
\arguments{
\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Flip the input horizontally around the y-axis.
}
\section{Targets}{

image, mask, bboxes, keypoints
}

\section{Image types}{

uint8, float32
}

