% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_IAAAdditiveGaussianNoise}
\alias{icevision_IAAAdditiveGaussianNoise}
\title{IAAAdditiveGaussianNoise}
\usage{
icevision_IAAAdditiveGaussianNoise(
  loc = 0,
  scale = list(2.55, 12.75),
  per_channel = FALSE,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{loc}{loc}

\item{scale}{scale}

\item{per_channel}{per_channel}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Add gaussian noise to the input image.
}
\section{Targets}{

image
}

