% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_IAAAffine}
\alias{icevision_IAAAffine}
\title{IAAAffine}
\usage{
icevision_IAAAffine(
  scale = 1,
  translate_percent = NULL,
  translate_px = NULL,
  rotate = 0,
  shear = 0,
  order = 1,
  cval = 0,
  mode = "reflect",
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{scale}{scale}

\item{translate_percent}{translate_percent}

\item{translate_px}{translate_px}

\item{rotate}{rotate}

\item{shear}{shear}

\item{order}{order}

\item{cval}{cval}

\item{mode}{mode}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None

None
}
\description{
Place a regular grid of points on the input and randomly move the neighbourhood of these point around
}
\details{
via affine transformations. Note: This class introduce interpolation artifacts to mask if it has values other than {0;1}
}
\section{Targets}{

image, mask
}

