% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_IAACropAndPad}
\alias{icevision_IAACropAndPad}
\title{IAACropAndPad}
\usage{
icevision_IAACropAndPad(
  px = NULL,
  percent = NULL,
  pad_mode = "constant",
  pad_cval = 0,
  keep_size = TRUE,
  always_apply = FALSE,
  p = 1
)
}
\arguments{
\item{px}{px}

\item{percent}{percent}

\item{pad_mode}{pad_mode}

\item{pad_cval}{pad_cval}

\item{keep_size}{keep_size}

\item{always_apply}{always_apply}

\item{p}{p}
}
\description{
Transform for segmentation task.
}
