% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_IAAEmboss}
\alias{icevision_IAAEmboss}
\title{IAAEmboss}
\usage{
icevision_IAAEmboss(
  alpha = list(0.2, 0.5),
  strength = list(0.2, 0.7),
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{alpha}{alpha}

\item{strength}{strength}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Emboss the input image and overlays the result with the original image.
}
\section{Targets}{

image
}

