% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_IAAPerspective}
\alias{icevision_IAAPerspective}
\title{IAAPerspective}
\usage{
icevision_IAAPerspective(
  scale = list(0.05, 0.1),
  keep_size = TRUE,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{scale}{scale}

\item{keep_size}{keep_size}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Perform a random four point perspective transform of the input.
}
\details{
Note: This class introduce interpolation artifacts to mask if it has values other than {0;1}
}
\section{Targets}{

image, mask
}

