% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_IAAPiecewiseAffine}
\alias{icevision_IAAPiecewiseAffine}
\title{IAAPiecewiseAffine}
\usage{
icevision_IAAPiecewiseAffine(
  scale = list(0.03, 0.05),
  nb_rows = 4,
  nb_cols = 4,
  order = 1,
  cval = 0,
  mode = "constant",
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{scale}{scale}

\item{nb_rows}{nb_rows}

\item{nb_cols}{nb_cols}

\item{order}{order}

\item{cval}{cval}

\item{mode}{mode}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Place a regular grid of points on the input and randomly move the neighbourhood of these point around
}
\details{
via affine transformations. Note: This class introduce interpolation artifacts to mask if it has values other than {0;1}
}
\section{Targets}{

image, mask
}

