% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_IAASharpen}
\alias{icevision_IAASharpen}
\title{IAASharpen}
\usage{
icevision_IAASharpen(
  alpha = list(0.2, 0.5),
  lightness = list(0.5, 1),
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{alpha}{alpha}

\item{lightness}{lightness}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Sharpen the input image and overlays the result with the original image.
}
\section{Targets}{

image
}

