% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_IAASuperpixels}
\alias{icevision_IAASuperpixels}
\title{IAASuperpixels}
\usage{
icevision_IAASuperpixels(
  p_replace = 0.1,
  n_segments = 100,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{p_replace}{p_replace}

\item{n_segments}{n_segments}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Completely or partially transform the input image to its superpixel representation. Uses skimage's version
}
\details{
of the SLIC algorithm. May be slow.
}
\section{Targets}{

image
}

