% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_ImageCompression}
\alias{icevision_ImageCompression}
\title{ImageCompression}
\usage{
icevision_ImageCompression(
  quality_lower = 99,
  quality_upper = 100,
  compression_type = 0,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{quality_lower}{quality_lower}

\item{quality_upper}{quality_upper}

\item{compression_type}{compression_type}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Decrease Jpeg, WebP compression of an image.
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

