% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_ImageOnlyIAATransform}
\alias{icevision_ImageOnlyIAATransform}
\title{ImageOnlyIAATransform}
\usage{
icevision_ImageOnlyIAATransform(always_apply = FALSE, p = 0.5)
}
\arguments{
\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Transform applied to image only.
}
