% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_ImageOnlyTransform}
\alias{icevision_ImageOnlyTransform}
\title{ImageOnlyTransform}
\usage{
icevision_ImageOnlyTransform(always_apply = FALSE, p = 0.5)
}
\arguments{
\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Transform applied to image only.
}
