% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_InvertImg}
\alias{icevision_InvertImg}
\title{InvertImg}
\usage{
icevision_InvertImg(always_apply = FALSE, p = 0.5)
}
\arguments{
\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Invert the input image by subtracting pixel values from 255.
}
\section{Targets}{

image
}

\section{Image types}{

uint8
}

