% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_ISONoise}
\alias{icevision_ISONoise}
\title{ISONoise}
\usage{
icevision_ISONoise(
  color_shift = list(0.01, 0.05),
  intensity = list(0.1, 0.5),
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{color_shift}{color_shift}

\item{intensity}{intensity}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Apply camera sensor noise.
}
\section{Targets}{

image
}

\section{Image types}{

uint8
}

