% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_MaskDropout}
\alias{icevision_MaskDropout}
\title{MaskDropout}
\usage{
icevision_MaskDropout(
  max_objects = 1,
  image_fill_value = 0,
  mask_fill_value = 0,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{max_objects}{max_objects}

\item{image_fill_value}{image_fill_value}

\item{mask_fill_value}{mask_fill_value}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Image & mask augmentation that zero out mask and image regions corresponding
}
\details{
to randomly chosen object instance from mask. Mask must be single-channel image, zero values treated as background.
Image can be any number of channels. Inspired by https://www.kaggle.com/c/severstal-steel-defect-detection/discussion/114254
}
