% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_MedianBlur}
\alias{icevision_MedianBlur}
\title{MedianBlur}
\usage{
icevision_MedianBlur(blur_limit = 7, always_apply = FALSE, p = 0.5)
}
\arguments{
\item{blur_limit}{blur_limit}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Blur the input image using a median filter with a random aperture linear size.
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

