% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_MultiplicativeNoise}
\alias{icevision_MultiplicativeNoise}
\title{MultiplicativeNoise}
\usage{
icevision_MultiplicativeNoise(
  multiplier = list(0.9, 1.1),
  per_channel = FALSE,
  elementwise = FALSE,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{multiplier}{multiplier}

\item{per_channel}{per_channel}

\item{elementwise}{elementwise}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Multiply image to random number or array of numbers.
}
\section{Targets}{

image
}

\section{Image types}{

Any
}

