% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_OpticalDistortion}
\alias{icevision_OpticalDistortion}
\title{OpticalDistortion}
\usage{
icevision_OpticalDistortion(
  distort_limit = 0.05,
  shift_limit = 0.05,
  interpolation = 1,
  border_mode = 4,
  value = NULL,
  mask_value = NULL,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{distort_limit}{distort_limit}

\item{shift_limit}{shift_limit}

\item{interpolation}{interpolation}

\item{border_mode}{border_mode}

\item{value}{value}

\item{mask_value}{mask_value}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
OpticalDistortion
}
\details{
distort_limit (float, (float, float)): If distort_limit is a single float, the range will be (-distort_limit, distort_limit). Default: (-0.05, 0.05). shift_limit (float, (float, float))): If shift_limit is a single float, the range will be (-shift_limit, shift_limit). Default: (-0.05, 0.05). interpolation (OpenCV flag): flag that is used to specify the interpolation algorithm. Should be one of: cv2.INTER_NEAREST, cv2.INTER_LINEAR, cv2.INTER_CUBIC, cv2.INTER_AREA, cv2.INTER_LANCZOS4. Default: cv2.INTER_LINEAR. border_mode (OpenCV flag): flag that is used to specify the pixel extrapolation method. Should be one of: cv2.BORDER_CONSTANT, cv2.BORDER_REPLICATE, cv2.BORDER_REFLECT, cv2.BORDER_WRAP, cv2.BORDER_REFLECT_101. Default: cv2.BORDER_REFLECT_101 value (int, float, list of ints, list of float): padding value if border_mode is cv2.BORDER_CONSTANT. mask_value (int, float, list of ints, list of float): padding value if border_mode is cv2.BORDER_CONSTANT applied for masks. Targets: image, mask Image types: uint8, float32
}
\section{Targets}{

image, mask
}

\section{Image types}{

uint8, float32
}

