% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_Posterize}
\alias{icevision_Posterize}
\title{Posterize}
\usage{
icevision_Posterize(num_bits = 4, always_apply = FALSE, p = 0.5)
}
\arguments{
\item{num_bits}{num_bits}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Reduce the number of bits for each color channel.
}
\section{Targets}{

image
}

\section{Image types}{

uint8
}

