% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_RandomContrast}
\alias{icevision_RandomContrast}
\title{RandomContrast}
\usage{
icevision_RandomContrast(limit = 0.2, always_apply = FALSE, p = 0.5)
}
\arguments{
\item{limit}{limit}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Randomly change contrast of the input image.
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

