% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_RandomCrop}
\alias{icevision_RandomCrop}
\title{RandomCrop}
\usage{
icevision_RandomCrop(height, width, always_apply = FALSE, p = 1)
}
\arguments{
\item{height}{height}

\item{width}{width}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Crop a random part of the input.
}
\section{Targets}{

image, mask, bboxes, keypoints
}

\section{Image types}{

uint8, float32
}

