% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_RandomCropNearBBox}
\alias{icevision_RandomCropNearBBox}
\title{RandomCropNearBBox}
\usage{
icevision_RandomCropNearBBox(max_part_shift = 0.3, always_apply = FALSE, p = 1)
}
\arguments{
\item{max_part_shift}{max_part_shift}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Crop bbox from image with random shift by x,y coordinates
}
\section{Targets}{

image, mask, bboxes, keypoints
}

\section{Image types}{

uint8, float32
}

