% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_RandomFog}
\alias{icevision_RandomFog}
\title{RandomFog}
\usage{
icevision_RandomFog(
  fog_coef_lower = 0.3,
  fog_coef_upper = 1,
  alpha_coef = 0.08,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{fog_coef_lower}{fog_coef_lower}

\item{fog_coef_upper}{fog_coef_upper}

\item{alpha_coef}{alpha_coef}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Simulates fog for the image
}
\details{
From https://github.com/UjjwalSaxena/Automold--Road-Augmentation-Library
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

