% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_RandomGridShuffle}
\alias{icevision_RandomGridShuffle}
\title{RandomGridShuffle}
\usage{
icevision_RandomGridShuffle(grid = list(3, 3), always_apply = FALSE, p = 0.5)
}
\arguments{
\item{grid}{grid}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Random shuffle grid's cells on image.
}
\section{Targets}{

image, mask
}

\section{Image types}{

uint8, float32
}

