% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_RandomRain}
\alias{icevision_RandomRain}
\title{RandomRain}
\usage{
icevision_RandomRain(
  slant_lower = -10,
  slant_upper = 10,
  drop_length = 20,
  drop_width = 1,
  drop_color = list(200, 200, 200),
  blur_value = 7,
  brightness_coefficient = 0.7,
  rain_type = NULL,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{slant_lower}{should be in range [-20, 20].}

\item{slant_upper}{should be in range [-20, 20].}

\item{drop_length}{should be in range [0, 100].}

\item{drop_width}{should be in range [1, 5]. drop_color (list of (r, g, b)): rain lines color. blur_value (int): rainy view are blurry brightness_coefficient (float): rainy days are usually shady. Should be in range [0, 1].}

\item{drop_color}{drop_color}

\item{blur_value}{blur_value}

\item{brightness_coefficient}{brightness_coefficient}

\item{rain_type}{One of [NULL, "drizzle", "heavy", "torrestial"]}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Adds rain effects.
}
\details{
From https://github.com/UjjwalSaxena/Automold--Road-Augmentation-Library
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

