% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_RandomResizedCrop}
\alias{icevision_RandomResizedCrop}
\title{RandomResizedCrop}
\usage{
icevision_RandomResizedCrop(
  height,
  width,
  scale = list(0.08, 1),
  ratio = list(0.75, 1.33333333333333),
  interpolation = 1,
  always_apply = FALSE,
  p = 1
)
}
\arguments{
\item{height}{height}

\item{width}{width}

\item{scale}{scale}

\item{ratio}{ratio}

\item{interpolation}{interpolation}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Torchvision's variant of crop a random part of the input and rescale it to some size.
}
\section{Targets}{

image, mask, bboxes, keypoints
}

\section{Image types}{

uint8, float32
}

