% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_RandomScale}
\alias{icevision_RandomScale}
\title{RandomScale}
\usage{
icevision_RandomScale(
  scale_limit = 0.1,
  interpolation = 1L,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{scale_limit}{scale_limit}

\item{interpolation}{interpolation}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Randomly resize the input. Output image size is different from the input image size.
}
\section{Targets}{

image, mask, bboxes, keypoints
}

\section{Image types}{

uint8, float32
}

