% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_RandomShadow}
\alias{icevision_RandomShadow}
\title{RandomShadow}
\usage{
icevision_RandomShadow(
  shadow_roi = list(0, 0.5, 1, 1),
  num_shadows_lower = 1,
  num_shadows_upper = 2,
  shadow_dimension = 5,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{shadow_roi}{shadow_roi}

\item{num_shadows_lower}{num_shadows_lower}

\item{num_shadows_upper}{num_shadows_upper}

\item{shadow_dimension}{shadow_dimension}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Simulates shadows for the image
}
\details{
From https://github.com/UjjwalSaxena/Automold--Road-Augmentation-Library
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

