% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_RandomSnow}
\alias{icevision_RandomSnow}
\title{RandomSnow}
\usage{
icevision_RandomSnow(
  snow_point_lower = 0.1,
  snow_point_upper = 0.3,
  brightness_coeff = 2.5,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{snow_point_lower}{snow_point_lower}

\item{snow_point_upper}{snow_point_upper}

\item{brightness_coeff}{brightness_coeff}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Bleach out some pixel values simulating snow.
}
\details{
From https://github.com/UjjwalSaxena/Automold--Road-Augmentation-Library
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

