% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_datasplitters.R
\name{icevision_RandomSplitter}
\alias{icevision_RandomSplitter}
\title{RandomSplitter}
\usage{
icevision_RandomSplitter(probs, seed = NULL)
}
\arguments{
\item{probs}{`Sequence` of probabilities that must sum to one. The length of
the `Sequence` is the number of groups to to split the items into.}

\item{seed}{Internal seed used for shuffling the items. Define this if you need reproducible results.}
}
\value{
None
}
\description{
Randomly splits items.
}
