% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_RandomSunFlare}
\alias{icevision_RandomSunFlare}
\title{RandomSunFlare}
\usage{
icevision_RandomSunFlare(
  flare_roi = list(0, 0, 1, 0.5),
  angle_lower = 0,
  angle_upper = 1,
  num_flare_circles_lower = 6,
  num_flare_circles_upper = 10,
  src_radius = 400,
  src_color = list(255, 255, 255),
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{flare_roi}{flare_roi}

\item{angle_lower}{angle_lower}

\item{angle_upper}{angle_upper}

\item{num_flare_circles_lower}{num_flare_circles_lower}

\item{num_flare_circles_upper}{num_flare_circles_upper}

\item{src_radius}{src_radius}

\item{src_color}{src_color}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Simulates Sun Flare for the image
}
\details{
From https://github.com/UjjwalSaxena/Automold--Road-Augmentation-Library
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

