% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_read_rgb_image}
\alias{icevision_read_rgb_image}
\title{Read_rgb_image}
\usage{
icevision_read_rgb_image(path)
}
\arguments{
\item{path}{path}
}
\value{
None
}
\description{
Read_rgb_image
}
