% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_Resize}
\alias{icevision_Resize}
\title{Resize}
\usage{
icevision_Resize(height, width, interpolation = 1, always_apply = FALSE, p = 1)
}
\arguments{
\item{height}{height}

\item{width}{width}

\item{interpolation}{interpolation}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Resize the input to the given height and width.
}
\section{Targets}{

image, mask, bboxes, keypoints
}

\section{Image types}{

uint8, float32
}

