% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_resize_and_pad}
\alias{icevision_resize_and_pad}
\title{Resize_and_pad}
\usage{
icevision_resize_and_pad(
  size,
  pad = partial(icevision_PadIfNeeded, border_mode = 0, value = c(124L, 116L, 104L))
)
}
\arguments{
\item{size}{size}

\item{pad}{pad}
}
\value{
None
}
\description{
Resize_and_pad
}
