% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_Rotate}
\alias{icevision_Rotate}
\title{Rotate}
\usage{
icevision_Rotate(
  limit = 90,
  interpolation = 1,
  border_mode = 4,
  value = NULL,
  mask_value = NULL,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{limit}{limit}

\item{interpolation}{interpolation}

\item{border_mode}{border_mode}

\item{value}{value}

\item{mask_value}{mask_value}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Rotate the input by an angle selected randomly from the uniform distribution.
}
\section{Targets}{

image, mask, bboxes, keypoints
}

\section{Image types}{

uint8, float32
}

