% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_ShiftScaleRotate}
\alias{icevision_ShiftScaleRotate}
\title{ShiftScaleRotate}
\usage{
icevision_ShiftScaleRotate(
  shift_limit = 0.0625,
  scale_limit = 0.1,
  rotate_limit = 45,
  interpolation = 1,
  border_mode = 4,
  value = NULL,
  mask_value = NULL,
  shift_limit_x = NULL,
  shift_limit_y = NULL,
  always_apply = FALSE,
  p = 0.5
)

icevision_ShiftScaleRotate(
  shift_limit = 0.0625,
  scale_limit = 0.1,
  rotate_limit = 45,
  interpolation = 1,
  border_mode = 4,
  value = NULL,
  mask_value = NULL,
  shift_limit_x = NULL,
  shift_limit_y = NULL,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{shift_limit}{shift_limit}

\item{scale_limit}{scale_limit}

\item{rotate_limit}{rotate_limit}

\item{interpolation}{interpolation}

\item{border_mode}{border_mode}

\item{value}{value}

\item{mask_value}{mask_value}

\item{shift_limit_x}{shift_limit_x}

\item{shift_limit_y}{shift_limit_y}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None

None
}
\description{
Randomly apply affine transforms: translate, scale and rotate the input.

Randomly apply affine transforms: translate, scale and rotate the input.
}
\section{Targets}{

image, mask, keypoints


image, mask, keypoints
}

\section{Image types}{

uint8, float32


uint8, float32
}

