% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_datasplitters.R
\name{icevision_SingleSplitSplitter}
\alias{icevision_SingleSplitSplitter}
\title{SingleSplitSplitter}
\usage{
icevision_SingleSplitSplitter(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
all items in a single group, without shuffling.
}
\description{
SingleSplitSplitter
}
