% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_SmallestMaxSize}
\alias{icevision_SmallestMaxSize}
\title{SmallestMaxSize}
\usage{
icevision_SmallestMaxSize(
  max_size = 1024,
  interpolation = 1,
  always_apply = FALSE,
  p = 1
)
}
\arguments{
\item{max_size}{max_size}

\item{interpolation}{interpolation}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Rescale an image so that minimum side is equal to max_size, keeping the aspect ratio of the initial image.
}
\section{Targets}{

image, mask, bboxes, keypoints
}

\section{Image types}{

uint8, float32
}

