% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_Solarize}
\alias{icevision_Solarize}
\title{Solarize}
\usage{
icevision_Solarize(threshold = 128, always_apply = FALSE, p = 0.5)
}
\arguments{
\item{threshold}{threshold}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Invert all pixel values above a threshold.
}
\section{Targets}{

image
}

\section{Image types}{

any
}

