% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_ToFloat}
\alias{icevision_ToFloat}
\title{ToFloat}
\usage{
icevision_ToFloat(max_value = NULL, always_apply = FALSE, p = 1)
}
\arguments{
\item{max_value}{max_value}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Divide pixel values by `max_value` to get a float32 output array where all values lie in the range [0, 1.0].
}
\details{
If `max_value` is NULL the transform will try to infer the maximum value by inspecting the data type of the input
image. See Also: :class:`~albumentations.augmentations.transforms.FromFloat`
}
\section{See Also}{

:class:`~albumentations.augmentations.transforms.FromFloat`
}

\section{Targets}{

image
}

\section{Image types}{

any type
}

