% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_ToGray}
\alias{icevision_ToGray}
\title{ToGray}
\usage{
icevision_ToGray(always_apply = FALSE, p = 0.5)
}
\arguments{
\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Convert the input RGB image to grayscale. If the mean pixel value for the resulting image is greater
than 127, invert the resulting grayscale image.
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

