% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_ToSepia}
\alias{icevision_ToSepia}
\title{ToSepia}
\usage{
icevision_ToSepia(always_apply = FALSE, p = 0.5)
}
\arguments{
\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Applies sepia filter to the input RGB image
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

