% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{icevision_VerticalFlip}
\alias{icevision_VerticalFlip}
\title{VerticalFlip}
\usage{
icevision_VerticalFlip(always_apply = FALSE, p = 0.5)
}
\arguments{
\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Flip the input vertically around the x-axis.
}
\section{Targets}{

image, mask, bboxes, keypoints
}

\section{Image types}{

uint8, float32
}

