% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{icnr_init}
\alias{icnr_init}
\title{Icnr_init}
\usage{
icnr_init(x, scale = 2, init = nn()$init$kaiming_normal_)
}
\arguments{
\item{x}{tensor}

\item{scale}{int, scale}

\item{init}{initializer}
}
\value{
None
}
\description{
ICNR init of `x`, with `scale` and `init` function
}
