% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_datasplitters.R
\name{IDMap}
\alias{IDMap}
\title{IDMap}
\usage{
IDMap(initial_names = NULL)
}
\arguments{
\item{initial_names}{initial_names}
}
\value{
None
}
\description{
Works like a dictionary that automatically assign values for new keys.
}
