% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{Image_open}
\alias{Image_open}
\title{Image_open}
\usage{
Image_open(fp, mode = "r")
}
\arguments{
\item{fp}{fp}

\item{mode}{mode}
}
\value{
None
}
\description{
Opens and identifies the given image file.
}
