% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loaders.R
\name{ImageDataLoaders_from_df}
\alias{ImageDataLoaders_from_df}
\title{ImageDataLoaders_from_df}
\usage{
ImageDataLoaders_from_df(
  cls,
  df,
  path = ".",
  valid_pct = 0.2,
  seed = NULL,
  fn_col = 0,
  folder = NULL,
  suff = "",
  label_col = 1,
  label_delim = NULL,
  y_block = NULL,
  valid_col = NULL,
  item_tfms = NULL,
  batch_tfms = NULL,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  device = NULL
)
}
\arguments{
\item{cls}{cls}

\item{df}{df}

\item{path}{path}

\item{valid_pct}{valid_pct}

\item{seed}{seed}

\item{fn_col}{fn_col}

\item{folder}{folder}

\item{suff}{suff}

\item{label_col}{label_col}

\item{label_delim}{label_delim}

\item{y_block}{y_block}

\item{valid_col}{valid_col}

\item{item_tfms}{item_tfms}

\item{batch_tfms}{batch_tfms}

\item{bs}{bs}

\item{val_bs}{val_bs}

\item{shuffle_train}{shuffle_train}

\item{device}{device}
}
\description{
Create from `df` using `fn_col` and `label_col`
}
\details{

}
