% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loaders.R
\name{ImageDataLoaders_from_folder}
\alias{ImageDataLoaders_from_folder}
\title{from_folder}
\usage{
ImageDataLoaders_from_folder(
  path,
  train = "train",
  valid = "valid",
  valid_pct = NULL,
  seed = NULL,
  vocab = NULL,
  item_tfms = NULL,
  batch_tfms = NULL,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  device = NULL,
  size = NULL,
  ...
)
}
\arguments{
\item{path}{path}

\item{train}{train}

\item{valid}{valid}

\item{valid_pct}{valid_pct}

\item{seed}{seed}

\item{vocab}{vocab}

\item{item_tfms}{item_tfms}

\item{batch_tfms}{batch_tfms}

\item{bs}{bs}

\item{val_bs}{val_bs}

\item{shuffle_train}{shuffle_train}

\item{device}{device}

\item{cls}{cls}
}
\description{
Create from imagenet style dataset in `path` with `train` and `valid` subfolders (or provide `valid_pct`)
}
\details{

}
