% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loaders.R
\name{ImageDataLoaders_from_name_re}
\alias{ImageDataLoaders_from_name_re}
\title{From_name_re}
\usage{
ImageDataLoaders_from_name_re(
  path,
  fnames,
  pat,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  device = NULL,
  item_tfms = NULL,
  batch_tfms = NULL,
  ...
)
}
\arguments{
\item{path}{path}

\item{fnames}{fnames}

\item{pat}{pat}

\item{bs}{bs}

\item{val_bs}{val_bs}

\item{shuffle_train}{shuffle_train}

\item{device}{device}

\item{item_tfms}{item_tfms}

\item{batch_tfms}{batch_tfms}
}
\description{
Create from the name attrs of `fnames` in `path`s with re expression `pat`
}
