% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loaders.R
\name{ImageDataLoaders_from_path_re}
\alias{ImageDataLoaders_from_path_re}
\title{ImageDataLoaders_from_path_re}
\usage{
ImageDataLoaders_from_path_re(
  path,
  fnames,
  pat,
  valid_pct = 0.2,
  seed = NULL,
  item_tfms = NULL,
  batch_tfms = NULL,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  device = NULL
)
}
\arguments{
\item{path}{path}

\item{fnames}{fnames}

\item{pat}{pat}

\item{valid_pct}{valid_pct}

\item{seed}{seed}

\item{item_tfms}{item_tfms}

\item{batch_tfms}{batch_tfms}

\item{bs}{bs}

\item{val_bs}{val_bs}

\item{shuffle_train}{shuffle_train}

\item{device}{device}
}
\description{
Create from list of `fnames` in `path`s with re expression `pat`
}
