% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R
\name{imagenet_stats}
\alias{imagenet_stats}
\title{Imagenet_stats}
\usage{
imagenet_stats()
}
\description{
list() -> empty list
}
\details{
list(iterable) -> list initialized from iterable's items If the argument is a list, the return value is the same object.
}
