% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_model.R
\name{InceptionModule}
\alias{InceptionModule}
\title{InceptionModule}
\usage{
InceptionModule(
  ni,
  nb_filters = 32,
  kss = c(39, 19, 9),
  bottleneck_size = 32,
  stride = 1
)
}
\arguments{
\item{ni}{number of input channels}

\item{nb_filters}{the number of filters}

\item{kss}{kernel size}

\item{bottleneck_size}{bottleneck size}

\item{stride}{stride}
}
\value{
module
}
\description{
The inception Module from `ni` inputs to len('kss')*`nb_filters`+`bottleneck_size`
}
