% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{IndexSplitter}
\alias{IndexSplitter}
\title{IndexSplitter}
\usage{
IndexSplitter(valid_idx)
}
\arguments{
\item{valid_idx}{The indices to use for the validation set (defaults to a random split otherwise)}
}
\description{
Split `items` so that `val_idx` are in the validation set and the others in the training set
}
