% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{init}
\alias{init}
\title{Wandb init}
\usage{
init(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
wandb Run object

None
}
\description{
Initialize a wandb Run.
}
\section{see https}{
//docs.wandb.com/library/init
}

