% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{init_default}
\alias{init_default}
\title{Init_default}
\usage{
init_default(m, func = nn$init$kaiming_normal_)
}
\arguments{
\item{m}{m parameter}

\item{func}{function}
}
\value{
None
}
\description{
Initialize `m` weights with `func` and set `bias` to 0.
}
