% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{init_linear}
\alias{init_linear}
\title{Init_linear}
\usage{
init_linear(m, act_func = NULL, init = "auto", bias_std = 0.01)
}
\arguments{
\item{m}{parameter}

\item{act_func}{activation function}

\item{init}{initializer}

\item{bias_std}{bias standard deviation}
}
\value{
None
}
\description{
Init_linear
}
