% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_fastai.R
\name{install_fastai}
\alias{install_fastai}
\title{Install fastai}
\usage{
install_fastai(
  version,
  gpu = FALSE,
  cuda_version = "10.1",
  overwrite = FALSE,
  extra_pkgs = c("kaggle", "transformers", "timm", "fastinference[interp]", "shap",
    "blurr", "icevision[all]"),
  skip_git_pkgs = FALSE,
  TPU = FALSE
)
}
\arguments{
\item{version}{specify version}

\item{gpu}{installation of gpu}

\item{cuda_version}{if gpu true, then cuda version is required. By default it is 10.1}

\item{overwrite}{will install all the dependencies}

\item{extra_pkgs}{character vector of additional packages}

\item{skip_git_pkgs}{skip installation from Github}

\item{TPU}{official way to install Pytorch-XLA 1.7}
}
\value{
None
}
\description{
Install fastai
}
