% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{InstanceNorm}
\alias{InstanceNorm}
\title{InstanceNorm}
\usage{
InstanceNorm(
  nf,
  ndim = 2,
  norm_type = 5,
  affine = TRUE,
  eps = 1e-05,
  momentum = 0.1,
  track_running_stats = FALSE
)
}
\arguments{
\item{nf}{input shape}

\item{ndim}{dimension number}

\item{norm_type}{normalization type}

\item{affine}{affine}

\item{eps}{epsilon}

\item{momentum}{momentum}

\item{track_running_stats}{track running statistics}
}
\value{
None
}
\description{
InstanceNorm layer with `nf` features and `ndim` initialized depending on `norm_type`.
}
