% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{JaccardCoeff}
\alias{JaccardCoeff}
\title{JaccardCoeff}
\usage{
JaccardCoeff(axis = 1)
}
\arguments{
\item{axis}{axis}
}
\value{
None
}
\description{
Implementation of the Jaccard coefficient that is lighter in RAM
}
