% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{kg}
\alias{kg}
\title{Kaggle API}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
kg
}
\description{
Kaggle API
}
\keyword{datasets}
