% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{LabeledBBox}
\alias{LabeledBBox}
\title{LabeledBBox}
\usage{
LabeledBBox(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
Basic type for a list of bounding boxes in an image
}
