% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_models.R
\name{language_model_learner}
\alias{language_model_learner}
\title{language_model_learner}
\usage{
language_model_learner(
  dls,
  arch,
  config = NULL,
  drop_mult = 1,
  backwards = FALSE,
  pretrained = TRUE,
  pretrained_fnames = NULL,
  opt_func = Adam(),
  lr = 0.001,
  cbs = NULL,
  metrics = NULL,
  path = NULL,
  model_dir = "models",
  wd = NULL,
  wd_bn_bias = FALSE,
  train_bn = TRUE,
  moms = list(0.95, 0.85, 0.95),
  ...
)
}
\arguments{
\item{dls}{dls}

\item{arch}{arch}

\item{config}{config}

\item{drop_mult}{drop_mult}

\item{backwards}{backwards}

\item{pretrained}{pretrained}

\item{pretrained_fnames}{pretrained_fnames}

\item{opt_func}{opt_func}

\item{lr}{lr}

\item{cbs}{cbs}

\item{metrics}{metrics}

\item{path}{path}

\item{model_dir}{model_dir}

\item{wd}{wd}

\item{wd_bn_bias}{wd_bn_bias}

\item{train_bn}{train_bn}

\item{moms}{moms}

\item{loss_func}{loss_func}

\item{splitter}{splitter}
}
\description{
Create a `Learner` with a language model from `dls` and `arch`.
}
\details{

}
