% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_hook.R
\name{layer_info}
\alias{layer_info}
\title{Layer_info}
\usage{
layer_info(learn, ...)
}
\arguments{
\item{learn}{learner/model}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Return layer infos of `model` on `xb` (only support batch first inputs)
}
