% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{Learner}
\alias{Learner}
\title{Learner}
\usage{
Learner(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
Learner
}
\examples{
\dontrun{

model = LitModel()

data = Data_Loaders(model$train_dataloader(), model$val_dataloader())$cuda()

learn = Learner(data, model, loss_func = F$cross_entropy, opt_func = Adam,
                metrics = accuracy)

}

}
