% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{LightingTfm}
\alias{LightingTfm}
\title{LightingTfm}
\usage{
LightingTfm(fs, ...)
}
\arguments{
\item{fs}{fs}

\item{...}{parameters to pass}
}
\value{
None
}
\description{
Apply `fs` to the logits
}
