% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{LinBnDrop}
\alias{LinBnDrop}
\title{LinBnDrop}
\usage{
LinBnDrop(n_in, n_out, bn = TRUE, p = 0, act = NULL, lin_first = FALSE)
}
\arguments{
\item{n_in}{input shape}

\item{n_out}{output shape}

\item{bn}{bn}

\item{p}{probability}

\item{act}{activation}

\item{lin_first}{linear first}
}
\value{
None
}
\description{
Module grouping `BatchNorm1d`, `Dropout` and `Linear` layers
}
