% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_models.R
\name{LinearDecoder}
\alias{LinearDecoder}
\title{LinearDecoder}
\usage{
LinearDecoder(n_out, n_hid, output_p = 0.1, tie_encoder = NULL, bias = TRUE)
}
\arguments{
\item{n_out}{n_out}

\item{n_hid}{n_hid}

\item{output_p}{output_p}

\item{tie_encoder}{tie_encoder}

\item{bias}{bias}
}
\description{
To go on top of a RNNCore module and create a Language Model.
}
