% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{LMDataLoader}
\alias{LMDataLoader}
\title{LMDataLoader}
\usage{
LMDataLoader(
  dataset,
  lens = NULL,
  cache = 2,
  bs = 64,
  seq_len = 72,
  num_workers = 0,
  shuffle = FALSE,
  verbose = FALSE,
  do_setup = TRUE,
  pin_memory = FALSE,
  timeout = 0L,
  batch_size = NULL,
  drop_last = FALSE,
  indexed = NULL,
  n = NULL,
  device = NULL
)
}
\arguments{
\item{dataset}{dataset}

\item{lens}{lens}

\item{cache}{cache}

\item{bs}{bs}

\item{seq_len}{seq_len}

\item{num_workers}{num_workers}

\item{shuffle}{shuffle}

\item{verbose}{verbose}

\item{do_setup}{do_setup}

\item{pin_memory}{pin_memory}

\item{timeout}{timeout}

\item{batch_size}{batch_size}

\item{drop_last}{drop_last}

\item{indexed}{indexed}

\item{n}{n}

\item{device}{device}
}
\description{
A `DataLoader` suitable for language modeling
}
\details{

}
