% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R, R/text_learner.R
\name{LMLearner}
\alias{LMLearner}
\title{LMLearner}
\usage{
LMLearner(
  dls,
  model,
  alpha = 2,
  beta = 1,
  moms = list(0.8, 0.7, 0.8),
  loss_func = NULL,
  opt_func = Adam(),
  lr = 0.001,
  splitter = trainable_params(),
  cbs = NULL,
  metrics = NULL,
  path = NULL,
  model_dir = "models",
  wd = NULL,
  wd_bn_bias = FALSE,
  train_bn = TRUE
)

LMLearner(
  dls,
  model,
  alpha = 2,
  beta = 1,
  moms = list(0.8, 0.7, 0.8),
  loss_func = NULL,
  opt_func = Adam(),
  lr = 0.001,
  splitter = trainable_params(),
  cbs = NULL,
  metrics = NULL,
  path = NULL,
  model_dir = "models",
  wd = NULL,
  wd_bn_bias = FALSE,
  train_bn = TRUE
)
}
\arguments{
\item{dls}{dls}

\item{model}{model}

\item{alpha}{alpha}

\item{beta}{beta}

\item{moms}{moms}

\item{loss_func}{loss_func}

\item{opt_func}{opt_func}

\item{lr}{lr}

\item{splitter}{splitter}

\item{cbs}{cbs}

\item{metrics}{metrics}

\item{path}{path}

\item{model_dir}{model_dir}

\item{wd}{wd}

\item{wd_bn_bias}{wd_bn_bias}

\item{train_bn}{train_bn}
}
\description{
Add functionality to `TextLearner` when dealingwith a language model

Add functionality to `TextLearner` when dealing with a language model
}
\details{

}
