% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_learner.R
\name{LMLearner_predict}
\alias{LMLearner_predict}
\title{LMLearner_predict}
\usage{
LMLearner_predict(
  text,
  n_words = 1,
  no_unk = TRUE,
  temperature = 1,
  min_p = NULL,
  no_bar = FALSE,
  decoder = decode_spec_tokens(),
  only_last_word = FALSE
)
}
\arguments{
\item{text}{text}

\item{n_words}{n_words}

\item{no_unk}{no_unk}

\item{temperature}{temperature}

\item{min_p}{min_p}

\item{no_bar}{no_bar}

\item{decoder}{decoder}

\item{only_last_word}{only_last_word}
}
\value{
None
}
\description{
Return `text` and the `n_words` that come after
}
