% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{load_dataset}
\alias{load_dataset}
\title{Load_dataset}
\usage{
load_dataset(test_path, bs = 4, num_workers = 4)
}
\arguments{
\item{test_path}{test path (directory)}

\item{bs}{batch size}

\item{num_workers}{number of workers}
}
\value{
None
}
\description{
A helper function for getting a DataLoader for images in the folder `test_path`, with batch size `bs`, and number of workers `num_workers`
}
