% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_learner.R
\name{load_ignore_keys}
\alias{load_ignore_keys}
\title{Load_ignore_keys}
\usage{
load_ignore_keys(model, wgts)
}
\arguments{
\item{model}{model}

\item{wgts}{wgts}
}
\value{
None
}
\description{
Load `wgts` in `model` ignoring the names of the keys, just taking parameters in order
}
