% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{load_image}
\alias{load_image}
\title{Load_image}
\usage{
load_image(fn, mode = NULL)
}
\arguments{
\item{fn}{fn}

\item{mode}{mode}
}
\value{
None
}
\description{
Open and load a `PIL.Image` and convert to `mode`
}
