% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner_fits.R
\name{load_learner}
\alias{load_learner}
\title{Load_learner}
\usage{
load_learner(fname, cpu = TRUE)
}
\arguments{
\item{fname}{fname}

\item{cpu}{cpu or not}
}
\value{
learner object
}
\description{
Load a `Learner` object in `fname`, optionally putting it on the `cpu`
}
