% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_learner.R
\name{load_model_text}
\alias{load_model_text}
\title{load_model_text}
\usage{
load_model_text(
  file,
  model,
  opt,
  with_opt = NULL,
  device = NULL,
  strict = TRUE
)
}
\arguments{
\item{file}{file}

\item{model}{model}

\item{opt}{opt}

\item{with_opt}{with_opt}

\item{device}{device}

\item{strict}{strict}
}
\description{
Load `model` from `file` along with `opt` (if available, and if `with_opt`)
}
