% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_core.R
\name{load_tokenized_csv}
\alias{load_tokenized_csv}
\title{load_tokenized_csv}
\usage{
load_tokenized_csv(fname)
}
\arguments{
\item{fname}{fname}
}
\description{
Utility function to quickly load a tokenized csv and the corresponding counter
}
