% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_modules.R
\name{loaders}
\alias{loaders}
\title{loaders}
\usage{
loaders()
}
\description{
dict() -> new empty dictionary
}
\details{
dict(mapping) -> new dictionary initialized from a mapping object's (key, value) pairs
dict(iterable) -> new dictionary initialized as if via: d = {} for k, v in iterable: d[k] = v
dict(**kwargs) -> new dictionary initialized with the name=value pairs in the keyword argument list. For example: dict(one=1, two=2)
}
