% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{log.fastai.torch_core.TensorMask}
\alias{log.fastai.torch_core.TensorMask}
\title{Log}
\usage{
\method{log}{fastai.torch_core.TensorMask}(x, base = exp(1))
}
\arguments{
\item{x}{tensor}

\item{base}{base parameter}
}
\value{
tensor
}
\description{
Log
}
