% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{log}
\alias{log}
\alias{log.fastai.torch_core.TensorMask}
\alias{log.torch.Tensor}
\title{Log}
\usage{
\method{log}{fastai.torch_core.TensorMask}(a, b)

\method{log}{torch.Tensor}(a, b)
}
\description{
Log

Log
}
