% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{log1p.fastai.torch_core.TensorMask}
\alias{log1p.fastai.torch_core.TensorMask}
\title{Log1p}
\usage{
\method{log1p}{fastai.torch_core.TensorMask}(x)
}
\arguments{
\item{x}{tensor}
}
\value{
tensor
}
\description{
Log1p
}
