% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{log1p}
\alias{log1p}
\alias{log1p.fastai.torch_core.TensorMask}
\alias{log1p.torch.Tensor}
\title{Log1p}
\usage{
\method{log1p}{fastai.torch_core.TensorMask}(a, b)

\method{log1p}{torch.Tensor}(a, b)
}
\description{
Log1p

Log1p
}
