% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{log2}
\alias{log2}
\alias{log2.fastai.torch_core.TensorMask}
\alias{log2.torch.Tensor}
\title{Log2}
\usage{
log2.fastai.torch_core.TensorMask(a, b)

log2.torch.Tensor(a, b)
}
\description{
Log2

Log2
}
