% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{logical_and}
\alias{logical_and}
\alias{&.fastai.torch_core.TensorMask}
\alias{&.torch.Tensor}
\title{Logical_and}
\usage{
\method{&}{fastai.torch_core.TensorMask}(a, b)

\method{&}{torch.Tensor}(a, b)
}
\description{
Logical_and

Logical_and
}
