% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{logical_or}
\alias{logical_or}
\alias{|.fastai.torch_core.TensorMask}
\alias{!.fastai.torch_core.TensorMask}
\alias{|.torch.Tensor}
\alias{!.torch.Tensor}
\title{Logical_or}
\usage{
\method{|}{fastai.torch_core.TensorMask}(a, b)

\method{!}{fastai.torch_core.TensorMask}(a)

\method{|}{torch.Tensor}(a, b)

\method{!}{torch.Tensor}(a)
}
\description{
Logical_or

Logical_not

Logical_or

Logical_not
}
