% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{login}
\alias{login}
\title{Wandb login}
\usage{
login(anonymous = NULL, key = NULL, relogin = NULL, host = NULL, force = NULL)
}
\arguments{
\item{anonymous}{must,never,allow,false,true}

\item{key}{API key (secret)}

\item{relogin}{relogin or not}

\item{host}{host address}

\item{force}{whether to force a user to be logged into wandb when running a script}
}
\value{
None
}
\description{
Log in to W&B.
}
