% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{lr_find}
\alias{lr_find}
\title{Lr_find}
\usage{
lr_find(
  object,
  start_lr = 1e-07,
  end_lr = 10,
  num_it = 100,
  stop_div = TRUE,
  suggestions = TRUE,
  ...
)
}
\arguments{
\item{object}{learner}

\item{start_lr}{starting learning rate}

\item{end_lr}{end learning rate}

\item{num_it}{number of iterations}

\item{stop_div}{stop div or not}

\item{suggestions}{suggestions}

\item{...}{additional arguments to pass}
}
\value{
data frame
}
\description{
Launch a mock training to find a good learning rate, return lr_min, lr_steep if `suggestions` is TRUE
}
\examples{

\dontrun{

model \%>\% lr_find()
model \%>\% plot_lr_find(dpi = 200)

}

}
