% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{lr_find_}
\alias{lr_find_}
\title{lr_find}
\usage{
lr_find_(object)
}
\arguments{
\item{object}{learner}
}
\description{
Launch a mock training to find a good learning rate, return lr_min, lr_steep if `suggestions` is TRUE
}
