% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{make_vocab}
\alias{make_vocab}
\title{make_vocab}
\usage{
make_vocab(count, min_freq = 3, max_vocab = 60000, special_toks = NULL)
}
\arguments{
\item{count}{count}

\item{min_freq}{min_freq}

\item{max_vocab}{max_vocab}

\item{special_toks}{special_toks}
}
\description{
Create a vocab of `max_vocab` size from `Counter` `count` with items present more than `min_freq`
}
