% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{Mask_create}
\alias{Mask_create}
\title{Mask_create}
\usage{
Mask_create(enc = NULL, dec = NULL, split_idx = NULL, order = NULL)
}
\arguments{
\item{enc}{encoder}

\item{dec}{decoder}

\item{split_idx}{split by index}

\item{order}{order}
}
\value{
None
}
\description{
Delegates (`__call__`,`decode`,`setup`) to (`encodes`,`decodes`,`setups`) if `split_idx` matches
}
