% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{==.fastai.torch_core.TensorMask}
\alias{==.fastai.torch_core.TensorMask}
\title{Equal}
\usage{
\method{==}{fastai.torch_core.TensorMask}(a, b)
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}
}
\value{
tensor
}
\description{
Equal
}
\examples{
\dontrun{

aa = tensor(1:10)
aa == aa

}
}
