% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medical.R
\name{mask_from_blur}
\alias{mask_from_blur}
\title{Mask from blur}
\usage{
mask_from_blur(img, window, sigma = 0.3, thresh = 0.05, remove_max = TRUE)
}
\arguments{
\item{img}{image}

\item{window}{windowing effect}

\item{sigma}{sigma}

\item{thresh}{thresholf point}

\item{remove_max}{remove maximum or not}
}
\description{
Mask from blur
}
