% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_mask_rcnn.R
\name{mask_rcnn_predict_dl}
\alias{mask_rcnn_predict_dl}
\title{Mask RCNN predict dataloader}
\usage{
mask_rcnn_predict_dl(model, infer_dl, show_pbar = TRUE)
}
\arguments{
\item{model}{model}

\item{infer_dl}{infer_dl}

\item{show_pbar}{show_pbar}
}
\value{
None
}
\description{
Mask RCNN predict dataloader
}
