% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{mask_tensor}
\alias{mask_tensor}
\title{mask_tensor}
\usage{
mask_tensor(x, p = 0.5, neutral = 0, batch = FALSE)
}
\arguments{
\item{x}{x}

\item{p}{p}

\item{neutral}{neutral}

\item{batch}{batch}
}
\description{
Mask elements of `x` with `neutral` with probability `1-p`
}
