% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{MaskBlock}
\alias{MaskBlock}
\title{MaskBlock}
\usage{
MaskBlock(codes = NULL)
}
\arguments{
\item{codes}{codes}
}
\value{
block
}
\description{
A `TransformBlock` for segmentation masks, potentially with `codes`
}
