% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_models.R
\name{masked_concat_pool}
\alias{masked_concat_pool}
\title{masked_concat_pool}
\usage{
masked_concat_pool(output, mask, bptt)
}
\arguments{
\item{output}{output}

\item{mask}{mask}

\item{bptt}{bptt}
}
\description{
Pool `MultiBatchEncoder` outputs into one vector [last_hidden, max_pool, avg_pool]
}
