% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_spectrogram.R
\name{MaskFreq}
\alias{MaskFreq}
\title{Mask Freq}
\usage{
MaskFreq(num_masks = 1, size = 20, start = NULL, val = NULL)
}
\arguments{
\item{num_masks}{number of masks}

\item{size}{size}

\item{start}{starting point}

\item{val}{value}
}
\value{
None
}
\description{
Google SpecAugment frequency masking from https://arxiv.org/abs/1904.08779.
}
