% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_spectrogram.R
\name{MaskTime}
\alias{MaskTime}
\title{MaskTime}
\usage{
MaskTime(num_masks = 1, size = 20, start = NULL, val = NULL)
}
\arguments{
\item{num_masks}{number masks}

\item{size}{size}

\item{start}{start point}

\item{val}{val}
}
\value{
None
}
\description{
Google SpecAugment time masking from https://arxiv.org/abs/1904.08779.
}
