% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_learner.R
\name{match_embeds}
\alias{match_embeds}
\title{match_embeds}
\usage{
match_embeds(old_wgts, old_vocab, new_vocab)
}
\arguments{
\item{old_wgts}{old_wgts}

\item{old_vocab}{old_vocab}

\item{new_vocab}{new_vocab}
}
\description{
Convert the embedding in `old_wgts` to go from `old_vocab` to `new_vocab`.
}
