% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{matmul}
\alias{matmul}
\alias{\%*\%.fastai.torch_core.TensorMask}
\alias{\%*\%.torch.Tensor}
\title{Matmul}
\usage{
`\%*\%.fastai.torch_core.TensorMask`(a, b)

`\%*\%.torch.Tensor`(a, b)
}
\description{
Matmul

Matmul
}
