% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{MatthewsCorrCoefMulti}
\alias{MatthewsCorrCoefMulti}
\title{MatthewsCorrCoefMulti}
\usage{
MatthewsCorrCoefMulti(thresh = 0.5, sigmoid = TRUE, sample_weight = NULL)
}
\arguments{
\item{thresh}{thresh}

\item{sigmoid}{sigmoid}

\item{sample_weight}{sample_weight}
}
\description{
Matthews correlation coefficient for multi-label classification problems
}
