% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{max.fastai.torch_core.TensorMask}
\alias{max.fastai.torch_core.TensorMask}
\title{Max}
\usage{
\method{max}{fastai.torch_core.TensorMask}(a, ..., na.rm = FALSE)
}
\arguments{
\item{a}{tensor}

\item{...}{additional parameters}

\item{na.rm}{remove NAs}
}
\value{
tensor
}
\description{
Max
}
\examples{
\dontrun{

aa = tensor(1:10)
max(aa)

}

}
