% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{max}
\alias{max}
\alias{max.fastai.torch_core.TensorMask}
\alias{max.torch.Tensor}
\title{Max}
\usage{
\method{max}{fastai.torch_core.TensorMask}(a, ..., na.rm = FALSE)

\method{max}{torch.Tensor}(a, ..., na.rm = FALSE)
}
\description{
Max

Max
}
