% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{MaxPool}
\alias{MaxPool}
\title{MaxPool}
\usage{
MaxPool(ks = 2, stride = NULL, padding = 0, ndim = 2, ceil_mode = FALSE)
}
\arguments{
\item{ks}{ks}

\item{stride}{stride}

\item{padding}{padding}

\item{ndim}{ndim}

\item{ceil_mode}{ceil_mode}
}
\value{
None
}
\description{
nn.MaxPool layer for `ndim`
}
