% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_core.R
\name{maybe_unsqueeze}
\alias{maybe_unsqueeze}
\title{Maybe_unsqueeze}
\usage{
maybe_unsqueeze(x)
}
\arguments{
\item{x}{R array/matrix/tensor}
}
\value{
array
}
\description{
Add empty dimension if it is a rank 1 tensor/array
}
