% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_mixup_preds_rnn.R
\name{MCDropoutCallback}
\alias{MCDropoutCallback}
\title{MCDropoutCallback}
\usage{
MCDropoutCallback(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
Turns on dropout during inference, allowing you to call
Learner$get_preds multiple times to approximate your model
uncertainty using Monte Carlo Dropout. https://arxiv.org/pdf/1506.02142.pdf
}
