% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{median.fastai.torch_core.TensorMask}
\alias{median.fastai.torch_core.TensorMask}
\title{Median of tensor}
\usage{
\method{median}{fastai.torch_core.TensorMask}(a, ...)
}
\description{
Median of tensor
}
