% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{MergeLayer}
\alias{MergeLayer}
\title{MergeLayer}
\usage{
MergeLayer(dense = FALSE)
}
\arguments{
\item{dense}{dense}
}
\value{
None
}
\description{
Merge a shortcut with the result of the module by adding them or concatenating them if `dense=TRUE`.
}
