% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{min}
\alias{min}
\alias{min.torch.Tensor}
\title{Min}
\usage{
\method{min}{torch.Tensor}(a, ..., na.rm = FALSE)
}
\arguments{
\item{a}{tensor}

\item{...}{additional parameters}

\item{na.rm}{remove NAs}
}
\value{
tensor
}
\description{
Min
}
