% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{MishJitAutoFn}
\alias{MishJitAutoFn}
\title{MishJitAutoFn}
\usage{
MishJitAutoFn(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
Records operation history and defines formulas for differentiating ops.
}
