% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_mixup_preds_rnn.R
\name{MixHandler}
\alias{MixHandler}
\title{MixHandler}
\usage{
MixHandler(alpha = 0.5)
}
\arguments{
\item{alpha}{alpha}
}
\value{
None
}
\description{
A handler class for implementing `MixUp` style scheduling
}
