% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{mobilenet_v2}
\alias{mobilenet_v2}
\title{mobilenet_v2}
\usage{
mobilenet_v2(pretrained = FALSE, progress = TRUE)
}
\arguments{
\item{pretrained}{pretrained}

\item{progress}{progress}
}
\description{
Constructs a MobileNetV2 architecture from
}
\details{
`"MobileNetV2: Inverted Residuals and Linear Bottlenecks" <https://arxiv.org/abs/1801.04381>`_.
}
