% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_hook.R
\name{model_sizes}
\alias{model_sizes}
\title{Model_sizes}
\usage{
model_sizes(m, size = list(64, 64))
}
\arguments{
\item{m}{m parameter}

\item{size}{size}
}
\value{
None
}
\description{
Pass a dummy input through the model `m` to get the various sizes of activations.
}
