% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_mixup_preds_rnn.R
\name{ModelResetter}
\alias{ModelResetter}
\title{ModelResetter}
\usage{
ModelResetter(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
Callback that resets the model at each validation/training step
}
