% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{most_confused}
\alias{most_confused}
\title{Most_confused}
\usage{
most_confused(object, min_val = 1)
}
\arguments{
\item{object}{learner}

\item{min_val}{min_val}
}
\description{
Sorted descending list of largest non-diagonal entries of confusion matrix, presented as actual, predicted, number of occurrences.
}
