% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{mse}
\alias{mse}
\title{MSE}
\usage{
mse(inp, targ)
}
\arguments{
\item{inp}{predictions}

\item{targ}{targets}
}
\value{
None
}
\description{
Mean squared error between `inp` and `targ`.
}
\examples{

\dontrun{

model = dls \%>\% tabular_learner(layers=c(200,100,100,200),
metrics = list(mse(),rmse()) )

}

}
