% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{MultiCategorize}
\alias{MultiCategorize}
\title{MultiCategorize}
\usage{
MultiCategorize(vocab = NULL, add_na = FALSE)
}
\arguments{
\item{vocab}{vocab}

\item{add_na}{add_na}
}
\description{
Reversible transform of multi-category strings to `vocab` id
}
