% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{MultiCategoryBlock}
\alias{MultiCategoryBlock}
\title{MultiCategoryBlock}
\usage{
MultiCategoryBlock(encoded = FALSE, vocab = NULL, add_na = FALSE)
}
\arguments{
\item{encoded}{encoded}

\item{vocab}{vocab}

\item{add_na}{add_na}
}
\value{
Block object
}
\description{
`TransformBlock` for multi-label categorical targets
}
