% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{multiplygit add -A && git commit -m 'staging all files'}
\alias{multiplygit add -A && git commit -m 'staging all files'}
\alias{*.torch.Tensor}
\title{Multiply}
\usage{
\method{*}{torch.Tensor}(a, b)
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}
}
\value{
tensor
}
\description{
Multiply
}
