% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{MultiTargetLoss}
\alias{MultiTargetLoss}
\title{MultiTargetLoss}
\usage{
MultiTargetLoss(...)
}
\arguments{
\item{...}{additional arguments}
}
\value{
None
}
\description{
Provides the ability to apply different loss functions to multi-modal targets/predictions
}
