% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{nn}
\alias{nn}
\title{Core Pytorch module}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
nn
}
\description{
Core Pytorch module
}
\keyword{datasets}
