% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{nn_module}
\alias{nn_module}
\title{Fastai NN module}
\usage{
nn_module(model_fn, name = "Custom_Model", gpu = TRUE)
}
\arguments{
\item{model_fn}{pass custom model function}

\item{name}{set name for nn_module}

\item{gpu}{move model to GPU}
}
\value{
None
}
\description{
Fastai NN module
}
