% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{NoneReduce}
\alias{NoneReduce}
\title{NoneReduce}
\usage{
NoneReduce(loss_func)
}
\arguments{
\item{loss_func}{loss function}
}
\value{
None
}
\description{
A context manager to evaluate `loss_func` with none reduce.
}
