% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{norm_apply_denorm}
\alias{norm_apply_denorm}
\title{norm_apply_denorm}
\usage{
norm_apply_denorm(x, f, nrm)
}
\arguments{
\item{x}{x}

\item{f}{f}

\item{nrm}{nrm}
}
\description{
Normalize `x` with `nrm`, then apply `f`, then denormalize
}
\details{

}
