% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{Normalize}
\alias{Normalize}
\title{Normalize}
\usage{
Normalize(cat_names, cont_names)
}
\arguments{
\item{cat_names}{cat_names}

\item{cont_names}{cont_names}
}
\description{
Normalize the continuous variables.
}
