% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R
\name{Normalize_from_stats}
\alias{Normalize_from_stats}
\title{Normalize from stats}
\usage{
Normalize_from_stats(mean, std, dim = 1, ndim = 4, cuda = TRUE)
}
\arguments{
\item{mean}{mean}

\item{std}{standard deviation}

\item{dim}{dimension}

\item{ndim}{number of dimensions}

\item{cuda}{cuda or not}
}
\value{
list
}
\description{
Normalize from stats
}
