% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_tabular.R
\name{NormalizeTS}
\alias{NormalizeTS}
\title{NormalizeTS}
\usage{
NormalizeTS(enc = NULL, dec = NULL, split_idx = NULL, order = NULL)
}
\arguments{
\item{enc}{encoder}

\item{dec}{decoder}

\item{split_idx}{split by index}

\item{order}{order}
}
\value{
None
}
\description{
Normalize the x variables.
}
