% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{not__mask}
\alias{not__mask}
\alias{!.fastai.torch_core.TensorMask}
\title{Logical_not}
\usage{
\method{!}{fastai.torch_core.TensorMask}(x)
}
\arguments{
\item{x}{tensor}
}
\value{
tensor
}
\description{
Logical_not
}
