% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{not_equal_to_mask_}
\alias{not_equal_to_mask_}
\alias{!=.fastai.torch_core.TensorMask}
\title{Not equal}
\usage{
\method{!=}{fastai.torch_core.TensorMask}(a, b)
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}
}
\value{
tensor
}
\description{
Not equal
}
