% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_hook.R
\name{num_features_model}
\alias{num_features_model}
\title{Num_features_model}
\usage{
num_features_model(m)
}
\arguments{
\item{m}{m parameter}
}
\value{
None
}
\description{
Return the number of output features for `m`.
}
