% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{Numericalize}
\alias{Numericalize}
\title{Numericalize}
\usage{
Numericalize(
  vocab = NULL,
  min_freq = 3,
  max_vocab = 60000,
  special_toks = NULL,
  pad_tok = NULL
)
}
\arguments{
\item{vocab}{vocab}

\item{min_freq}{min_freq}

\item{max_vocab}{max_vocab}

\item{special_toks}{special_toks}

\item{pad_tok}{pad_tok}
}
\value{
None
}
\description{
Reversible transform of tokenized texts to numericalized ids
}
