% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{one_batch}
\alias{one_batch}
\title{One batch}
\usage{
one_batch(object, ..., convert = FALSE)
}
\arguments{
\item{object}{data loader}

\item{...}{additional parameters to pass}

\item{convert}{to R matrix}
}
\value{
tensor
}
\description{
One batch
}
\examples{

\dontrun{

# get batch from data loader
batch = dls \%>\% one_batch()

}

}
