% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_readers.R
\name{open_image}
\alias{open_image}
\title{open_image}
\usage{
open_image(fn, div = TRUE, convert_mode = "RGB", after_open = NULL, ...)
}
\arguments{
\item{fn}{fn}

\item{div}{div}

\item{convert_mode}{convert_mode}

\item{after_open}{after_open}

\item{cls}{cls}
}
\description{
Return `Image` object created from image in file `fn`.
}
\details{

}
