% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R
\name{open_mask}
\alias{open_mask}
\title{open_mask}
\usage{
open_mask(fn, div = FALSE, convert_mode = "L", after_open = NULL)
}
\arguments{
\item{fn}{fn}

\item{div}{div}

\item{convert_mode}{convert_mode}

\item{after_open}{after_open}
}
\description{
Return `ImageSegment` object create from mask in file `fn`. If `div`, divides pixel values by 255.
}
\details{

}
