% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_signal.R
\name{OpenAudio}
\alias{OpenAudio}
\title{OpenAudio}
\usage{
OpenAudio(items)
}
\arguments{
\item{items}{vector, items}
}
\value{
None
}
\description{
Transform that creates AudioTensors from a list of files.
}
