% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{optim_metric}
\alias{optim_metric}
\title{optim_metric}
\usage{
optim_metric(f, argname, bounds, tol = 0.01, do_neg = TRUE, get_x = FALSE)
}
\arguments{
\item{f}{f}

\item{argname}{argname}

\item{bounds}{bounds}

\item{tol}{tol}

\item{do_neg}{do_neg}

\item{get_x}{get_x}
}
\description{
Replace metric `f` with a version that optimizes argument `argname`
}
