% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{or_mask}
\alias{or_mask}
\alias{|.fastai.torch_core.TensorMask}
\title{Logical_or}
\usage{
\method{|}{fastai.torch_core.TensorMask}(a, b)
}
\arguments{
\item{x}{tensor}

\item{y}{tensor}
}
\value{
tensor
}
\description{
Logical_or
}
