% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner_fits.R
\name{os_environ_tpu}
\alias{os_environ_tpu}
\title{An environment supporting TPUs}
\usage{
os_environ_tpu(text = "COLAB_TPU_ADDR")
}
\arguments{
\item{text}{string to pass to environment}
}
\value{
None
}
\description{
An environment supporting TPUs
}
