% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{pad_conv_norm_relu}
\alias{pad_conv_norm_relu}
\title{Pad_conv_norm_relu}
\usage{
pad_conv_norm_relu(
  ch_in,
  ch_out,
  pad_mode,
  norm_layer,
  ks = 3,
  bias = TRUE,
  pad = 1,
  stride = 1,
  activ = TRUE,
  init = nn()$init$kaiming_normal_,
  init_gain = 0.02
)
}
\arguments{
\item{ch_in}{input}

\item{ch_out}{output}

\item{pad_mode}{padding mode}

\item{norm_layer}{normalization layer}

\item{ks}{kernel size}

\item{bias}{bias}

\item{pad}{padding}

\item{stride}{stride}

\item{activ}{activation}

\item{init}{initializer}

\item{init_gain}{init gain}
}
\value{
None
}
\description{
Pad_conv_norm_relu
}
