% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{pad_input}
\alias{pad_input}
\title{Pad_input}
\usage{
pad_input(
  samples,
  pad_idx = 1,
  pad_fields = 0,
  pad_first = FALSE,
  backwards = FALSE
)
}
\arguments{
\item{samples}{samples}

\item{pad_idx}{pad_idx}

\item{pad_fields}{pad_fields}

\item{pad_first}{pad_first}

\item{backwards}{backwards}
}
\value{
None
}
\description{
Function that collect `samples` and adds padding
}
