% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{pad_input_chunk}
\alias{pad_input_chunk}
\title{pad_input_chunk}
\usage{
pad_input_chunk(samples, pad_idx = 1L, pad_first = TRUE, seq_len = 72L)
}
\arguments{
\item{samples}{samples}

\item{pad_idx}{pad_idx}

\item{pad_first}{pad_first}

\item{seq_len}{seq_len}
}
\description{
Pad `samples` by adding padding by chunks of size `seq_len`
}
\details{

}
