% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{PadIfNeeded}
\alias{PadIfNeeded}
\title{PadIfNeeded}
\usage{
PadIfNeeded(
  min_height = 1024,
  min_width = 1024,
  pad_height_divisor = NULL,
  pad_width_divisor = NULL,
  border_mode = 4,
  value = NULL,
  mask_value = NULL,
  always_apply = FALSE,
  p = 1
)
}
\arguments{
\item{min_height}{min_height}

\item{min_width}{min_width}

\item{pad_height_divisor}{pad_height_divisor}

\item{pad_width_divisor}{pad_width_divisor}

\item{border_mode}{border_mode}

\item{value}{value}

\item{mask_value}{mask_value}

\item{always_apply}{always_apply}

\item{p}{p}
}
\description{
Pad side of the image / max if side is less than desired number.
}
\section{Targets}{

image, mask, bbox, keypoints
}

\section{Image types}{

uint8, float32
}

