% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{parallel}
\alias{parallel}
\title{parallel}
\usage{
parallel(f, items, ...)
}
\arguments{
\item{f}{f}

\item{items}{items}
}
\description{
Applies `func` in parallel to `items`, using `n_workers`
}
\details{

}
