% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_schedule.R
\name{ParamScheduler}
\alias{ParamScheduler}
\title{ParamScheduler}
\usage{
ParamScheduler(scheds)
}
\arguments{
\item{scheds}{scheds}
}
\value{
None
}
\description{
Schedule hyper-parameters according to `scheds`
}
