% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_parsers.R
\name{parse}
\alias{parse}
\title{Parse}
\usage{
parse(
  data_splitter = NULL,
  idmap = NULL,
  autofix = TRUE,
  show_pbar = TRUE,
  cache_filepath = NULL
)
}
\arguments{
\item{data_splitter}{How to split the parsed data, defaults to a [0.8, 0.2] random split.}

\item{idmap}{Maps from filenames to unique ids, pass an `IDMap()` if you need this information.}

\item{autofix}{autofix}

\item{show_pbar}{Whether or not to show a progress bar while parsing the data.}

\item{cache_filepath}{Path to save records in pickle format. Defaults to NULL, e.g. if the user does not specify a path, no saving nor loading happens.}
}
\value{
A list of records for each split defined by data_splitter.
}
\description{
Loops through all data points parsing the required fields.
}
