% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_parsers.R
\name{parsers_AreasMixin}
\alias{parsers_AreasMixin}
\title{AreasMixin}
\usage{
parsers_AreasMixin(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
Adds areas method to parser
}
