% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_parsers.R
\name{parsers_BBoxesMixin}
\alias{parsers_BBoxesMixin}
\title{BBoxesMixin}
\usage{
parsers_BBoxesMixin(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
Adds bboxes method to parser
}
