% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_parsers.R
\name{parsers_FasterRCNN}
\alias{parsers_FasterRCNN}
\title{Faster RCNN}
\usage{
parsers_FasterRCNN(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
Parser with required mixins for Faster RCNN.
}
