% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_parsers.R
\name{parsers_FilepathMixin}
\alias{parsers_FilepathMixin}
\title{FilepathMixin}
\usage{
parsers_FilepathMixin(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
Adds filepath method to parser
}
