% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_parsers.R
\name{parsers_ImageidMixin}
\alias{parsers_ImageidMixin}
\title{Imageid Mixin}
\usage{
parsers_ImageidMixin(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
Adds imageid method to parser
}
