% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_parsers.R
\name{parsers_IsCrowdsMixin}
\alias{parsers_IsCrowdsMixin}
\title{IsCrowdsMixin}
\usage{
parsers_IsCrowdsMixin(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
Adds iscrowds method to parser
}
