% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_parsers.R
\name{parsers_LabelsMixin}
\alias{parsers_LabelsMixin}
\title{LabelsMixin}
\usage{
parsers_LabelsMixin(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
Adds labels method to parser
}
