% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_parsers.R
\name{parsers_MaskRCNN}
\alias{parsers_MaskRCNN}
\title{Mask RCNN}
\usage{
parsers_MaskRCNN(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
Parser with required mixins for Mask RCNN.
}
