% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_parsers.R
\name{parsers_MasksMixin}
\alias{parsers_MasksMixin}
\title{MasksMixin}
\usage{
parsers_MasksMixin(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
Adds masks method to parser
}
