% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_parsers.R
\name{parsers_SizeMixin}
\alias{parsers_SizeMixin}
\title{SizeMixin}
\usage{
parsers_SizeMixin(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
Adds image_width_height method to parser
}
