% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_parsers.R
\name{parsers_voc}
\alias{parsers_voc}
\title{Voc parser}
\usage{
parsers_voc(annotations_dir, images_dir, class_map, masks_dir = NULL)
}
\arguments{
\item{annotations_dir}{annotations_dir}

\item{images_dir}{images_dir}

\item{class_map}{class_map}

\item{masks_dir}{masks_dir}
}
\value{
None
}
\description{
Voc parser
}
