% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{partial}
\alias{partial}
\title{Partial}
\usage{
partial(...)
}
\arguments{
\item{...}{additional arguments}
}
\value{
None
}
\description{
partial(func, *args, **keywords) - new function with partial application
}
\examples{

\dontrun{

generator = basic_generator(out_size = 64, n_channels = 3, n_extra_layers = 1)
critic    = basic_critic(in_size = 64, n_channels = 3, n_extra_layers = 1,
                         act_cls = partial(nn$LeakyReLU, negative_slope = 0.2))

}

}
