% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab.R
\name{pca}
\alias{pca}
\title{PCA}
\usage{
pca(object, k = 3, convert = TRUE)
}
\arguments{
\item{object}{an object to apply PCA}

\item{k}{number of dimensions}

\item{convert}{to R matrix}
}
\value{
tensor
}
\description{
Compute PCA of `x` with `k` dimensions.
}
