% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{PearsonCorrCoef}
\alias{PearsonCorrCoef}
\title{PearsonCorrCoef}
\usage{
PearsonCorrCoef(
  dim_argmax = NULL,
  activation = "no",
  thresh = NULL,
  to_np = FALSE,
  invert_arg = FALSE,
  flatten = TRUE
)
}
\arguments{
\item{dim_argmax}{dim_argmax}

\item{activation}{activation}

\item{thresh}{thresh}

\item{to_np}{to_np}

\item{invert_arg}{invert_arg}

\item{flatten}{flatten}
}
\value{
None
}
\description{
Pearson correlation coefficient for regression problem
}
