% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{PixelShuffle_ICNR}
\alias{PixelShuffle_ICNR}
\title{PixelShuffle_ICNR}
\usage{
PixelShuffle_ICNR(
  ni,
  nf = NULL,
  scale = 2,
  blur = FALSE,
  norm_type = 3,
  act_cls = nn$ReLU
)
}
\arguments{
\item{ni}{input shape}

\item{nf}{number of features / outputs}

\item{scale}{scale}

\item{blur}{blur}

\item{norm_type}{normalziation type}

\item{act_cls}{activation}
}
\value{
None
}
\description{
Upsample by `scale` from `ni` filters to `nf` (default `ni`), using `nn.PixelShuffle`.
}
