% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{plot.list}
\alias{plot.list}
\title{Plot tensor}
\usage{
plot.list(x, y, ..., dpi = 100)
}
\arguments{
\item{x}{x axis}

\item{y}{y axis}

\item{...}{additional arguments}

\item{dpi}{dots per inch}
}
\value{
None
}
\description{
Plot tensor
}
