% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_finder.R
\name{plot_bs_find}
\alias{plot_bs_find}
\title{Plot_bs_find}
\usage{
plot_bs_find(object, ..., dpi = 250)
}
\arguments{
\item{object}{model}

\item{...}{additional arguments}

\item{dpi}{dots per inch}
}
\value{
None
}
\description{
Plot_bs_find
}
