% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{plot_confusion_matrix}
\usage{
plot_confusion_matrix(
  interp,
  normalize = FALSE,
  title = "Confusion matrix",
  cmap = "Blues",
  norm_dec = 2,
  plot_txt = TRUE,
  figsize = c(19.2, 10.8),
  ...,
  dpi = 90
)
}
\arguments{
\item{interp}{interpretation object}

\item{normalize}{normalize}

\item{title}{title}

\item{cmap}{cmap}

\item{norm_dec}{norm_dec}

\item{plot_txt}{plot_txt}
}
\description{
Plot the confusion matrix, with `title` and using `cmap`.
}
