% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{plot_lr_find}
\alias{plot_lr_find}
\title{plot_lr_find}
\usage{
plot_lr_find(object, skip_end = 5, dpi = 250)
}
\arguments{
\item{skip_end}{skip_end}
}
\description{
Plot the result of an LR Finder test (won't work if you didn't do `learn.lr_find()` before)
}
