% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{plot_top_losses}
\alias{plot_top_losses}
\title{plot_top_losses}
\usage{
plot_top_losses(
  interp,
  k,
  largest = TRUE,
  figsize = c(19.2, 10.8),
  ...,
  dpi = NULL
)
}
\arguments{
\item{interp}{interpretation object}

\item{k}{k}

\item{largest}{largest}
}
\description{
plot_top_losses
}
