% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{+.torch.nn.modules.container.Sequential}
\alias{+.torch.nn.modules.container.Sequential}
\title{Add layers to Sequential}
\usage{
\method{+}{torch.nn.modules.container.Sequential}(a, b)
}
\arguments{
\item{a}{sequential model}

\item{b}{layer}
}
\value{
model
}
\description{
Add layers to Sequential
}
