% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{PointScaler}
\alias{PointScaler}
\title{PointScaler}
\usage{
PointScaler(do_scale = TRUE, y_first = FALSE)
}
\arguments{
\item{do_scale}{do_scale}

\item{y_first}{y_first}
}
\description{
Scale a tensor representing points
}
