% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{PooledSelfAttention2d}
\alias{PooledSelfAttention2d}
\title{PooledSelfAttention2d}
\usage{
PooledSelfAttention2d(n_channels)
}
\arguments{
\item{n_channels}{number of channels}
}
\value{
None
}
\description{
Pooled self attention layer for 2d.
}
