% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{PoolFlatten}
\alias{PoolFlatten}
\title{PoolFlatten}
\usage{
PoolFlatten(pool_type = "Avg")
}
\arguments{
\item{pool_type}{pooling type}
}
\value{
None
}
\description{
Combine `nn.AdaptiveAvgPool2d` and `Flatten`.
}
