% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_models.R
\name{PoolingLinearClassifier}
\alias{PoolingLinearClassifier}
\title{PoolingLinearClassifier}
\usage{
PoolingLinearClassifier(dims, ps, bptt, y_range = NULL)
}
\arguments{
\item{dims}{dims}

\item{ps}{ps}

\item{bptt}{bptt}

\item{y_range}{y_range}
}
\value{
None
}
\description{
Create a linear classifier with pooling
}
