% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{pow}
\alias{pow}
\alias{^.fastai.torch_core.TensorMask}
\alias{^.torch.Tensor}
\title{Pow}
\usage{
\method{^}{fastai.torch_core.TensorMask}(a, b)

\method{^}{torch.Tensor}(a, b)
}
\description{
Pow

Pow
}
