% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{pre_process_squad}
\alias{pre_process_squad}
\title{Pre_process_squad}
\usage{
pre_process_squad(row, hf_arch, hf_tokenizer)
}
\arguments{
\item{row}{row in dataframe}

\item{hf_arch}{architecture}

\item{hf_tokenizer}{tokenizer}
}
\value{
None
}
\description{
Pre_process_squad
}
