% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{Precision}
\alias{Precision}
\title{Precision}
\usage{
Precision(
  axis = -1,
  labels = NULL,
  pos_label = 1,
  average = "binary",
  sample_weight = NULL
)
}
\arguments{
\item{axis}{axis}

\item{labels}{labels}

\item{pos_label}{pos_label}

\item{average}{average}

\item{sample_weight}{sample_weight}
}
\value{
None
}
\description{
Precision for single-label classification problems
}
