% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict.fastai.learner.Learner}
\alias{predict.fastai.learner.Learner}
\title{Predict}
\usage{
\method{predict}{fastai.learner.Learner}(object, row, ...)
}
\arguments{
\item{object}{the model}

\item{row}{row}

\item{...}{additional arguments to pass}
}
\value{
data frame
}
\description{
Prediction on `item`, fully decoded, loss function decoded and probabilities
}
