% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_transform.R
\name{predict.fastai2.tabular.learner.TabularLearner}
\alias{predict.fastai2.tabular.learner.TabularLearner}
\title{Predict}
\usage{
\method{predict}{fastai2.tabular.learner.TabularLearner}(object, row)
}
\arguments{
\item{object}{the model}

\item{row}{row}
}
\description{
Prediction on `item`, fully decoded, loss function decoded and probabilities
}
