% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_config.R
\name{preprocess_audio_folder}
\alias{preprocess_audio_folder}
\title{Preprocess_audio_folder}
\usage{
preprocess_audio_folder(
  path,
  folders = NULL,
  output_dir = NULL,
  sample_rate = 16000,
  force_mono = TRUE,
  crop_signal_to = NULL
)
}
\arguments{
\item{path}{path}

\item{folders}{folders}

\item{output_dir}{output_dir}

\item{sample_rate}{sample_rate}

\item{force_mono}{force_mono}

\item{crop_signal_to}{crop_signal_to}
}
\value{
None
}
\description{
Preprocess audio files in `path` in parallel using `n_workers`
}
