% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_config.R
\name{PreprocessAudio}
\alias{PreprocessAudio}
\title{Preprocess Audio}
\usage{
PreprocessAudio(sample_rate = 16000, force_mono = TRUE, crop_signal_to = NULL)
}
\arguments{
\item{sample_rate}{sample rate}

\item{force_mono}{force mono or not}

\item{crop_signal_to}{int, crop signal}
}
\value{
None
}
\description{
Creates an audio tensor and run the basic preprocessing transforms on it.
}
\details{
Used while preprocessing the audios, this is not a `Transform`.
}
