% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{print.fastai.learner.Learner}
\alias{print.fastai.learner.Learner}
\title{Print model}
\usage{
\method{print}{fastai.learner.Learner}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{additional parameters to pass}
}
\value{
None
}
\description{
Print model
}
