% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{print.fastai.tabular.learner.TabularLearner}
\alias{print.fastai.tabular.learner.TabularLearner}
\title{Print tabular model}
\usage{
\method{print}{fastai.tabular.learner.TabularLearner}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{additional parameters to pass}
}
\value{
None
}
\description{
Print tabular model
}
