% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{qhadam_step}
\alias{qhadam_step}
\title{Qhadam_step}
\usage{
qhadam_step(p, lr, mom, sqr_mom, sqr_avg, nu_1, nu_2, step, grad_avg, eps, ...)
}
\arguments{
\item{p}{p}

\item{lr}{learning rate}

\item{mom}{momentum}

\item{sqr_mom}{sqr momentum}

\item{sqr_avg}{sqr average}

\item{nu_1}{nu_1}

\item{nu_2}{nu_2}

\item{step}{step}

\item{grad_avg}{gradient average}

\item{eps}{epsilon}

\item{...}{additional arguments to pass}
}
\value{
None
}
\description{
Qhadam_step
}
