% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_models.R
\name{QRNN}
\alias{QRNN}
\title{QRNN}
\usage{
QRNN(
  input_size,
  hidden_size,
  n_layers = 1,
  batch_first = TRUE,
  dropout = 0,
  bidirectional = FALSE,
  save_prev_x = FALSE,
  zoneout = 0,
  window = NULL,
  output_gate = TRUE
)
}
\arguments{
\item{input_size}{input_size}

\item{hidden_size}{hidden_size}

\item{n_layers}{n_layers}

\item{batch_first}{batch_first}

\item{dropout}{dropout}

\item{bidirectional}{bidirectional}

\item{save_prev_x}{save_prev_x}

\item{zoneout}{zoneout}

\item{window}{window}

\item{output_gate}{output_gate}
}
\description{
Apply a multiple layer Quasi-Recurrent Neural Network (QRNN) to an input sequence.
}
