% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_models.R
\name{QRNNLayer}
\alias{QRNNLayer}
\title{QRNNLayer}
\usage{
QRNNLayer(
  input_size,
  hidden_size = NULL,
  save_prev_x = FALSE,
  zoneout = 0,
  window = 1,
  output_gate = TRUE,
  batch_first = TRUE,
  backward = FALSE
)
}
\arguments{
\item{input_size}{input_size}

\item{hidden_size}{hidden_size}

\item{save_prev_x}{save_prev_x}

\item{zoneout}{zoneout}

\item{window}{window}

\item{output_gate}{output_gate}

\item{batch_first}{batch_first}

\item{backward}{backward}
}
\description{
Apply a single layer Quasi-Recurrent Neural Network (QRNN) to an input sequence.
}
\details{

}
