% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{radam_step}
\alias{radam_step}
\title{Radam_step}
\usage{
radam_step(p, lr, mom, step, sqr_mom, grad_avg, sqr_avg, eps, beta, ...)
}
\arguments{
\item{p}{p}

\item{lr}{learning rate}

\item{mom}{momentum}

\item{step}{step}

\item{sqr_mom}{sqr momentum}

\item{grad_avg}{grad average}

\item{sqr_avg}{sqr average}

\item{eps}{epsilon}

\item{beta}{beta}

\item{...}{additional arguments to pass}
}
\value{
None
}
\description{
Step for RAdam with `lr` on `p`
}
