% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{RandomBrightnessContrast}
\alias{RandomBrightnessContrast}
\title{RandomBrightnessContrast}
\usage{
RandomBrightnessContrast(
  brightness_limit = 0.2,
  contrast_limit = 0.2,
  brightness_by_max = TRUE,
  always_apply = FALSE,
  p = 0.5
)
}
\arguments{
\item{brightness_limit}{brightness_limit}

\item{contrast_limit}{contrast_limit}

\item{brightness_by_max}{brightness_by_max}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Randomly change brightness and contrast of the input image.
}
\section{Targets}{

image
}

\section{Image types}{

uint8, float32
}

