% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{RandomCrop}
\alias{RandomCrop}
\title{RandomCrop}
\usage{
RandomCrop(size, ...)
}
\arguments{
\item{size}{size}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Randomly crop an image to `size`
}
