% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{RandomResizedCrop}
\alias{RandomResizedCrop}
\title{RandomResizedCrop}
\usage{
RandomResizedCrop(
  size,
  min_scale = 0.08,
  ratio = list(0.75, 1.33333333333333),
  resamples = list(2, 0),
  val_xtra = 0.14
)
}
\arguments{
\item{size}{size}

\item{min_scale}{minimum scale}

\item{ratio}{ratio}

\item{resamples}{resamples}

\item{val_xtra}{validation xtra}
}
\value{
None
}
\description{
Picks a random scaled crop of an image and resize it to `size`
}
