% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_albumentations.R
\name{RandomSizedBBoxSafeCrop}
\alias{RandomSizedBBoxSafeCrop}
\title{RandomSizedBBoxSafeCrop}
\usage{
RandomSizedBBoxSafeCrop(
  height,
  width,
  erosion_rate = 0,
  interpolation = 1,
  always_apply = FALSE,
  p = 1
)
}
\arguments{
\item{height}{height}

\item{width}{width}

\item{erosion_rate}{erosion_rate}

\item{interpolation}{interpolation}

\item{always_apply}{always_apply}

\item{p}{p}
}
\value{
None
}
\description{
Crop a random part of the input and rescale it to some size without loss of bboxes.
}
\section{Targets}{

image, mask, bboxes
}

\section{Image types}{

uint8, float32
}

