% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{RandPair}
\alias{RandPair}
\title{RandPair}
\usage{
RandPair(itemsB)
}
\arguments{
\item{itemsB}{a random image from domain B}
}
\value{
None
}
\description{
Returns a random image from domain B, resulting in a random pair of images from domain A and B.
}
