% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{RandTransform}
\alias{RandTransform}
\title{RandTransform}
\usage{
RandTransform(p = 1, nm = NULL, before_call = NULL, ...)
}
\arguments{
\item{p}{p}

\item{nm}{nm}

\item{before_call}{before_call}
}
\description{
A transform that before_call its state at each `__call__`
}
