% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{ranger}
\alias{ranger}
\title{Ranger}
\usage{
ranger(
  p,
  lr,
  mom = 0.95,
  wd = 0.01,
  eps = 1e-06,
  sqr_mom = 0.99,
  beta = 0,
  decouple_wd = TRUE
)
}
\arguments{
\item{p}{p}

\item{lr}{learning rate}

\item{mom}{momentum}

\item{wd}{weight decay}

\item{eps}{epsilon}

\item{sqr_mom}{sqr momentum}

\item{beta}{beta}

\item{decouple_wd}{decouple weight decay}
}
\value{
None
}
\description{
Convenience method for `Lookahead` with `RAdam`
}
