% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{RatioResize}
\alias{RatioResize}
\title{RatioResize}
\usage{
RatioResize(max_sz, resamples = list(2, 0), ...)
}
\arguments{
\item{max_sz}{max_sz}

\item{resamples}{resamples}
}
\description{
Resizes the biggest dimension of an image to `max_sz` maintaining the aspect ratio
}
