% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_tabular.R
\name{ReadTSBatch}
\alias{ReadTSBatch}
\title{ReadTSBatch}
\usage{
ReadTSBatch(to)
}
\arguments{
\item{to}{output from TSDataTable function}
}
\value{
None
}
\description{
A transform that always take lists as items
}
