% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{RecallMulti}
\alias{RecallMulti}
\title{RecallMulti}
\usage{
RecallMulti(
  thresh = 0.5,
  sigmoid = TRUE,
  labels = NULL,
  pos_label = 1,
  average = "macro",
  sample_weight = NULL
)
}
\arguments{
\item{thresh}{thresh}

\item{sigmoid}{sigmoid}

\item{labels}{labels}

\item{pos_label}{pos_label}

\item{average}{average}

\item{sample_weight}{sample_weight}
}
\value{
None
}
\description{
Recall for multi-label classification problems
}
