% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fit}
\alias{predict}
\alias{median}
\alias{dev.off}
\alias{sd}
\alias{capture.output}
\alias{download.file}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{fit}}}

  \item{grDevices}{\code{\link[grDevices:dev]{dev.off}}}

  \item{stats}{\code{\link[stats]{median}}, \code{\link[stats]{predict}}, \code{\link[stats]{sd}}}

  \item{utils}{\code{\link[utils]{capture.output}}, \code{\link[utils]{download.file}}}
}}

