% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{RegressionBlock}
\alias{RegressionBlock}
\title{RegressionBlock}
\usage{
RegressionBlock(n_out = NULL)
}
\arguments{
\item{n_out}{n_out}
}
\value{
Block object
}
\description{
`TransformBlock` for float targets
}
