% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_preprocess.R
\name{RemoveSilence}
\alias{RemoveSilence}
\title{Remove Silence}
\usage{
RemoveSilence(remove_type = RemoveType$Trim$value, threshold = 20, pad_ms = 20)
}
\arguments{
\item{remove_type}{remove type from RemoveType module}

\item{threshold}{threshold point}

\item{pad_ms}{pad ms}
}
\value{
None
}
\description{
Split signal at points of silence greater than 2*pad_ms
}
