% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_core.R
\name{replace_maj}
\alias{replace_maj}
\title{Replace_maj}
\usage{
replace_maj(t)
}
\arguments{
\item{t}{text}
}
\value{
string
}
\description{
Replace tokens in ALL CAPS by their lower version and add `TK_UP` before.
}
