% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_core.R
\name{replace_wrep}
\alias{replace_wrep}
\title{Replace_wrep}
\usage{
replace_wrep(t)
}
\arguments{
\item{t}{text}
}
\value{
string
}
\description{
Replace word repetitions: word word word word -- TK_WREP 4 word
}
