% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_model.R
\name{res_block_1d}
\alias{res_block_1d}
\title{Res_block_1d}
\usage{
res_block_1d(nf, ks = c(5, 3))
}
\arguments{
\item{nf}{number of features}

\item{ks}{kernel size}
}
\value{
block
}
\description{
Resnet block as described in the paper.
}
