% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_preprocess.R
\name{Resample}
\alias{Resample}
\title{Resample}
\usage{
Resample(sr_new)
}
\arguments{
\item{sr_new}{input}
}
\value{
None
}
\description{
Resample using faster polyphase technique and avoiding FFT computation
}
