% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{reshape}
\alias{reshape}
\title{Reshape}
\usage{
reshape(x, h, w, resample = 0)
}
\arguments{
\item{x}{tensor}

\item{h}{height}

\item{w}{width}

\item{resample}{resample value}
}
\value{
None
}
\description{
resize x to (w,h)
}
