% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{resize_max}
\alias{resize_max}
\title{Resize_max}
\usage{
resize_max(img, resample = 0, max_px = NULL, max_h = NULL, max_w = NULL)
}
\arguments{
\item{resample}{resample}

\item{max_px}{max_px}

\item{max_h}{max_h}

\item{max_w}{max_w}

\item{x}{x}
}
\value{
None
}
\description{
`resize` `x` to `max_px`, or `max_h`, or `max_w`
}
