% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{ResizeBatch}
\alias{ResizeBatch}
\title{ResizeBatch}
\usage{
ResizeBatch(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
Reshape x to size, keeping batch dim the same size
}
