% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_signal.R
\name{ResizeSignal}
\alias{ResizeSignal}
\title{Resize Signal}
\usage{
ResizeSignal(duration, pad_mode = AudioPadType$Zeros)
}
\arguments{
\item{duration}{int, duration}

\item{pad_mode}{padding mode}
}
\value{
None
}
\description{
Crops signal to be length specified in ms by duration, padding if needed
}
