% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{ResNet}
\alias{ResNet}
\title{ResNet}
\usage{
ResNet(
  block,
  layers,
  num_classes = 1000,
  zero_init_residual = FALSE,
  groups = 1,
  width_per_group = 64,
  replace_stride_with_dilation = NULL,
  norm_layer = NULL
)
}
\arguments{
\item{block}{the blocks that need to passed to ResNet}

\item{layers}{the layers to pass to ResNet}

\item{num_classes}{the number of classes}

\item{zero_init_residual}{logical, initializer}

\item{groups}{the groups}

\item{width_per_group}{the width per group}

\item{replace_stride_with_dilation}{logical, replace stride with dilation}

\item{norm_layer}{norm_layer}
}
\description{
Base class for all neural network modules.
}
