% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{resnet34}
\alias{resnet34}
\title{resnet34}
\usage{
resnet34(pretrained = FALSE, progress)
}
\arguments{
\item{pretrained}{pretrained}

\item{progress}{progress}
}
\description{
ResNet-34 model from
}
\details{
`"Deep Residual Learning for Image Recognition" <https://arxiv.org/pdf/1512.03385.pdf>`_
}
