% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{resnet50}
\alias{resnet50}
\title{Resnet50}
\usage{
resnet50(pretrained = FALSE, progress)
}
\arguments{
\item{pretrained}{pretrained or not}

\item{progress}{to see progress bar or not}
}
\value{
model
}
\description{
Resnet50
}
\details{
"Deep Residual Learning for Image Recognition" <https://arxiv.org/pdf/1512.03385.pdf>
}
