% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{resnet_generator}
\alias{resnet_generator}
\title{Resnet_generator}
\usage{
resnet_generator(
  ch_in,
  ch_out,
  n_ftrs = 64,
  norm_layer = NULL,
  dropout = 0,
  n_blocks = 9,
  pad_mode = "reflection"
)
}
\arguments{
\item{ch_in}{input}

\item{ch_out}{output}

\item{n_ftrs}{filter}

\item{norm_layer}{normalziation layer}

\item{dropout}{dropout rate}

\item{n_blocks}{number of blocks}

\item{pad_mode}{paddoing mode}
}
\value{
None
}
\description{
Resnet_generator
}
