% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastaibuilt.R
\name{RetinaNet}
\alias{RetinaNet}
\title{RetinaNet}
\usage{
RetinaNet(
  encoder,
  n_classes,
  final_bias = 0,
  chs = 256,
  n_anchors = 9,
  flatten = TRUE
)
}
\arguments{
\item{encoder}{encoder}

\item{n_classes}{n_classes}

\item{final_bias}{final_bias}

\item{chs}{chs}

\item{n_anchors}{n_anchors}

\item{flatten}{flatten}
}
\description{
Implements RetinaNet from https://arxiv.org/abs/1708.02002
}
\details{

}
