% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{rgb2hsv}
\alias{rgb2hsv}
\title{rgb2hsv}
\usage{
rgb2hsv(img)
}
\arguments{
\item{img}{img}
}
\description{
Converts a RGB image to an HSV image.
}
\details{
Note: Will not work on logit space images.
}
