% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{rmse}
\alias{rmse}
\title{RMSE}
\usage{
rmse(preds, targs)
}
\arguments{
\item{preds}{predictions}

\item{targs}{targets}
}
\value{
None
}
\description{
Root mean squared error
}
\examples{

\dontrun{

model = dls \%>\% tabular_learner(layers=c(200,100,100,200),
metrics = list(mse(),rmse()) )

}

}
