% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_mixup_preds_rnn.R
\name{RNNRegularizer}
\alias{RNNRegularizer}
\title{RNNRegularizer}
\usage{
RNNRegularizer(alpha = 0, beta = 0)
}
\arguments{
\item{alpha}{alpha}

\item{beta}{beta}
}
\value{
None
}
\description{
`Callback` that adds AR and TAR regularization in RNN training
}
