% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{RocAuc}
\alias{RocAuc}
\title{RocAuc}
\usage{
RocAuc(
  axis = -1,
  average = "macro",
  sample_weight = NULL,
  max_fpr = NULL,
  multi_class = "ovr"
)
}
\arguments{
\item{axis}{axis}

\item{average}{average}

\item{sample_weight}{sample_weight}

\item{max_fpr}{max_fpr}

\item{multi_class}{multi_class}
}
\description{
Area Under the Receiver Operating Characteristic
Curve for single-label multiclass classification problems
}
