% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{RocAucBinary}
\alias{RocAucBinary}
\title{RocAucBinary}
\usage{
RocAucBinary(
  axis = -1,
  average = "macro",
  sample_weight = NULL,
  max_fpr = NULL,
  multi_class = "raise"
)
}
\arguments{
\item{axis}{axis}

\item{average}{average}

\item{sample_weight}{sample_weight}

\item{max_fpr}{max_fpr}

\item{multi_class}{multi_class}
}
\value{
None
}
\description{
Area Under the Receiver Operating Characteristic Curve for single-label binary classification problems
}
\examples{

\dontrun{

model = dls \%>\% tabular_learner(layers=c(200,100,100,200),
config = tabular_config(embed_p = 0.3, use_bn = FALSE),
metrics = list(accuracy, RocAucBinary(),
               Precision(), Recall(),
               F1Score()))

}

}
