% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R
\name{Rotate}
\alias{Rotate}
\title{Rotate}
\usage{
Rotate(
  max_deg = 10,
  p = 0.5,
  draw = NULL,
  size = NULL,
  mode = "bilinear",
  pad_mode = "reflection",
  align_corners = TRUE,
  batch = FALSE
)
}
\arguments{
\item{max_deg}{max_deg}

\item{p}{p}

\item{draw}{draw}

\item{size}{size}

\item{mode}{mode}

\item{pad_mode}{pad_mode}

\item{align_corners}{align_corners}

\item{batch}{batch}
}
\description{
Apply a random rotation of at most `max_deg` with probability `p` to a batch of images
}
\details{

}
