% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{rotate_mat}
\alias{rotate_mat}
\title{Rotate_mat}
\usage{
rotate_mat(x, max_deg = 10, p = 0.5, draw = NULL, batch = FALSE)
}
\arguments{
\item{x}{x}

\item{max_deg}{max_deg}

\item{p}{p}

\item{draw}{draw}

\item{batch}{batch}
}
\value{
None
}
\description{
Return a random rotation matrix with `max_deg` and `p`
}
