% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{round.fastai.torch_core.TensorMask}
\alias{round.fastai.torch_core.TensorMask}
\title{Round}
\usage{
\method{round}{fastai.torch_core.TensorMask}(x, digits = 0)
}
\arguments{
\item{x}{tensor}

\item{digits}{decimal}
}
\value{
tensor
}
\description{
Round
}
