% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{round}
\alias{round}
\alias{round.torch.Tensor}
\title{Round}
\usage{
\method{round}{torch.Tensor}(x, digits = 0)
}
\arguments{
\item{x}{tensor}

\item{digits}{decimal}
}
\value{
tensor
}
\description{
Round
}
