% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{Saturation}
\alias{Saturation}
\title{Saturation}
\usage{
Saturation(max_lighting = 0.2, p = 0.75, draw = NULL, batch = FALSE)
}
\arguments{
\item{max_lighting}{max_lighting}

\item{p}{p}

\item{draw}{draw}

\item{batch}{batch}
}
\value{
None
}
\description{
Apply change in saturation of `max_lighting` to batch of images with probability `p`.
}
