% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_schedule.R
\name{SchedCos}
\alias{SchedCos}
\title{SchedCos}
\usage{
SchedCos(start, end)
}
\arguments{
\item{start}{start}

\item{end}{end}
}
\value{
None
}
\description{
Cosine schedule function from `start` to `end`
}
