% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_schedule.R
\name{SchedExp}
\alias{SchedExp}
\title{SchedExp}
\usage{
SchedExp(start, end)
}
\arguments{
\item{start}{start}

\item{end}{end}
}
\value{
None
}
\description{
Exponential schedule function from `start` to `end`
}
