% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_schedule.R
\name{SchedLin}
\alias{SchedLin}
\title{SchedLin}
\usage{
SchedLin(start, end)
}
\arguments{
\item{start}{start}

\item{end}{end}
}
\value{
None
}
\description{
Linear schedule function from `start` to `end`
}
