% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_schedule.R
\name{SchedNo}
\alias{SchedNo}
\title{SchedNo}
\usage{
SchedNo(start, end)
}
\arguments{
\item{start}{start}

\item{end}{end}
}
\value{
None
}
\description{
Constant schedule function with `start` value
}
