% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_schedule.R
\name{SchedPoly}
\alias{SchedPoly}
\title{SchedPoly}
\usage{
SchedPoly(start, end, power)
}
\arguments{
\item{start}{start}

\item{end}{end}

\item{power}{power}
}
\value{
None
}
\description{
Polynomial schedule (of `power`) function from `start` to `end`
}
