% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_data.R
\name{SegmentationDataLoaders_from_label_func}
\alias{SegmentationDataLoaders_from_label_func}
\title{from_label_func}
\usage{
SegmentationDataLoaders_from_label_func(
  path,
  fnames,
  label_func,
  valid_pct = 0.2,
  seed = NULL,
  codes = NULL,
  item_tfms = NULL,
  batch_tfms = NULL,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  device = NULL
)
}
\arguments{
\item{path}{path}

\item{fnames}{fnames}

\item{label_func}{label_func}

\item{valid_pct}{valid_pct}

\item{seed}{seed}

\item{codes}{codes}

\item{item_tfms}{item_tfms}

\item{batch_tfms}{batch_tfms}

\item{bs}{bs}

\item{val_bs}{val_bs}

\item{shuffle_train}{shuffle_train}

\item{device}{device}

\item{cls}{cls}
}
\description{
Create from list of `fnames` in `path`s with `label_func`.
}
\details{

}
