% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{SelfAttention}
\alias{SelfAttention}
\title{SelfAttention}
\usage{
SelfAttention(n_channels)
}
\arguments{
\item{n_channels}{number of channels}
}
\value{
None
}
\description{
Self attention layer for `n_channels`.
}
