% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{SEModule}
\alias{SEModule}
\title{SEModule}
\usage{
SEModule(ch, reduction, act_cls = nn$ReLU)
}
\arguments{
\item{ch}{ch}

\item{reduction}{reduction}

\item{act_cls}{activation}
}
\value{
None
}
\description{
SEModule
}
