% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{SeparableBlock}
\alias{SeparableBlock}
\title{SeparableBlock}
\usage{
SeparableBlock(expansion, ni, nf, reduction = 16, stride = 1, base_width = 4)
}
\arguments{
\item{expansion}{decoder}

\item{ni}{number of inputs}

\item{nf}{number of features}

\item{reduction}{number of reduction}

\item{stride}{number of stride}

\item{base_width}{base width}
}
\value{
Block object
}
\description{
SeparableBlock
}
