% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{sgd_step}
\alias{sgd_step}
\title{Sgd_step}
\usage{
sgd_step(p, lr, ...)
}
\arguments{
\item{p}{p}

\item{lr}{learning rate}

\item{...}{additional arguments to pass}
}
\value{
None
}
\description{
Sgd_step
}
\examples{

\dontrun{

tst_param = function(val, grad = NULL) {
  "Create a tensor with `val` and a gradient of `grad` for testing"
  res = tensor(val) \%>\% float()

  if(is.null(grad)) {
    grad = tensor(val / 10)
  } else {
    grad = tensor(grad)
  }

  res$grad = grad \%>\% float()
  res
}
p = tst_param(1., 0.1)
sgd_step(p, 1.)

}

}
