% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_spectrogram.R
\name{SGRoll}
\alias{SGRoll}
\title{SGRoll}
\usage{
SGRoll(max_shift_pct = 0.5, direction = 0)
}
\arguments{
\item{max_shift_pct}{maximum shift percentage}

\item{direction}{direction}
}
\value{
None
}
\description{
Shifts spectrogram along x-axis wrapping around to other side
}
