% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_model.R
\name{Shortcut}
\alias{Shortcut}
\title{Shortcut}
\usage{
Shortcut(ni, nf, act_fn = nn$ReLU(inplace = TRUE))
}
\arguments{
\item{ni}{number of input channels}

\item{nf}{number of features}

\item{act_fn}{activation}
}
\value{
None
}
\description{
Merge a shortcut with the result of the module by adding them. Adds Conv, BN and ReLU
}
