% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R, R/callbacks_training.R
\name{ShortEpochCallback}
\alias{ShortEpochCallback}
\title{ShortEpochCallback}
\usage{
ShortEpochCallback(pct = 0.01, short_valid = TRUE)

ShortEpochCallback(pct = 0.01, short_valid = TRUE)
}
\arguments{
\item{pct}{percentage}

\item{short_valid}{short_valid or not}

\item{...}{parameters to pass}
}
\value{
None

None
}
\description{
Fit just `pct` of an epoch, then stop
}
