% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{show}
\alias{show}
\title{Show}
\usage{
show(img, frames = 1, scale = TRUE, ...)
}
\arguments{
\item{img}{image object}

\item{frames}{number of frames}

\item{scale}{scale}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Adds functionality to view dicom images where each file may have more than 1 frame
}
