% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_core.R
\name{show_array}
\alias{show_array}
\title{Show_array}
\usage{
show_array(
  array,
  ax = NULL,
  figsize = NULL,
  title = NULL,
  ctx = NULL,
  tx = NULL
)
}
\arguments{
\item{array}{R array}

\item{ax}{axis}

\item{figsize}{figure size}

\item{title}{title, text}

\item{ctx}{ctx}

\item{tx}{tx}
}
\value{
None
}
\description{
Show an array on `ax`.
}
\examples{

\dontrun{

arr = as.array(1:10)
show_array(arr,title = 'My R array') \%>\% plot(dpi = 200)

}



}
