% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{show_images}
\alias{show_images}
\title{show_images}
\usage{
show_images(
  ims,
  nrows = 1,
  ncols = NULL,
  titles = NULL,
  figsize = NULL,
  imsize = 3,
  add_vert = 0
)
}
\arguments{
\item{ims}{ims}

\item{nrows}{nrows}

\item{ncols}{ncols}

\item{titles}{titles}

\item{figsize}{figsize}

\item{imsize}{imsize}

\item{add_vert}{add_vert}
}
\description{
Show all images `ims` as subplots with `rows` using `titles`
}
