% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{show_results}
\alias{show_results}
\title{Show_results}
\usage{
show_results(
  object,
  ds_idx = 1,
  dl = NULL,
  max_n = 9,
  shuffle = TRUE,
  dpi = 90,
  ...
)
}
\arguments{
\item{object}{model}

\item{ds_idx}{ds_idx}

\item{dl}{dl}

\item{max_n}{max_n}

\item{shuffle}{shuffle}

\item{dpi}{dots per inch}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Show some predictions on `ds_idx`-th dataset or `dl`
}
