% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icevision_utils.R
\name{show_samples}
\alias{show_samples}
\title{show_samples}
\usage{
show_samples(
  samples,
  class_map = NULL,
  denormalize_fn = denormalize_imagenet(),
  display_label = TRUE,
  display_bbox = TRUE,
  display_mask = TRUE,
  ncols = 1,
  figsize = NULL,
  show = FALSE,
  dpi = 100
)
}
\arguments{
\item{samples}{samples}

\item{class_map}{class_map}

\item{denormalize_fn}{denormalize_fn}

\item{display_label}{display_label}

\item{display_bbox}{display_bbox}

\item{display_mask}{display_mask}

\item{ncols}{ncols}

\item{figsize}{figsize}

\item{show}{show}

\item{dpi}{dots per inch}
}
\value{
None
}
\description{
show_samples
}
