% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{ShowCycleGANImgsCallback}
\alias{ShowCycleGANImgsCallback}
\title{ShowCycleGANImgsCallback}
\usage{
ShowCycleGANImgsCallback(imgA = FALSE, imgB = TRUE, show_img_interval = 10)
}
\arguments{
\item{imgA}{img from A domain}

\item{imgB}{img from B domain}

\item{show_img_interval}{show image interval}
}
\value{
None
}
\description{
Update the progress bar with input and prediction images
}
