% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid}
\usage{
sigmoid(input, eps = 1e-07)
}
\arguments{
\item{input}{inputs}

\item{eps}{epsilon}
}
\value{
None
}
\description{
Same as `torch$sigmoid`, plus clamping to `(eps,1-eps)
}
