% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{sigmoid_}
\alias{sigmoid_}
\title{Sigmoid_}
\usage{
sigmoid_(input, eps = 1e-07)
}
\arguments{
\item{input}{input}

\item{eps}{eps}
}
\value{
None
}
\description{
Same as `torch$sigmoid_`, plus clamping to `(eps,1-eps)
}
