% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{sigmoid_range}
\alias{sigmoid_range}
\title{Sigmoid_range}
\usage{
sigmoid_range(x, low, high)
}
\arguments{
\item{x}{tensor}

\item{low}{low value}

\item{high}{high value}
}
\value{
None
}
\description{
Sigmoid function with range `(low, high)`
}
