% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{SigmoidRange}
\alias{SigmoidRange}
\title{SigmoidRange}
\usage{
SigmoidRange(low, high)
}
\arguments{
\item{low}{low value}

\item{high}{high value}
}
\value{
None
}
\description{
Sigmoid module with range `(low, high)`
}
