% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_signal.R
\name{SignalCutout}
\alias{SignalCutout}
\title{Signal Cutout}
\usage{
SignalCutout(p = 0.5, max_cut_pct = 0.15)
}
\arguments{
\item{p}{p}

\item{max_cut_pct}{max cut percentage}
}
\value{
None
}
\description{
Randomly zeros some portion of the signal
}
