% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_signal.R
\name{SignalLoss}
\alias{SignalLoss}
\title{Signal Loss}
\usage{
SignalLoss(p = 0.5, max_loss_pct = 0.15)
}
\arguments{
\item{p}{probability}

\item{max_loss_pct}{max loss percentage}
}
\value{
None
}
\description{
Randomly loses some portion of the signal
}
