% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_signal.R
\name{SignalShifter}
\alias{SignalShifter}
\title{Signal Shifter}
\usage{
SignalShifter(
  p = 0.5,
  max_pct = 0.2,
  max_time = NULL,
  direction = 0,
  roll = FALSE
)
}
\arguments{
\item{p}{probability}

\item{max_pct}{max percentage}

\item{max_time}{maximum time}

\item{direction}{direction}

\item{roll}{roll or not}
}
\value{
None
}
\description{
Randomly shifts the audio signal by `max_pct` %.
}
\details{
direction must be -1(left) 0(bidirectional) or 1(right).
}
