% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{SimpleCNN}
\alias{SimpleCNN}
\title{SimpleCNN}
\usage{
SimpleCNN(filters, kernel_szs = NULL, strides = NULL, bn = TRUE)
}
\arguments{
\item{filters}{filters number}

\item{kernel_szs}{kernel size}

\item{strides}{strides}

\item{bn}{bn}
}
\value{
None
}
\description{
Create a simple CNN with `filters`.
}
