% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{SimpleSelfAttention}
\alias{SimpleSelfAttention}
\title{SimpleSelfAttention}
\usage{
SimpleSelfAttention(n_in, ks = 1, sym = FALSE)
}
\arguments{
\item{n_in}{inputs}

\item{ks}{kernel size}

\item{sym}{sym}
}
\value{
None
}
\description{
Same as `nn()$Module`, but no need for subclasses to call `super()$__init__`
}
