% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{skm_to_fastai}
\alias{skm_to_fastai}
\title{skm_to_fastai}
\usage{
skm_to_fastai(
  func,
  is_class = TRUE,
  thresh = NULL,
  axis = -1,
  activation = NULL,
  ...
)
}
\arguments{
\item{func}{func}

\item{is_class}{is_class}

\item{thresh}{thresh}

\item{axis}{axis}

\item{activation}{activation}
}
\description{
Convert `func` from sklearn.metrics to a fastai metric
}
\details{

}
