% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{slice}
\alias{slice}
\title{Slice}
\usage{
slice(...)
}
\value{
sliced object
}
\description{
Slice
}
\details{
slice(start, stop[, step]) Create a slice object. This is used for extended slicing (e.g. a[0:10:2]).
}
