% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{sort.fastai.torch_core.TensorMask}
\alias{sort.fastai.torch_core.TensorMask}
\title{Sort}
\usage{
\method{sort}{fastai.torch_core.TensorMask}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{tensor}

\item{decreasing}{the order}

\item{...}{additional parameters to pass}
}
\value{
tensor
}
\description{
Sort
}
