% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{sort}
\alias{sort}
\alias{sort.fastai.torch_core.TensorMask}
\alias{sort.torch.Tensor}
\title{Sort}
\usage{
\method{sort}{fastai.torch_core.TensorMask}(x, decreasing = FALSE, ...)

\method{sort}{torch.Tensor}(x, decreasing = FALSE, ...)
}
\description{
Sort

Sort
}
