% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{SortedDL}
\alias{SortedDL}
\title{SortedDL}
\usage{
SortedDL(
  dataset,
  sort_func = NULL,
  res = NULL,
  bs = 64,
  shuffle = FALSE,
  num_workers = NULL,
  verbose = FALSE,
  do_setup = TRUE,
  pin_memory = FALSE,
  timeout = 0,
  batch_size = NULL,
  drop_last = FALSE,
  indexed = NULL,
  n = NULL,
  device = NULL
)
}
\arguments{
\item{dataset}{dataset}

\item{sort_func}{sort_func}

\item{res}{res}

\item{bs}{bs}

\item{shuffle}{shuffle}

\item{num_workers}{num_workers}

\item{verbose}{verbose}

\item{do_setup}{do_setup}

\item{pin_memory}{pin_memory}

\item{timeout}{timeout}

\item{batch_size}{batch_size}

\item{drop_last}{drop_last}

\item{indexed}{indexed}

\item{n}{n}

\item{device}{device}
}
\description{
A `DataLoader` that goes throught the item in the order given by `sort_func`
}
\details{

}
