% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_core.R
\name{SpacyTokenizer}
\alias{SpacyTokenizer}
\title{SpacyTokenizer}
\usage{
SpacyTokenizer(lang = "en", special_toks = NULL, buf_sz = 5000)
}
\arguments{
\item{lang}{language}

\item{special_toks}{special tokenizers}

\item{buf_sz}{buffer size}
}
\value{
none
}
\description{
Spacy tokenizer for `lang`
}
