% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{SpearmanCorrCoef}
\alias{SpearmanCorrCoef}
\title{SpearmanCorrCoef}
\usage{
SpearmanCorrCoef(
  dim_argmax = NULL,
  axis = 0,
  nan_policy = "propagate",
  activation = "no",
  thresh = NULL,
  to_np = FALSE,
  invert_arg = FALSE,
  flatten = TRUE
)
}
\arguments{
\item{dim_argmax}{dim_argmax}

\item{axis}{axis}

\item{nan_policy}{nan_policy}

\item{activation}{activation}

\item{thresh}{thresh}

\item{to_np}{to_np}

\item{invert_arg}{invert_arg}

\item{flatten}{flatten}
}
\description{
Spearman correlation coefficient for regression problem
}
