% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_spectrogram.R
\name{SpectrogramTransformer}
\alias{SpectrogramTransformer}
\title{Spectrogram Transformer}
\usage{
SpectrogramTransformer(mel = TRUE, to_db = TRUE)
}
\arguments{
\item{mel}{mel-spectrogram or not}

\item{to_db}{to decibels}
}
\value{
None
}
\description{
Creates a factory for creating AudioToSpec
}
\details{
transforms with different parameters
}
