% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{split_by_idx}
\alias{split_by_idx}
\title{Split by idx}
\usage{
split_by_idx(object, valid_idx)
}
\arguments{
\item{valid_idx}{valid_idx}
}
\description{
Split the data according to the indexes in `valid_idx`.
}
