% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{sqrd}
\alias{sqrd}
\alias{sqrt.fastai.torch_core.TensorMask}
\alias{sqrt.torch.Tensor}
\title{Sqrt}
\usage{
\method{sqrt}{fastai.torch_core.TensorMask}(a)

\method{sqrt}{torch.Tensor}(a)
}
\description{
Sqrt

Sqrt
}
