% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{squeezenet1_1}
\alias{squeezenet1_1}
\title{squeezenet1_1}
\usage{
squeezenet1_1(pretrained = FALSE, progress = TRUE)
}
\arguments{
\item{pretrained}{pretrained}

\item{progress}{progress}
}
\description{
SqueezeNet 1.1 model from the `official SqueezeNet repo
}
\details{
<https://github.com/DeepScale/SqueezeNet/tree/master/SqueezeNet_v1.1>`_.
SqueezeNet 1.1 has 2.4x less computation and slightly fewer parameters
than SqueezeNet 1.0, without sacrificing accuracy.
}
