% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_data.R
\name{stack_train_valid}
\alias{stack_train_valid}
\title{Stack_train_valid}
\usage{
stack_train_valid(df_train, df_valid)
}
\arguments{
\item{df_train}{train data}

\item{df_valid}{validation data}
}
\value{
data frame
}
\description{
Stack df_train and df_valid, adds `valid_col`=TRUE/FALSE for df_valid/df_train
}
