% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{step_stat}
\alias{step_stat}
\title{step_stat}
\usage{
step_stat(p, step = 0, ...)
}
\arguments{
\item{p}{p}

\item{step}{step}

\item{...}{additional args to pass}
}
\description{
Register the number of steps done in `state` for `p`
}
