% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{sub_mask}
\alias{sub_mask}
\alias{-.fastai.torch_core.TensorMask}
\title{Sub}
\usage{
\method{-}{fastai.torch_core.TensorMask}(a, b)
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}
}
\value{
tensor
}
\description{
Sub
}
