% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{subplots}
\alias{subplots}
\title{Subplots}
\usage{
subplots(nrows = 2, ncols = 2, figsize = NULL, imsize = 4, add_vert = 0)
}
\arguments{
\item{nrows}{number of rows}

\item{ncols}{number of columns}

\item{figsize}{figure size}

\item{imsize}{image size}

\item{add_vert}{add vertical}
}
\value{
plot object
}
\description{
Subplots
}
