% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{summarization_splitter}
\alias{summarization_splitter}
\title{Summarization_splitter}
\usage{
summarization_splitter(m, arch)
}
\arguments{
\item{m}{splitter parameter}

\item{arch}{architecture}
}
\value{
None
}
\description{
Custom param splitter for summarization models
}
