% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{summary.fastai.basic_train.Learner}
\alias{summary.fastai.basic_train.Learner}
\title{Summary}
\usage{
\method{summary}{fastai.basic_train.Learner}(m, n = 70)
}
\arguments{
\item{m}{m}
}
\description{
Print a summary of `m` using a output text width of `n` chars
}
