% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{summary.fastai.learner.Learner}
\alias{summary.fastai.learner.Learner}
\title{Summary}
\usage{
\method{summary}{fastai.learner.Learner}(object, ...)
}
\arguments{
\item{object}{model}

\item{...}{additional arguments to pass}
}
\value{
None
}
\description{
Summary
}
\examples{

\dontrun{

summary(model)

}

}
