% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R, R/tabular.R
\name{summary.fastai.tabular.learner.TabularLearner}
\alias{summary.fastai.tabular.learner.TabularLearner}
\title{Summary}
\usage{
\method{summary}{fastai.tabular.learner.TabularLearner}(object, ...)

\method{summary}{fastai.tabular.learner.TabularLearner}(object, ...)
}
\arguments{
\item{object}{model}

\item{...}{additional parameters to pass}
}
\value{
None

None
}
\description{
Print a summary of `m` using a output text width of `n` chars
}
\examples{

\dontrun{

summary(model)

}

}
