% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{summary.fastai2.tabular.learner.TabularLearner}
\alias{summary.fastai2.tabular.learner.TabularLearner}
\title{Summary}
\usage{
\method{summary}{fastai2.tabular.learner.TabularLearner}(object)
}
\arguments{
\item{m}{m}
}
\description{
Print a summary of `m` using a output text width of `n` chars
}
