% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastinference.R
\name{summary_plot}
\alias{summary_plot}
\title{Summary_plot}
\usage{
summary_plot(object, dpi = 200, ...)
}
\arguments{
\item{object}{ShapInterpretation object}

\item{dpi}{dots per inch}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Displays the SHAP values (which can be interpreted for feature importance)
}
