% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_model.R
\name{tabular_config}
\alias{tabular_config}
\title{Tabular_config}
\usage{
tabular_config(
  ps = NULL,
  embed_p = 0,
  y_range = NULL,
  use_bn = TRUE,
  bn_final = FALSE,
  bn_cont = TRUE,
  act_cls = nn$ReLU(inplace = TRUE)
)
}
\arguments{
\item{ps}{ps}

\item{embed_p}{embed proportion}

\item{y_range}{y_range}

\item{use_bn}{use batch normalization}

\item{bn_final}{batch normalization final}

\item{bn_cont}{batch normalization}

\item{act_cls}{activation}
}
\value{
None
}
\description{
Convenience function to easily create a config for `TabularModel`
}
