% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{tabular_learner}
\alias{tabular_learner}
\title{Tabular_learner}
\usage{
tabular_learner(
  dls,
  layers = NULL,
  emb_szs = NULL,
  config = NULL,
  n_out = NULL,
  y_range = NULL,
  loss_func = NULL,
  opt_func = Adam(),
  lr = 0.001,
  splitter = trainable_params(),
  cbs = NULL,
  metrics = NULL,
  path = NULL,
  model_dir = "models",
  wd = NULL,
  wd_bn_bias = FALSE,
  train_bn = TRUE,
  moms = list(0.95, 0.85, 0.95)
)
}
\arguments{
\item{dls}{dls}

\item{layers}{layers}

\item{emb_szs}{emb_szs}

\item{config}{config}

\item{n_out}{n_out}

\item{y_range}{y_range}

\item{loss_func}{loss_func}

\item{opt_func}{opt_func}

\item{lr}{lr}

\item{splitter}{splitter}

\item{cbs}{cbs}

\item{metrics}{metrics}

\item{path}{path}

\item{model_dir}{model_dir}

\item{wd}{wd}

\item{wd_bn_bias}{wd_bn_bias}

\item{train_bn}{train_bn}

\item{moms}{moms}
}
\description{
Get a `Learner` using `dls`, with `metrics`, including a `TabularModel` created using the remaining params.
}
