% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{tabular_TabularDataBunch_from_df}
\alias{tabular_TabularDataBunch_from_df}
\title{from_df}
\usage{
tabular_TabularDataBunch_from_df(
  path,
  df,
  dep_var,
  valid_idx,
  procs = NULL,
  cat_names = NULL,
  cont_names = NULL,
  classes = NULL,
  test_df = NULL,
  bs = 64,
  val_bs = NULL,
  num_workers = 6L,
  dl_tfms = NULL,
  device = NULL,
  collate_fn = data_collate(),
  no_check = FALSE
)
}
\arguments{
\item{path}{path}

\item{df}{df}

\item{dep_var}{dep_var}

\item{valid_idx}{valid_idx}

\item{procs}{procs}

\item{cat_names}{cat_names}

\item{cont_names}{cont_names}

\item{classes}{classes}

\item{test_df}{test_df}

\item{bs}{bs}

\item{val_bs}{val_bs}

\item{num_workers}{num_workers}

\item{dl_tfms}{dl_tfms}

\item{device}{device}

\item{collate_fn}{collate_fn}

\item{no_check}{no_check}
}
\description{
Create a `DataBunch` from `df` and `valid_idx` with `dep_var`. `kwargs` are passed to `DataBunch.create`.
}
