% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{tabular_TabularList_from_df}
\alias{tabular_TabularList_from_df}
\title{Tabular List from dataframe}
\usage{
tabular_TabularList_from_df(
  df,
  cat_names = NULL,
  cont_names = NULL,
  procs = NULL
)
}
\arguments{
\item{df}{df}

\item{cat_names}{cat_names}

\item{cont_names}{cont_names}

\item{procs}{procs}

\item{cls}{cls}
}
\description{
Get the list of inputs in the `col` of `path/csv_name`.
}
