% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{tabular_TabularList_get_emb_szs}
\alias{tabular_TabularList_get_emb_szs}
\title{Get_emb_szs}
\usage{
tabular_TabularList_get_emb_szs(sz_dict = NULL)
}
\arguments{
\item{sz_dict}{sz_dict}
}
\description{
Return the default embedding sizes suitable for this data or takes the ones in `sz_dict`.
}
\details{

}
