% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_transform.R
\name{tabular_TabularList_show_xys}
\alias{tabular_TabularList_show_xys}
\title{Show_xys}
\usage{
tabular_TabularList_show_xys(xs, ys)
}
\arguments{
\item{xs}{xs}

\item{ys}{ys}
}
\description{
Show the `xs` (inputs) and `ys` (targets).
}
\details{

}
