% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_transform.R
\name{tabular_TabularList_show_xyzs}
\alias{tabular_TabularList_show_xyzs}
\title{show_xyzs}
\usage{
tabular_TabularList_show_xyzs(xs, ys, zs)
}
\arguments{
\item{xs}{xs}

\item{ys}{ys}

\item{zs}{zs}
}
\description{
Show `xs` (inputs), `ys` (targets) and `zs` (predictions).
}
\details{

}
