% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{TabularDataBunch}
\alias{TabularDataBunch}
\title{TabularDataBunch}
\usage{
TabularDataBunch(
  train_dl,
  valid_dl,
  fix_dl = NULL,
  test_dl = NULL,
  device = NULL,
  dl_tfms = NULL,
  path = getwd(),
  collate_fn = data_collate(),
  no_check = FALSE
)
}
\arguments{
\item{train_dl}{train_dl}

\item{valid_dl}{valid_dl}

\item{fix_dl}{fix_dl}

\item{test_dl}{test_dl}

\item{device}{device}

\item{dl_tfms}{dl_tfms}

\item{path}{path}

\item{collate_fn}{collate_fn}

\item{no_check}{no_check}
}
\description{
Create a `DataBunch` suitable for tabular data.
}
