% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{TabularDataTable}
\alias{TabularDataTable}
\title{TabularDataTable}
\usage{
TabularDataTable(
  df,
  procs = NULL,
  cat_names = NULL,
  cont_names = NULL,
  y_names = NULL,
  y_block = NULL,
  splits = NULL,
  do_setup = TRUE,
  device = NULL,
  inplace = FALSE,
  reduce_memory = TRUE
)
}
\arguments{
\item{df}{A DataFrame of your data}

\item{procs}{procs}

\item{cat_names}{the names of the categorical variables}

\item{cont_names}{the names of the continuous variables}

\item{y_names}{the names of the dependent variables}

\item{y_block}{the TransformBlock to use for the target}

\item{splits}{How to split your data}

\item{do_setup}{A parameter for if Tabular will run the data through the procs upon initialization}

\item{device}{cuda or cpu}

\item{inplace}{If True, Tabular will not keep a separate copy of your original DataFrame in memory}

\item{reduce_memory}{fastai will attempt to reduce the overall memory usage}
}
\description{
A `Tabular` object with transforms
}
