% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_model.R
\name{TabularModel}
\alias{TabularModel}
\title{TabularModel}
\usage{
TabularModel(
  emb_szs,
  n_cont,
  out_sz,
  layers,
  ps = NULL,
  embed_p = 0,
  y_range = NULL,
  use_bn = TRUE,
  bn_final = FALSE,
  bn_cont = TRUE,
  act_cls = nn$ReLU(inplace = TRUE)
)
}
\arguments{
\item{emb_szs}{emb_szs}

\item{n_cont}{n_cont}

\item{out_sz}{out_sz}

\item{layers}{layers}

\item{ps}{ps}

\item{embed_p}{embed_p}

\item{y_range}{y_range}

\item{use_bn}{use_bn}

\item{bn_final}{bn_final}

\item{bn_cont}{bn_cont}

\item{act_cls}{act_cls}
}
\value{
None
}
\description{
Basic model for tabular data.
}
