% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_tabular.R
\name{TabularTSDataloader}
\alias{TabularTSDataloader}
\title{TabularTSDataloader}
\usage{
TabularTSDataloader(
  dataset,
  bs = 16,
  shuffle = FALSE,
  after_batch = NULL,
  num_workers = 0,
  verbose = FALSE,
  do_setup = TRUE,
  pin_memory = FALSE,
  timeout = 0,
  batch_size = NULL,
  drop_last = FALSE,
  indexed = NULL,
  n = NULL,
  device = NULL
)
}
\arguments{
\item{dataset}{data set}

\item{bs}{batch size}

\item{shuffle}{shuffle or not}

\item{after_batch}{after batch}

\item{num_workers}{the number of workers}

\item{verbose}{verbose}

\item{do_setup}{A parameter for if Tabular will run the data through the procs upon initialization}

\item{pin_memory}{pin memory or not}

\item{timeout}{timeout}

\item{batch_size}{batch size}

\item{drop_last}{drop last}

\item{indexed}{indexed}

\item{n}{n}

\item{device}{device name}
}
\value{
None
}
\description{
Transformed `DataLoader`
}
