% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_signal.R
\name{tar_extract_at_filename}
\alias{tar_extract_at_filename}
\title{Tar_extract_at_filename}
\usage{
tar_extract_at_filename(fname, dest)
}
\arguments{
\item{fname}{folder name}

\item{dest}{destination}
}
\value{
None
}
\description{
Extract `fname` to `dest`/`fname.name` folder using `tarfile`
}
