% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{tensor}
\alias{tensor}
\title{tensor}
\usage{
tensor(...)
}
\arguments{
\item{...}{image}
}
\description{
Like `torch.as_tensor`, but handle lists too, and can pass multiple vector elements directly.
}
