% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_image.R
\name{==.fastai.torch_core.TensorImage}
\alias{==.fastai.torch_core.TensorImage}
\title{Equal}
\usage{
\method{==}{fastai.torch_core.TensorImage}(a, b)
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}
}
\value{
tensor
}
\description{
Equal
}
