% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{TensorBBox_create}
\alias{TensorBBox_create}
\title{TensorBBox_create}
\usage{
TensorBBox_create(x, img_size = NULL)
}
\arguments{
\item{x}{tensor}

\item{img_size}{image size}
}
\value{
None
}
\description{
TensorBBox_create
}
