% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{TensorPoint_create}
\alias{TensorPoint_create}
\title{TensorPoint_create}
\usage{
TensorPoint_create(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
Delegates (`__call__`,`decode`,`setup`) to (`encodes`,`decodes`,`setups`) if `split_idx` matches
}
