% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_learner.R
\name{text_classifier_learner}
\alias{text_classifier_learner}
\title{text_classifier_learner}
\usage{
text_classifier_learner(
  dls,
  arch,
  seq_len = 72,
  config = NULL,
  backwards = FALSE,
  pretrained = TRUE,
  drop_mult = 0.5,
  n_out = NULL,
  lin_ftrs = NULL,
  ps = NULL,
  max_len = 1440,
  y_range = NULL,
  loss_func = NULL,
  opt_func = Adam,
  lr = 0.001,
  splitter = trainable_params,
  cbs = NULL,
  metrics = NULL,
  path = NULL,
  model_dir = "models",
  wd = NULL,
  wd_bn_bias = FALSE,
  train_bn = TRUE,
  moms = list(0.95, 0.85, 0.95)
)
}
\arguments{
\item{dls}{dls}

\item{arch}{arch}

\item{seq_len}{seq_len}

\item{config}{config}

\item{backwards}{backwards}

\item{pretrained}{pretrained}

\item{drop_mult}{drop_mult}

\item{n_out}{n_out}

\item{lin_ftrs}{lin_ftrs}

\item{ps}{ps}

\item{max_len}{max_len}

\item{y_range}{y_range}

\item{loss_func}{loss_func}

\item{opt_func}{opt_func}

\item{lr}{lr}

\item{splitter}{splitter}

\item{cbs}{cbs}

\item{metrics}{metrics}

\item{path}{path}

\item{model_dir}{model_dir}

\item{wd}{wd}

\item{wd_bn_bias}{wd_bn_bias}

\item{train_bn}{train_bn}

\item{moms}{moms}
}
\description{
Create a `Learner` with a text classifier from `dls` and `arch`.
}
