% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{TextBlock}
\alias{TextBlock}
\title{TextBlock}
\usage{
TextBlock(
  tok_tfm,
  vocab = NULL,
  is_lm = FALSE,
  seq_len = 72,
  backwards = FALSE,
  min_freq = 3,
  max_vocab = 60000,
  special_toks = NULL,
  pad_tok = NULL
)
}
\arguments{
\item{tok_tfm}{tok_tfm}

\item{vocab}{vocab}

\item{is_lm}{is_lm}

\item{seq_len}{seq_len}

\item{backwards}{backwards}

\item{min_freq}{min_freq}

\item{max_vocab}{max_vocab}

\item{special_toks}{special_toks}

\item{pad_tok}{pad_tok}
}
\value{
block object
}
\description{
A `TransformBlock` for texts
}
