% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{TextBlock_from_df}
\alias{TextBlock_from_df}
\title{from_df}
\usage{
TextBlock_from_df(
  text_cols,
  vocab = NULL,
  is_lm = FALSE,
  seq_len = 72,
  backwards = FALSE,
  min_freq = 3,
  max_vocab = 60000,
  tok = NULL,
  rules = NULL,
  sep = " ",
  n_workers = 6,
  mark_fields = NULL,
  res_col_name = "text"
)
}
\arguments{
\item{text_cols}{text_cols}

\item{vocab}{vocab}

\item{is_lm}{is_lm}

\item{seq_len}{seq_len}

\item{backwards}{backwards}

\item{min_freq}{min_freq}

\item{max_vocab}{max_vocab}

\item{tok}{tok}

\item{rules}{rules}

\item{sep}{sep}

\item{n_workers}{n_workers}

\item{mark_fields}{mark_fields}

\item{res_col_name}{res_col_name}

\item{cls}{cls}
}
\description{
Build a `TextBlock` from a dataframe using `text_cols`
}
\details{

}
