% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{TextBlock_from_folder}
\alias{TextBlock_from_folder}
\title{from_folder}
\usage{
TextBlock_from_folder(
  path,
  vocab = NULL,
  is_lm = FALSE,
  seq_len = 72,
  backwards = FALSE,
  min_freq = 3,
  max_vocab = 60000,
  tok = NULL,
  rules = NULL,
  extensions = NULL,
  folders = NULL,
  output_dir = NULL,
  skip_if_exists = TRUE,
  output_names = NULL,
  n_workers = 6,
  encoding = "utf8"
)
}
\arguments{
\item{path}{path}

\item{vocab}{vocab}

\item{is_lm}{is_lm}

\item{seq_len}{seq_len}

\item{backwards}{backwards}

\item{min_freq}{min_freq}

\item{max_vocab}{max_vocab}

\item{tok}{tok}

\item{rules}{rules}

\item{extensions}{extensions}

\item{folders}{folders}

\item{output_dir}{output_dir}

\item{skip_if_exists}{skip_if_exists}

\item{output_names}{output_names}

\item{n_workers}{n_workers}

\item{encoding}{encoding}
}
\description{
Build a `TextBlock` from a `path`
}
