% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{TextDataLoaders_from_csv}
\alias{TextDataLoaders_from_csv}
\title{TextDataLoaders_from_csv}
\usage{
TextDataLoaders_from_csv(
  path,
  csv_fname = "labels.csv",
  header = "infer",
  delimiter = NULL,
  valid_pct = 0.2,
  seed = NULL,
  text_col = 0,
  label_col = 1,
  label_delim = NULL,
  y_block = NULL,
  text_vocab = NULL,
  is_lm = FALSE,
  valid_col = NULL,
  tok_tfm = NULL,
  seq_len = 72,
  backwards = FALSE,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  device = NULL
)
}
\arguments{
\item{path}{path}

\item{csv_fname}{csv file name}

\item{header}{header}

\item{delimiter}{delimiter}

\item{valid_pct}{valid_ation percentage}

\item{seed}{random seed}

\item{text_col}{text column}

\item{label_col}{label column}

\item{label_delim}{label separator}

\item{y_block}{y_block}

\item{text_vocab}{text vocabulary}

\item{is_lm}{is_lm}

\item{valid_col}{valid column}

\item{tok_tfm}{tok_tfm}

\item{seq_len}{seq_len}

\item{backwards}{backwards}

\item{bs}{batch size}

\item{val_bs}{validation batch size}

\item{shuffle_train}{shuffle train data}

\item{device}{device}
}
\value{
text loader
}
\description{
Create from `csv` file in `path/csv_fname`
}
