% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{TextDataLoaders_from_folder}
\alias{TextDataLoaders_from_folder}
\title{TextDataLoaders_from_folder}
\usage{
TextDataLoaders_from_folder(
  path,
  train = "train",
  valid = "valid",
  valid_pct = NULL,
  seed = NULL,
  vocab = NULL,
  text_vocab = NULL,
  is_lm = FALSE,
  tok_tfm = NULL,
  seq_len = 72,
  backwards = FALSE,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  device = NULL
)
}
\arguments{
\item{path}{path}

\item{train}{train data}

\item{valid}{validation data}

\item{valid_pct}{validation percentage}

\item{seed}{random seed}

\item{vocab}{vocabulary}

\item{text_vocab}{text_vocab}

\item{is_lm}{is_lm}

\item{tok_tfm}{tok_tfm}

\item{seq_len}{seq_len}

\item{backwards}{backwards}

\item{bs}{batch size}

\item{val_bs}{validation batch size}

\item{shuffle_train}{shuffle train data}

\item{device}{device}
}
\value{
text loader
}
\description{
Create from imagenet style dataset in `path` with `train` and `valid` subfolders (or provide `valid_pct`)
}
