% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_learner.R
\name{TextLearner_load_pretrained}
\alias{TextLearner_load_pretrained}
\title{load_pretrained}
\usage{
TextLearner_load_pretrained(wgts_fname, vocab_fname, model = NULL)
}
\arguments{
\item{wgts_fname}{wgts_fname}

\item{vocab_fname}{vocab_fname}

\item{model}{model}
}
\description{
Load a pretrained model and adapt it to the data vocabulary.
}
