% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_spectrogram.R
\name{TfmResize}
\alias{TfmResize}
\title{TfmResize}
\usage{
TfmResize(size, interp_mode = "bilinear")
}
\arguments{
\item{size}{size}

\item{interp_mode}{interpolation mode}
}
\value{
None
}
\description{
Temporary fix to allow image resizing transform
}
