% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timm_loader.R
\name{timm_learner}
\alias{timm_learner}
\title{Timm_learner}
\usage{
timm_learner(dls, arch, ...)
}
\arguments{
\item{dls}{dataloader}

\item{arch}{model architecture}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Build a convnet style learner from `dls` and `arch` using the `timm` library
}
