% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{to_bytes_format}
\alias{to_bytes_format}
\title{To_bytes_format}
\usage{
to_bytes_format(img, format = "png")
}
\arguments{
\item{img}{image}

\item{format}{format}
}
\value{
None
}
\description{
Convert to bytes, default to PNG format
}
