% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{to_thumb}
\alias{to_thumb}
\title{To_thumb}
\usage{
to_thumb(img, h, w = NULL)
}
\arguments{
\item{img}{image}

\item{h}{height}

\item{w}{width}
}
\value{
None
}
\description{
Same as `thumbnail`, but uses a copy
}
