% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_core.R
\name{tokenize1}
\alias{tokenize1}
\title{tokenize1}
\usage{
tokenize1(text, tok, rules = NULL, post_rules = NULL)
}
\arguments{
\item{text}{text}

\item{tok}{tok}

\item{rules}{rules}

\item{post_rules}{post_rules}
}
\description{
Call `TokenizeWithRules` with a single text
}
