% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_core.R
\name{tokenize_csv}
\alias{tokenize_csv}
\title{tokenize_csv}
\usage{
tokenize_csv(
  fname,
  text_cols,
  outname = NULL,
  n_workers = 4,
  rules = NULL,
  mark_fields = NULL,
  tok = NULL,
  header = "infer",
  chunksize = 50000
)
}
\arguments{
\item{fname}{fname}

\item{text_cols}{text_cols}

\item{outname}{outname}

\item{n_workers}{n_workers}

\item{rules}{rules}

\item{mark_fields}{mark_fields}

\item{tok}{tok}

\item{header}{header}

\item{chunksize}{chunksize}
}
\description{
Tokenize texts in the `text_cols` of the csv `fname` in parallel using `n_workers`
}
