% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_core.R
\name{tokenize_df}
\alias{tokenize_df}
\title{Tokenize_df}
\usage{
tokenize_df(
  df,
  text_cols,
  n_workers = 6,
  rules = NULL,
  mark_fields = NULL,
  tok = NULL,
  tok_text_col = "text"
)
}
\arguments{
\item{df}{data frame}

\item{text_cols}{text columns}

\item{n_workers}{number of workers}

\item{rules}{rules}

\item{mark_fields}{mark_fields}

\item{tok}{tokenizer}

\item{tok_text_col}{tok_text_col}
}
\value{
None
}
\description{
Tokenize texts in `df[text_cols]` in parallel using `n_workers`
}
