% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_core.R
\name{tokenize_folder}
\alias{tokenize_folder}
\title{tokenize_folder}
\usage{
tokenize_folder(
  path,
  extensions = NULL,
  folders = NULL,
  output_dir = NULL,
  skip_if_exists = TRUE,
  output_names = NULL,
  n_workers = 6,
  rules = NULL,
  tok = NULL,
  encoding = "utf8"
)
}
\arguments{
\item{path}{path}

\item{extensions}{extensions}

\item{folders}{folders}

\item{output_dir}{output_dir}

\item{skip_if_exists}{skip_if_exists}

\item{output_names}{output_names}

\item{n_workers}{n_workers}

\item{rules}{rules}

\item{tok}{tok}

\item{encoding}{encoding}
}
\description{
Tokenize text files in `path` in parallel using `n_workers`
}
