% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_core.R
\name{tokenize_texts}
\alias{tokenize_texts}
\title{tokenize_texts}
\usage{
tokenize_texts(texts, n_workers = 6, rules = NULL, tok = NULL)
}
\arguments{
\item{texts}{texts}

\item{n_workers}{n_workers}

\item{rules}{rules}

\item{tok}{tok}
}
\description{
Tokenize `texts` in parallel using `n_workers`
}
