% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_core.R
\name{TokenizeWithRules}
\alias{TokenizeWithRules}
\title{TokenizeWithRules}
\usage{
TokenizeWithRules(tok, rules = NULL, post_rules = NULL)
}
\arguments{
\item{tok}{tok}

\item{rules}{rules}

\item{post_rules}{post_rules}
}
\description{
A wrapper around `tok` which applies `rules`, then tokenizes, then applies `post_rules`
}
