% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks_hook.R
\name{total_params}
\alias{total_params}
\title{Total_params}
\usage{
total_params(m)
}
\arguments{
\item{m}{m parameter}
}
\value{
None
}
\description{
Give the number of parameters of a module and if it's trainable or not
}
