% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{ToTensor}
\alias{ToTensor}
\title{ToTensor}
\usage{
ToTensor(enc = NULL, dec = NULL, split_idx = NULL, order = NULL)
}
\arguments{
\item{enc}{encoder}

\item{dec}{decoder}

\item{split_idx}{int, split by index}

\item{order}{order}
}
\value{
None
}
\description{
Convert item to appropriate tensor class
}
