% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab.R, R/tabular.R
\name{trainable_params}
\alias{trainable_params}
\title{trainable_params}
\usage{
trainable_params(m)

trainable_params(m)
}
\arguments{
\item{m}{m}
}
\description{
Return all trainable parameters of `m`

Return all trainable parameters of `m`
}
