% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{Transform}
\alias{Transform}
\title{Transform}
\usage{
Transform(enc = NULL, dec = NULL, split_idx = NULL, order = NULL)
}
\arguments{
\item{enc}{enc}

\item{dec}{dec}

\item{split_idx}{split_idx}

\item{order}{order}
}
\value{
None
}
\description{
Delegates (`__call__`,`decode`,`setup`) to (`encodes`,`decodes`,`setups`) if `split_idx` matches
}
