% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{TransformBlock}
\alias{TransformBlock}
\title{TransformBlock}
\usage{
TransformBlock(
  type_tfms = NULL,
  item_tfms = NULL,
  batch_tfms = NULL,
  dl_type = NULL,
  dls_kwargs = NULL
)
}
\arguments{
\item{type_tfms}{transformation type}

\item{item_tfms}{item transformation type}

\item{batch_tfms}{one or several transforms applied to the batches once they are formed}

\item{dl_type}{DL application}

\item{dls_kwargs}{additional arguments}
}
\value{
block
}
\description{
A basic wrapper that links defaults transforms for the data block API
}
