% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_modules.R
\name{TransformersTokenizer}
\alias{TransformersTokenizer}
\title{TransformersTokenizer}
\usage{
TransformersTokenizer(tokenizer)
}
\arguments{
\item{tokenizer}{tokenizer object}
}
\value{
None
}
\description{
TransformersTokenizer
}
