% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{trunc_normal_}
\alias{trunc_normal_}
\title{Trunc_normal_}
\usage{
trunc_normal_(x, mean = 0, std = 1)
}
\arguments{
\item{x}{tensor}

\item{mean}{mean}

\item{std}{standard deviation}
}
\value{
tensor
}
\description{
Truncated normal initialization (approximation)
}
