% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet.R
\name{unet_config}
\alias{unet_config}
\title{Unet_config}
\usage{
unet_config(
  blur = FALSE,
  blur_final = TRUE,
  self_attention = FALSE,
  y_range = NULL,
  last_cross = TRUE,
  bottle = FALSE,
  act_cls = nn()$ReLU,
  init = nn()$init$kaiming_normal_,
  norm_type = NULL
)
}
\arguments{
\item{blur}{blur is used to avoid checkerboard artifacts at each layer.}

\item{blur_final}{blur final is specific to the last layer.}

\item{self_attention}{self_attention determines if we use a self attention layer at the third block before the end.}

\item{y_range}{If y_range is passed, the last activations go through a sigmoid rescaled to that range.}

\item{last_cross}{last cros}

\item{bottle}{bottle}

\item{act_cls}{activation}

\item{init}{initializer}

\item{norm_type}{normalization type}
}
\value{
None
}
\description{
Convenience function to easily create a config for `DynamicUnet`
}
