% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet.R
\name{unet_learner}
\alias{unet_learner}
\title{Unet_learner}
\usage{
unet_learner(dls, arch, ...)
}
\arguments{
\item{dls}{dataloader}

\item{arch}{architecture}
}
\value{
None
}
\description{
Build a unet learner from `dls` and `arch`
}
