% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeze.R
\name{unfreeze}
\alias{unfreeze}
\title{Unfreeze a model}
\usage{
unfreeze(object)
}
\arguments{
\item{object}{A model}
}
\value{
Nothing
}
\description{
Unfreeze a model
}
\examples{

\dontrun{
learnR \%>\% unfreeze()
}
}
