% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{unique.fastai.torch_core.TensorMask}
\alias{unique.fastai.torch_core.TensorMask}
\title{Unique of tensor}
\usage{
\method{unique}{fastai.torch_core.TensorMask}(a, ...)
}
\description{
Unique of tensor
}
