% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_ADULT_SAMPLE}
\alias{URLs_ADULT_SAMPLE}
\title{ADULT_SAMPLE dataset}
\usage{
URLs_ADULT_SAMPLE(filename = "ADULT_SAMPLE", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download ADULT_SAMPLE dataset
}
