% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_AG_NEWS}
\alias{URLs_AG_NEWS}
\title{AG_NEWS dataset}
\usage{
URLs_AG_NEWS(filename = "AG_NEWS", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download AG_NEWS dataset
}
