% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_AMAZON_REVIEWSAMAZON_REVIEWS}
\alias{URLs_AMAZON_REVIEWSAMAZON_REVIEWS}
\title{AMAZON_REVIEWSAMAZON_REVIEWS dataset}
\usage{
URLs_AMAZON_REVIEWSAMAZON_REVIEWS(
  filename = AMAZON_REVIEWSAMAZON_REVIEWS,
  untar = TRUE
)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download AMAZON_REVIEWSAMAZON_REVIEWS dataset
}
