% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_BIWI_HEAD_POSE}
\alias{URLs_BIWI_HEAD_POSE}
\title{BIWI_HEAD_POSE dataset}
\usage{
URLs_BIWI_HEAD_POSE(filename = "BIWI_HEAD_POSE", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download BIWI_HEAD_POSE dataset
}
