% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_CAMVID}
\alias{URLs_CAMVID}
\title{CAMVID dataset}
\usage{
URLs_CAMVID(filename = "CAMVID", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download CAMVID dataset
}
