% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_CAMVID_TINY}
\alias{URLs_CAMVID_TINY}
\title{CAMVID_TINY dataset}
\usage{
URLs_CAMVID_TINY(filename = "CAMVID_TINY", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download CAMVID_TINY dataset
}
