% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_CARS}
\alias{URLs_CARS}
\title{CARS dataset}
\usage{
URLs_CARS(filename = "CARS", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download CARS dataset
}
