% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_CIFAR}
\alias{URLs_CIFAR}
\title{CIFAR dataset}
\usage{
URLs_CIFAR(filename = "CIFAR", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download CIFAR dataset
}
