% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_COCO_TINY}
\alias{URLs_COCO_TINY}
\title{COCO_TINY dataset}
\usage{
URLs_COCO_TINY(filename = "COCO_TINY", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download COCO_TINY dataset
}
