% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_CUB_200_2011}
\alias{URLs_CUB_200_2011}
\title{CUB_200_2011 dataset}
\usage{
URLs_CUB_200_2011(filename = CUB_200_2011, untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download CUB_200_2011 dataset
}
