% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_DBPEDIA}
\alias{URLs_DBPEDIA}
\title{DBPEDIA dataset}
\usage{
URLs_DBPEDIA(filename = "DBPEDIA", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download DBPEDIA dataset
}
