% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_DOGS}
\alias{URLs_DOGS}
\title{DOGS dataset}
\usage{
URLs_DOGS(filename = "DOGS", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download DOGS dataset
}
