% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_FOOD}
\alias{URLs_FOOD}
\title{FOOD dataset}
\usage{
URLs_FOOD(filename = "FOOD", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download FOOD dataset
}
