% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleGAN_models.R
\name{URLs_HORSE_2_ZEBRA}
\alias{URLs_HORSE_2_ZEBRA}
\title{HORSE_2_ZEBRA dataset}
\usage{
URLs_HORSE_2_ZEBRA(filename = "horse2zebra", unzip = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{unzip}{logical, whether to unzip the '.zip' file}
}
\value{
None
}
\description{
download HORSE_2_ZEBRA dataset
}
