% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_HUMAN_NUMBERS}
\alias{URLs_HUMAN_NUMBERS}
\title{HUMAN_NUMBERS dataset}
\usage{
URLs_HUMAN_NUMBERS(filename = "HUMAN_NUMBERS", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download HUMAN_NUMBERS dataset
}
