% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_IMAGENETTE}
\alias{URLs_IMAGENETTE}
\title{IMAGENETTE dataset}
\usage{
URLs_IMAGENETTE(filename = IMAGENETTE, untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download IMAGENETTE dataset
}
