% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_IMAGENETTE_160}
\alias{URLs_IMAGENETTE_160}
\title{IMAGENETTE_160 dataset}
\usage{
URLs_IMAGENETTE_160(filename = "IMAGENETTE_160", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download IMAGENETTE_160 dataset
}
