% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_IMAGEWOOF_160}
\alias{URLs_IMAGEWOOF_160}
\title{IMAGEWOOF_160 dataset}
\usage{
URLs_IMAGEWOOF_160(filename = "IMAGEWOOF_160", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download IMAGEWOOF_160 dataset
}
