% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_IMAGEWOOF_320}
\alias{URLs_IMAGEWOOF_320}
\title{IMAGEWOOF_320 dataset}
\usage{
URLs_IMAGEWOOF_320(filename = IMAGEWOOF_320, untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download IMAGEWOOF_320 dataset
}
