% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_LSUN_BEDROOMS}
\alias{URLs_LSUN_BEDROOMS}
\title{LSUN_BEDROOMS dataset}
\usage{
URLs_LSUN_BEDROOMS(filename = LSUN_BEDROOMS, untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download LSUN_BEDROOMS dataset
}
