% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_ML_SAMPLE}
\alias{URLs_ML_SAMPLE}
\title{ML_SAMPLE dataset}
\usage{
URLs_ML_SAMPLE(filename = "ML_SAMPLE", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download ML_SAMPLE dataset
}
