% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_MNIST}
\alias{URLs_MNIST}
\title{MNIST dataset}
\usage{
URLs_MNIST(filename = "MNIST", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download MNIST dataset
}
