% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_MNIST_SAMPLE}
\alias{URLs_MNIST_SAMPLE}
\title{MNIST_SAMPLE dataset}
\usage{
URLs_MNIST_SAMPLE(filename = MNIST_SAMPLE, untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download MNIST_SAMPLE dataset
}
