% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_MNIST_VAR_SIZE_TINY}
\alias{URLs_MNIST_VAR_SIZE_TINY}
\title{MNIST_VAR_SIZE_TINY dataset}
\usage{
URLs_MNIST_VAR_SIZE_TINY(filename = "MNIST_VAR_SIZE_TINY", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download MNIST_VAR_SIZE_TINY dataset
}
