% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_MOVIE_LENS_ML_100k}
\alias{URLs_MOVIE_LENS_ML_100k}
\title{MOVIE_LENS_ML_100k dataset}
\usage{
URLs_MOVIE_LENS_ML_100k(filename = "ml-100k", unzip = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download YELP_REVIEWS_POLARITY dataset
}
