% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_MT_ENG_FRA}
\alias{URLs_MT_ENG_FRA}
\title{MT_ENG_FRA dataset}
\usage{
URLs_MT_ENG_FRA(filename = MT_ENG_FRA, untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download MT_ENG_FRA dataset
}
