% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_OPENAI_TRANSFORMER}
\alias{URLs_OPENAI_TRANSFORMER}
\title{OPENAI_TRANSFORMER dataset}
\usage{
URLs_OPENAI_TRANSFORMER(filename = OPENAI_TRANSFORMER, untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download OPENAI_TRANSFORMER dataset
}
