% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_PASCAL_2012}
\alias{URLs_PASCAL_2012}
\title{PASCAL_2012 dataset}
\usage{
URLs_PASCAL_2012(filename = "PASCAL_2012", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download PASCAL_2012 dataset
}
