% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_PETS}
\alias{URLs_PETS}
\title{PETS dataset}
\usage{
URLs_PETS(filename = PETS, untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download PETS dataset
}
