% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_PLANET_SAMPLE}
\alias{URLs_PLANET_SAMPLE}
\title{PLANET_SAMPLE dataset}
\usage{
URLs_PLANET_SAMPLE(filename = "PLANET_SAMPLE", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download PLANET_SAMPLE dataset
}
