% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_PLANET_TINY}
\alias{URLs_PLANET_TINY}
\title{PLANET_TINY dataset}
\usage{
URLs_PLANET_TINY(filename = "PLANET_TINY", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download PLANET_TINY dataset
}
