% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_S3_IMAGE}
\alias{URLs_S3_IMAGE}
\title{S3_IMAGE dataset}
\usage{
URLs_S3_IMAGE(filename = S3_IMAGE, untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download S3_IMAGE dataset
}
