% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_S3_IMAGELOC}
\alias{URLs_S3_IMAGELOC}
\title{S3_IMAGELOC dataset}
\usage{
URLs_S3_IMAGELOC(filename = "S3_IMAGELOC", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download S3_IMAGELOC dataset
}
