% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_S3_MODEL}
\alias{URLs_S3_MODEL}
\title{S3_MODEL dataset}
\usage{
URLs_S3_MODEL(filename = "S3_MODEL", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download S3_MODEL dataset
}
