% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_S3_NLP}
\alias{URLs_S3_NLP}
\title{S3_NLP dataset}
\usage{
URLs_S3_NLP(filename = "S3_NLP", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download S3_NLP dataset
}
