% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_SIIM_SMALL}
\alias{URLs_SIIM_SMALL}
\title{SIIM_SMALL}
\usage{
URLs_SIIM_SMALL(filename = "SIIM_SMALL", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download YELP_REVIEWS_POLARITY dataset
}
