% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_SKIN_LESION}
\alias{URLs_SKIN_LESION}
\title{SKIN_LESION dataset}
\usage{
URLs_SKIN_LESION(filename = SKIN_LESION, untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download SKIN_LESION dataset
}
