% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_SOGOU_NEWS}
\alias{URLs_SOGOU_NEWS}
\title{SOGOU_NEWS dataset}
\usage{
URLs_SOGOU_NEWS(filename = "SOGOU_NEWS", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download SOGOU_NEWS dataset
}
