% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_download_tar.R
\name{URLs_SPEAKERS10}
\alias{URLs_SPEAKERS10}
\title{SPEAKERS10 dataset}
\usage{
URLs_SPEAKERS10(filename = "SPEAKERS10", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download SPEAKERS10 dataset
}
\examples{
\dontrun{

URLs_SPEAKERS10()

}

}
