% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_download_tar.R
\name{URLs_SPEECHCOMMANDS}
\alias{URLs_SPEECHCOMMANDS}
\title{SPEECHCOMMANDS dataset}
\usage{
URLs_SPEECHCOMMANDS(filename = "SPEECHCOMMANDS", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download SPEECHCOMMANDS dataset
}
\examples{
\dontrun{

URLs_SPEECHCOMMANDS()

}

}
