% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_WIKITEXT}
\alias{URLs_WIKITEXT}
\title{WIKITEXT dataset}
\usage{
URLs_WIKITEXT(filename = "WIKITEXT", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download WIKITEXT dataset
}
