% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_WIKITEXT_TINY}
\alias{URLs_WIKITEXT_TINY}
\title{WIKITEXT_TINY dataset}
\usage{
URLs_WIKITEXT_TINY(filename = WIKITEXT_TINY, untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download WIKITEXT_TINY dataset
}
