% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_WT103_BWD}
\alias{URLs_WT103_BWD}
\title{WT103_BWD dataset}
\usage{
URLs_WT103_BWD(filename = "WT103_BWD", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download WT103_BWD dataset
}
