% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_WT103_FWD}
\alias{URLs_WT103_FWD}
\title{WT103_FWD dataset}
\usage{
URLs_WT103_FWD(filename = WT103_FWD, untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download WT103_FWD dataset
}
