% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_YELP_REVIEWS}
\alias{URLs_YELP_REVIEWS}
\title{YELP_REVIEWS dataset}
\usage{
URLs_YELP_REVIEWS(filename = "YELP_REVIEWS", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\value{
None
}
\description{
download YELP_REVIEWS dataset
}
