% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_YELP_REVIEWS_POLARITY}
\alias{URLs_YELP_REVIEWS_POLARITY}
\title{YELP_REVIEWS_POLARITY dataset}
\usage{
URLs_YELP_REVIEWS_POLARITY(filename = "YELP_REVIEWS_POLARITY", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download YELP_REVIEWS_POLARITY dataset
}
