% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{vleaky_relu}
\alias{vleaky_relu}
\title{Vleaky_relu}
\usage{
vleaky_relu(input, inplace = TRUE)
}
\arguments{
\item{input}{input}

\item{inplace}{inplace}
}
\value{
None
}
\description{
`F$leaky_relu` with 0.3 slope
}
