% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_config.R
\name{Voice}
\alias{Voice}
\title{Voice}
\usage{
Voice(
  sample_rate = 16000,
  n_fft = 1024,
  win_length = NULL,
  hop_length = 128,
  f_min = 50,
  f_max = 8000,
  pad = 0,
  n_mels = 128,
  window_fn = torch$hann_window,
  power = 2,
  normalized = FALSE,
  wkwargs = NULL,
  mel = TRUE,
  to_db = FALSE
)
}
\arguments{
\item{sample_rate}{sample_rate}

\item{n_fft}{n_fft}

\item{win_length}{win_length}

\item{hop_length}{hop_length}

\item{f_min}{f_min}

\item{f_max}{f_max}

\item{pad}{pad}

\item{n_mels}{n_mels}

\item{window_fn}{window_fn}

\item{power}{power}

\item{normalized}{normalized}

\item{wkwargs}{wkwargs}

\item{mel}{mel}

\item{to_db}{to_db}
}
\value{
None
}
\description{
Voice
}
