% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{WandbCallback}
\alias{WandbCallback}
\title{WandbCallback}
\usage{
WandbCallback(
  log = "gradients",
  log_preds = TRUE,
  log_model = TRUE,
  log_dataset = FALSE,
  dataset_name = NULL,
  valid_dl = NULL,
  n_preds = 36,
  seed = 12345,
  reorder = TRUE
)
}
\arguments{
\item{log}{"gradients" (default), "parameters", "all" or None. Losses & metrics are always logged.}

\item{log_preds}{whether we want to log prediction samples (default to True).}

\item{log_model}{whether we want to log our model (default to True). This also requires SaveModelCallback.}

\item{log_dataset}{Options:
- False (default)
- True will log folder referenced by learn.dls.path.
- a path can be defined explicitly to reference which folder to log.
Note: subfolder "models" is always ignored.}

\item{dataset_name}{name of logged dataset (default to folder name).}

\item{valid_dl}{DataLoaders containing items used for prediction samples (default to random items from learn.dls.valid.}

\item{n_preds}{number of logged predictions (default to 36).}

\item{seed}{used for defining random samples.}

\item{reorder}{reorder or not}
}
\value{
None
}
\description{
Saves model topology, losses & metrics
}
