% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{Warp}
\alias{Warp}
\title{Warp}
\usage{
Warp(
  magnitude = 0.2,
  p = 0.5,
  draw_x = NULL,
  draw_y = NULL,
  size = NULL,
  mode = "bilinear",
  pad_mode = "reflection",
  batch = FALSE,
  align_corners = TRUE
)
}
\arguments{
\item{magnitude}{magnitude}

\item{p}{p}

\item{draw_x}{draw_x}

\item{draw_y}{draw_y}

\item{size}{size}

\item{mode}{mode}

\item{pad_mode}{pad_mode}

\item{batch}{batch}

\item{align_corners}{align_corners}
}
\description{
Apply perspective warping with `magnitude` and `p` on a batch of matrices
}
