% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastinference.R
\name{waterfall_plot}
\alias{waterfall_plot}
\title{Waterfall_plot}
\usage{
waterfall_plot(object, row_idx = NULL, class_id = 0, dpi = 200, ...)
}
\arguments{
\item{object}{ShapInterpretation object}

\item{row_idx}{is the index of the row chosen in test_data to be analyzed, which defaults to zero.}

\item{class_id}{Accepts a class_id which is used to indicate the class of interest for a
classification model. It can either be an int or str representation for a class of choice.}

\item{dpi}{dots per inch}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Plots an explanation of a single prediction as a waterfall plot. Accepts a row_index and class_id.
}
