% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{xception}
\alias{xception}
\title{xception}
\usage{
xception(c, k = 8, n_middle = 8)
}
\arguments{
\item{c}{c}

\item{k}{k}

\item{n_middle}{n_middle}
}
\description{
Preview version of Xception network. Not tested yet - use at own risk. No pretrained model yet.
}
\details{

}
