% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{xresnet18_deeper}
\alias{xresnet18_deeper}
\title{Xresnet18_deeper}
\usage{
xresnet18_deeper(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
model
}
\description{
Load model architecture
}
