% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{xresnet34_deeper}
\alias{xresnet34_deeper}
\title{Xresnet34_deeper}
\usage{
xresnet34_deeper(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
model
}
\description{
Load model architecture
}
