% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{xresnet50_deeper}
\alias{xresnet50_deeper}
\title{Xresnet50_deeper}
\usage{
xresnet50_deeper(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
model
}
\description{
Load model architecture
}
