% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{xse_resnext50_deeper}
\alias{xse_resnext50_deeper}
\title{xse_resnext50_deeper}
\usage{
xse_resnext50_deeper(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
model
}
\description{
Load model architecture
}
