% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{Zoom_}
\alias{Zoom_}
\title{Zoom}
\usage{
Zoom_(
  min_zoom = 1,
  max_zoom = 1.1,
  p = 0.5,
  draw = NULL,
  draw_x = NULL,
  draw_y = NULL,
  size = NULL,
  mode = "bilinear",
  pad_mode = "reflection",
  batch = FALSE,
  align_corners = TRUE
)
}
\arguments{
\item{min_zoom}{minimum zoom}

\item{max_zoom}{maximum zoom}

\item{p}{probability}

\item{draw}{draw}

\item{draw_x}{draw x}

\item{draw_y}{draw y}

\item{size}{size}

\item{mode}{mode}

\item{pad_mode}{pad mode}

\item{batch}{batch}

\item{align_corners}{align  corners or not}
}
\value{
None
}
\description{
Apply a random zoom of at most `max_zoom` with probability `p` to a batch of images
}
