% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{zoom_mat}
\alias{zoom_mat}
\title{Zoom_mat}
\usage{
zoom_mat(
  x,
  min_zoom = 1,
  max_zoom = 1.1,
  p = 0.5,
  draw = NULL,
  draw_x = NULL,
  draw_y = NULL,
  batch = FALSE
)
}
\arguments{
\item{x}{tensor}

\item{min_zoom}{minimum zoom}

\item{max_zoom}{maximum zoom}

\item{p}{probability}

\item{draw}{draw}

\item{draw_x}{draw x}

\item{draw_y}{draw y}

\item{batch}{batch}
}
\value{
None
}
\description{
Return a random zoom matrix with `max_zoom` and `p`
}
