% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R
\name{abs.fastai.torch_core.TensorMask}
\alias{abs.fastai.torch_core.TensorMask}
\title{Abs}
\usage{
\method{abs}{fastai.torch_core.TensorMask}(x)
}
\arguments{
\item{x}{tensor}
}
\value{
tensor
}
\description{
Abs
}
\examples{
\dontrun{

aa = tensor(-1:-10)
abs(aa)

}

}
