% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{AccumMetric}
\alias{AccumMetric}
\title{AccumMetric}
\usage{
AccumMetric(
  func,
  dim_argmax = NULL,
  activation = "no",
  thresh = NULL,
  to_np = FALSE,
  invert_arg = FALSE,
  flatten = TRUE,
  ...
)
}
\arguments{
\item{func}{function}

\item{dim_argmax}{dimension argmax}

\item{activation}{activation}

\item{thresh}{threshold point}

\item{to_np}{to matrix or not}

\item{invert_arg}{invert arguments}

\item{flatten}{flatten}

\item{...}{additional arguments to pass}
}
\value{
None
}
\description{
Stores predictions and targets on CPU in accumulate to perform final calculations with `func`.
}
