% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy}
\usage{
accuracy(inp, targ, axis = -1)
}
\arguments{
\item{inp}{predictions}

\item{targ}{targets}

\item{axis}{axis}
}
\value{
None
}
\description{
Compute accuracy with `targ` when `pred` is bs * n_classes
}
