% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{accuracy_multi}
\alias{accuracy_multi}
\title{Accuracy_multi}
\usage{
accuracy_multi(inp, targ, thresh = 0.5, sigmoid = TRUE)
}
\arguments{
\item{inp}{predictions}

\item{targ}{targets}

\item{thresh}{threshold point}

\item{sigmoid}{sigmoid}
}
\value{
None
}
\description{
Compute accuracy when `inp` and `targ` are the same size.
}
