% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{accuracy_thresh_expand}
\alias{accuracy_thresh_expand}
\title{accuracy_thresh_expand}
\usage{
accuracy_thresh_expand(y_pred, y_true, thresh = 0.5, sigmoid = TRUE)
}
\arguments{
\item{y_pred}{y_pred}

\item{y_true}{y_true}

\item{thresh}{thresh}

\item{sigmoid}{sigmoid}
}
\description{
Compute accuracy after expanding `y_true` to the size of `y_pred`.
}
