% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loss.R
\name{AdaptiveLoss}
\alias{AdaptiveLoss}
\title{AdaptiveLoss}
\usage{
AdaptiveLoss(crit)
}
\arguments{
\item{crit}{critic}
}
\value{
Loss object
}
\description{
Expand the `target` to match the `output` size before applying `crit`.
}
