% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{add}
\alias{add}
\alias{+.torch.Tensor}
\alias{sinh.torch.Tensor}
\title{Add}
\usage{
\method{+}{torch.Tensor}(a, b)

\method{sinh}{torch.Tensor}(x)
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}

\item{x}{tensor}
}
\value{
tensor

tensor
}
\description{
Add

Sinh
}
