% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{add}
\alias{add}
\alias{+.torch.Tensor}
\alias{floor.torch.Tensor}
\alias{ceiling.torch.Tensor}
\alias{cos.torch.Tensor}
\alias{cosh.torch.Tensor}
\alias{sin.torch.Tensor}
\alias{sinh.torch.Tensor}
\title{Add}
\usage{
\method{+}{torch.Tensor}(a, b)

\method{floor}{torch.Tensor}(a)

\method{ceiling}{torch.Tensor}(a)

\method{cos}{torch.Tensor}(a)

\method{cosh}{torch.Tensor}(a)

\method{sin}{torch.Tensor}(a)

\method{sinh}{torch.Tensor}(a)
}
\description{
Add

Floor

Ceil

Cos

Cosh

Sin

Sinh
}
