% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{affine_coord}
\alias{affine_coord}
\title{Aaffine_coord}
\usage{
affine_coord(
  x,
  mat = NULL,
  coord_tfm = NULL,
  sz = NULL,
  mode = "bilinear",
  pad_mode = "reflection",
  align_corners = TRUE,
  ...
)
}
\arguments{
\item{x}{x}

\item{mat}{mat}

\item{coord_tfm}{coord_tfm}

\item{sz}{sz}

\item{mode}{mode}

\item{pad_mode}{pad_mode}

\item{align_corners}{align_corners}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Aaffine_coord
}
