% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{alexnet}
\alias{alexnet}
\title{Alexnet}
\usage{
alexnet(pretrained = FALSE, progress)
}
\arguments{
\item{pretrained}{pretrained or not}

\item{progress}{to see progress bar or not}
}
\value{
model
}
\description{
AlexNet model architecture
}
\details{
"One weird trick..." <https://arxiv.org/abs/1404.5997>
}
\examples{

\dontrun{

alexnet(pretrained = FALSE, progress = TRUE)

}

}
