% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_config.R
\name{AudioBlock}
\alias{AudioBlock}
\title{AudioBlock}
\usage{
AudioBlock(
  cache_folder = NULL,
  sample_rate = 16000,
  force_mono = TRUE,
  crop_signal_to = NULL
)
}
\arguments{
\item{cache_folder}{cache_folder}

\item{sample_rate}{sample_rate}

\item{force_mono}{force_mono}

\item{crop_signal_to}{crop_signal_to}
}
\value{
None
}
\description{
A `TransformBlock` for audios
}
