% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_signal.R
\name{AudioGetter}
\alias{AudioGetter}
\title{AudioGetter}
\usage{
AudioGetter(suf = "", recurse = TRUE, folders = NULL)
}
\arguments{
\item{suf}{suffix}

\item{recurse}{recurse}

\item{folders}{folders}
}
\value{
None
}
\description{
Create `get_audio_files` partial function that searches path suffix `suf`
}
\details{
and passes along `kwargs`, only in `folders`, if specified.
}
