% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_signal.R
\name{AudioTensor_create}
\alias{AudioTensor_create}
\title{AudioTensor create}
\usage{
AudioTensor_create(
  fn,
  cache_folder = NULL,
  frame_offset = 0,
  num_frames = -1,
  normalize = TRUE,
  channels_first = TRUE
)
}
\arguments{
\item{fn}{function}

\item{cache_folder}{cache folder}

\item{frame_offset}{offset}

\item{num_frames}{number of frames}

\item{normalize}{apply normalization or not}

\item{channels_first}{channels first/last}
}
\value{
None
}
\description{
Creates audio tensor from file
}
