% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_signal.R
\name{AudioTensor_create}
\alias{AudioTensor_create}
\title{AudioTensor create}
\usage{
AudioTensor_create(
  fn,
  cache_folder = NULL,
  out = NULL,
  normalization = TRUE,
  channels_first = TRUE,
  num_frames = 0,
  offset = 0,
  signalinfo = NULL,
  encodinginfo = NULL,
  filetype = NULL
)
}
\arguments{
\item{fn}{function}

\item{cache_folder}{cache folder}

\item{out}{out format}

\item{normalization}{apply normalization or not}

\item{channels_first}{channels first/last}

\item{num_frames}{number of frames}

\item{offset}{offset}

\item{signalinfo}{signal info}

\item{encodinginfo}{encoding info}

\item{filetype}{the type of file}
}
\value{
None
}
\description{
Creates audio tensor from file
}
