% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{average_sqr_grad}
\alias{average_sqr_grad}
\title{average_sqr_grad}
\usage{
average_sqr_grad(p, sqr_mom, dampening = TRUE, sqr_avg = NULL, ...)
}
\arguments{
\item{p}{p}

\item{sqr_mom}{sqr_mom}

\item{dampening}{dampening}

\item{sqr_avg}{sqr_avg}

\item{...}{additional args to pass}
}
\description{
average_sqr_grad
}
