% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{AvgPool}
\alias{AvgPool}
\title{AvgPool}
\usage{
AvgPool(ks = 2, stride = NULL, padding = 0, ndim = 2, ceil_mode = FALSE)
}
\arguments{
\item{ks}{ks}

\item{stride}{stride}

\item{padding}{padding}

\item{ndim}{ndim}

\item{ceil_mode}{ceil_mode}
}
\value{
None
}
\description{
nn$AvgPool layer for `ndim`
}
