% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{basic_generator}
\alias{basic_generator}
\title{Basic_generator}
\usage{
basic_generator(
  out_size,
  n_channels,
  in_sz = 100,
  n_features = 64,
  n_extra_layers = 0,
  bias = NULL,
  ndim = 2,
  norm_type = 1,
  bn_1st = TRUE,
  act_cls = nn$ReLU,
  init = "auto",
  xtra = NULL,
  bias_std = 0.01,
  dilation = 1,
  groups = 1,
  ...
)
}
\arguments{
\item{out_size}{out_size}

\item{n_channels}{n_channels}

\item{in_sz}{in_sz}

\item{n_features}{The number of features}

\item{n_extra_layers}{The number of extra layers}

\item{bias}{bias}

\item{ndim}{ndim}

\item{norm_type}{norm_type}

\item{bn_1st}{bn_1st}

\item{act_cls}{act_cls}

\item{init}{init}

\item{xtra}{xtra}

\item{bias_std}{bias_std}

\item{dilation}{dilation}

\item{groups}{groups}

\item{...}{additional params to pass}

\item{transpose}{transpose}
}
\description{
A basic generator from `in_sz` to images `n_channels` x `out_size` x `out_size`.
}
