% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{basic_generator}
\alias{basic_generator}
\title{Basic generator}
\usage{
basic_generator(out_size, n_channels, ...)
}
\arguments{
\item{out_size}{out_size}

\item{n_channels}{n_channels}

\item{...}{additional params to pass}
}
\value{
generator object
}
\description{
A basic generator from `in_sz` to images `n_channels` x `out_size` x `out_size`.
}
\examples{
\dontrun{

generator = basic_generator(out_size = 64, n_channels = 3, n_extra_layers = 1)

}

}
