% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_config.R
\name{BasicSpectrogram}
\alias{BasicSpectrogram}
\title{BasicSpectrogram}
\usage{
BasicSpectrogram(
  n_fft = 400,
  win_length = NULL,
  hop_length = NULL,
  pad = 0,
  window_fn = torch$hann_window,
  power = 2,
  normalized = FALSE,
  wkwargs = NULL,
  mel = FALSE,
  to_db = TRUE
)
}
\arguments{
\item{n_fft}{number of fast fourier transforms}

\item{win_length}{windowing length}

\item{hop_length}{hopping length}

\item{pad}{padding}

\item{window_fn}{window function}

\item{power}{power}

\item{normalized}{normalized or not}

\item{wkwargs}{additional arguments}

\item{mel}{mel-spectrogram or not}

\item{to_db}{to decibels}
}
\value{
None
}
\description{
BasicSpectrogram
}
