% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_data.R
\name{BBoxLblBlock}
\alias{BBoxLblBlock}
\title{BBoxLblBlock}
\usage{
BBoxLblBlock(vocab = NULL, add_na = TRUE)
}
\arguments{
\item{vocab}{vocab}

\item{add_na}{add_na}
}
\description{
A `TransformBlock` for labeled bounding boxes, potentially with `vocab`
}
