% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{bt}
\alias{bt}
\title{Built ins}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
bt
}
\value{
module
}
\description{
Built ins
}
\keyword{datasets}
