% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{Categorify}
\alias{Categorify}
\title{Categorify}
\usage{
Categorify(cat_names, cont_names)
}
\arguments{
\item{cat_names}{cat_names}

\item{cont_names}{cont_names}
}
\description{
Transform the categorical variables to that type.
}
