% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{CategoryBlock}
\alias{CategoryBlock}
\title{CategoryBlock}
\usage{
CategoryBlock(vocab = NULL, sort = TRUE, add_na = FALSE)
}
\arguments{
\item{vocab}{vocab}

\item{sort}{sort}

\item{add_na}{add_na}
}
\description{
`TransformBlock` for single-label categorical targets
}
