% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{CategoryBlock}
\alias{CategoryBlock}
\title{CategoryBlock}
\usage{
CategoryBlock(vocab = NULL, sort = TRUE, add_na = FALSE)
}
\arguments{
\item{vocab}{vocab}

\item{sort}{sort}

\item{add_na}{add_na}
}
\value{
Block object
}
\description{
`TransformBlock` for single-label categorical targets
}
