% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_data.R
\name{clip_remove_empty}
\alias{clip_remove_empty}
\title{Clip_remove_empty}
\usage{
clip_remove_empty(bbox, label)
}
\arguments{
\item{bbox}{bbox}

\item{label}{label}
}
\value{
None
}
\description{
Clip bounding boxes with image border and label background the empty ones
}
