% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_learner.R
\name{cnn_config}
\alias{cnn_config}
\title{Cnn_config}
\usage{
cnn_config(
  cut = NULL,
  pretrained = TRUE,
  n_in = 3,
  init = nn$init$kaiming_normal_,
  custom_head = NULL,
  concat_pool = TRUE,
  lin_ftrs = NULL,
  ps = 0.5,
  bn_final = FALSE,
  lin_first = FALSE,
  y_range = NULL
)
}
\arguments{
\item{cut}{cut}

\item{pretrained}{pretrained}

\item{n_in}{number of in}

\item{init}{init}

\item{custom_head}{custom_head}

\item{concat_pool}{concat_pool}

\item{lin_ftrs}{lin_ftrs}

\item{ps}{ps}

\item{bn_final}{bn_final}

\item{lin_first}{lin_first}

\item{y_range}{y_range}
}
\value{
None
}
\description{
Convenience function to easily create a config for `create_cnn_model`
}
