% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab.R
\name{collab_learner}
\alias{collab_learner}
\title{Collab_learner}
\usage{
collab_learner(
  dls,
  n_factors = 50,
  use_nn = FALSE,
  emb_szs = NULL,
  layers = NULL,
  config = NULL,
  y_range = NULL,
  loss_func = NULL,
  opt_func = Adam(),
  lr = 0.001,
  splitter = trainable_params(),
  cbs = NULL,
  metrics = NULL,
  path = NULL,
  model_dir = "models",
  wd = NULL,
  wd_bn_bias = FALSE,
  train_bn = TRUE,
  moms = list(0.95, 0.85, 0.95)
)
}
\arguments{
\item{dls}{dls}

\item{n_factors}{n_factors}

\item{use_nn}{use_nn}

\item{emb_szs}{emb_szs}

\item{layers}{layers}

\item{config}{config}

\item{y_range}{y_range}

\item{loss_func}{loss_func}

\item{opt_func}{opt_func}

\item{lr}{lr}

\item{splitter}{splitter}

\item{cbs}{cbs}

\item{metrics}{metrics}

\item{path}{path}

\item{model_dir}{model_dir}

\item{wd}{wd}

\item{wd_bn_bias}{wd_bn_bias}

\item{train_bn}{train_bn}

\item{moms}{moms}
}
\description{
Create a Learner for collaborative filtering on `dls`.
}
