% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab.R
\name{CollabDataLoaders_from_dblock}
\alias{CollabDataLoaders_from_dblock}
\title{CollabDataLoaders_from_dblock}
\usage{
CollabDataLoaders_from_dblock(
  dblock,
  source,
  path = ".",
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  device = NULL
)
}
\arguments{
\item{dblock}{dblock}

\item{source}{source}

\item{path}{The folder where to work}

\item{bs}{The batch size}

\item{val_bs}{The batch size for the validation DataLoader (defaults to bs)}

\item{shuffle_train}{If we shuffle the training DataLoader or not}

\item{device}{device}
}
\value{
None
}
\description{
Create a dataloaders from a given `dblock`
}
