% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle_wrappers.R
\name{competition_download_file}
\alias{competition_download_file}
\title{competition_download_file}
\usage{
competition_download_file(
  competition,
  file_name,
  path = NULL,
  force = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{competition}{competition}

\item{file_name}{file_name}

\item{path}{path}

\item{force}{force}

\item{quiet}{quiet}
}
\description{
download a competition file to a designated location, or use
}
\details{
a default location Paramters
=========
competition: the name of the competition
file_name: the configuration file name
path: a path to download the file to
force: force the download if the file already exists (default FALSE)
quiet: suppress verbose output (default is FALSE)
}
\section{competition}{
 the name of the competition:
file_name: the configuration file name path: a path to download the file to force: force the download if the file already exists (default FALSE) quiet: suppress verbose output (default is FALSE)
}

\section{file_name}{
 the configuration file name:
path: a path to download the file to force: force the download if the file already exists (default FALSE) quiet: suppress verbose output (default is FALSE)
}

\section{path}{
 a path to download the file to:
force: force the download if the file already exists (default FALSE) quiet: suppress verbose output (default is FALSE)
}

\section{force}{
 force the download if the file already exists (default False):
quiet: suppress verbose output (default is FALSE)
}

