% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle_wrappers.R
\name{competition_download_files}
\alias{competition_download_files}
\title{Competition download files}
\usage{
competition_download_files(
  competition,
  path = NULL,
  force = FALSE,
  quiet = FALSE,
  unzip = FALSE
)
}
\arguments{
\item{competition}{the name of the competition}

\item{path}{a path to download the file to}

\item{force}{force the download if the file already exists (default FALSE)}

\item{quiet}{suppress verbose output (default is TRUE)}
}
\value{
None
}
\description{
Competition download files
}
