% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle_wrappers.R
\name{competition_leaderboard_download}
\alias{competition_leaderboard_download}
\title{Competition leaderboard download}
\usage{
competition_leaderboard_download(competition, path, quiet = TRUE)
}
\arguments{
\item{competition}{the name of the competition}

\item{path}{a path to download the file to}

\item{quiet}{suppress verbose output (default is TRUE)}
}
\value{
data frame
}
\description{
Download competition leaderboards
}
