% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle_wrappers.R
\name{competition_list_files}
\alias{competition_list_files}
\title{Competition list files}
\usage{
competition_list_files(competition)
}
\arguments{
\item{competition}{the name of the competition}
}
\value{
list of files
}
\description{
list files for competition
}
\examples{

\dontrun{

com_nm = 'titanic'
titanic_files = competition_list_files(com_nm)


}
}
