% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle_wrappers.R
\name{competition_submit}
\alias{competition_submit}
\title{competition_submit}
\usage{
competition_submit(file_name, message, competition, quiet = FALSE)
}
\arguments{
\item{file_name}{file_name}

\item{message}{message}

\item{competition}{competition}

\item{quiet}{quiet}
}
\description{
submit a competition!
}
\details{
Parameters
==========
file_name: the competition metadata file
message: the submission description
competition: the competition name
quiet: suppress verbose output (default is FALSE)
}
\section{file_name}{
 the competition metadata file:
message: the submission description competition: the competition name quiet: suppress verbose output (default is FALSE)
}

\section{message}{
 the submission description:
competition: the competition name quiet: suppress verbose output (default is FALSE)
}

\section{competition}{
 the competition name:
quiet: suppress verbose output (default is FALSE)
}

