% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle_wrappers.R
\name{competitions_list}
\alias{competitions_list}
\title{Competitions list}
\usage{
competitions_list(
  group = NULL,
  category = NULL,
  sort_by = NULL,
  page = 1,
  search = NULL
)
}
\arguments{
\item{group}{group to filter result to}

\item{category}{category to filter result to}

\item{sort_by}{how to sort the result, see valid_competition_sort_by for options}

\item{page}{the page to return (default is 1)}

\item{search}{a search term to use (default is empty string)}
}
\value{
list of competitions
}
\description{
Competitions list
}
