% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{CorpusBLEUMetric}
\alias{CorpusBLEUMetric}
\title{CorpusBLEUMetric}
\usage{
CorpusBLEUMetric(vocab_sz = 5000, axis = -1)
}
\arguments{
\item{vocab_sz}{vocab_sz}

\item{axis}{axis}
}
\value{
None
}
\description{
Blueprint for defining a metric
}
