% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_learner.R
\name{create_body}
\alias{create_body}
\title{create_body}
\usage{
create_body(arch, n_in = 3, pretrained = TRUE, cut = NULL)
}
\arguments{
\item{arch}{arch}

\item{n_in}{n_in}

\item{pretrained}{pretrained}

\item{cut}{cut}
}
\description{
Cut off the body of a typically pretrained `arch` as determined by `cut`
}
