% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_learner.R
\name{create_head}
\alias{create_head}
\title{Create_head}
\usage{
create_head(
  nf,
  n_out,
  lin_ftrs = NULL,
  ps = 0.5,
  concat_pool = TRUE,
  bn_final = FALSE,
  lin_first = FALSE,
  y_range = NULL
)
}
\arguments{
\item{nf}{nf}

\item{n_out}{n_out}

\item{lin_ftrs}{lin_ftrs}

\item{ps}{ps}

\item{concat_pool}{concat_pool}

\item{bn_final}{bn_final}

\item{lin_first}{lin_first}

\item{y_range}{y_range}
}
\value{
None
}
\description{
Model head that takes `nf` features, runs through `lin_ftrs`, and out `n_out` classes.
}
