% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{CSVLogger}
\alias{CSVLogger}
\title{CSVLogger}
\usage{
CSVLogger(fname = "history.csv", append = FALSE)
}
\arguments{
\item{fname}{file name}

\item{append}{append or not}
}
\value{
None
}
\description{
Basic class handling tweaks of the training loop by changing a `Learner` in various events
}
