% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{CSVLogger}
\alias{CSVLogger}
\title{CSVLogger}
\usage{
CSVLogger(fname = "history.csv", append = FALSE)
}
\arguments{
\item{fname}{file name}

\item{append}{append or not}
}
\value{
None
}
\description{
Basic class handling tweaks of the training loop by changing a `Learner` in various events
}
\examples{

\dontrun{

URLs_MNIST_SAMPLE()
# transformations
tfms = aug_transforms(do_flip = FALSE)
path = 'mnist_sample'
bs = 20

#load into memory
data = ImageDataLoaders_from_folder(path, batch_tfms = tfms, size = 26, bs = bs)


learn = cnn_learner(data, resnet18(), metrics = accuracy, path = getwd())

learn \%>\% fit_one_cycle(2, cbs = CSVLogger())

}

}
