% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{CSVLogger}
\alias{CSVLogger}
\title{CSVLogger}
\usage{
CSVLogger(learn, filename = "history", append = FALSE)
}
\arguments{
\item{learn}{learn}

\item{filename}{filename}

\item{append}{append}
}
\description{
A `LearnerCallback` that saves history of metrics while training `learn` into CSV `filename`.
}
