% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{DataBlock}
\alias{DataBlock}
\title{DataBlock}
\usage{
DataBlock(
  blocks = NULL,
  dl_type = NULL,
  getters = NULL,
  n_inp = NULL,
  item_tfms = NULL,
  batch_tfms = NULL,
  ...
)
}
\arguments{
\item{blocks}{blocks}

\item{dl_type}{dl_type}

\item{getters}{how to get dataet}

\item{n_inp}{n_inp is the number of elements in the tuples that should be considered part of the input and will default to 1 if tfms consists of one set of transforms}

\item{item_tfms}{One or several transforms applied to the items before batching them}

\item{batch_tfms}{One or several transforms applied to the batches once they are formed}

\item{...}{additional parameters to pass}
}
\value{
block
}
\description{
Generic container to quickly build `Datasets` and `DataLoaders`
}
