% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{dataloaders}
\alias{dataloaders}
\title{dataloaders}
\usage{
dataloaders(object, ...)
}
\arguments{
\item{...}{additional parameters to pass}

\item{source}{source}
}
\description{
Create a `DataLoaders` object from `source`
}
