% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{dataloaders}
\alias{dataloaders}
\title{Dataloaders}
\usage{
dataloaders(
  object,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  n = NULL,
  ...
)
}
\arguments{
\item{object}{object}

\item{bs}{bs}

\item{val_bs}{val_bs}

\item{shuffle_train}{shuffle_train}

\item{n}{n}

\item{...}{parameters to pass}
}
\description{
Get a `DataLoaders`
}
