% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{dataloaders}
\alias{dataloaders}
\title{Dataloaders from dls object}
\usage{
dataloaders(object, ...)
}
\arguments{
\item{object}{model}

\item{...}{additional parameters to pass}
}
\description{
Create a `DataLoaders` object from `source`
}
\examples{
\dontrun{

dls = TabularDataTable(df, procs, cat_names, cont_names,
y_names = dep_var, splits = list(tr_idx, ts_idx) ) \%>\%
  dataloaders(bs = 50)

}

}
