% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{dataloaders}
\alias{dataloaders}
\title{Dataloaders from dls object}
\usage{
dataloaders(object, ...)
}
\arguments{
\item{...}{additional parameters to pass}
}
\description{
Create a `DataLoaders` object from `source`
}
