% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_aug_spectrogram.R
\name{Delta}
\alias{Delta}
\title{Delta}
\usage{
Delta(width = 9)
}
\arguments{
\item{width}{width}
}
\description{
Creates delta with order 1 and 2 from spectrogram
}
\details{
and concatenate with the original
}
