% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{densenet161}
\alias{densenet161}
\title{Densenet161}
\usage{
densenet161(pretrained = FALSE, progress)
}
\arguments{
\item{pretrained}{pretrained or not}

\item{progress}{to see progress bar or not}
}
\value{
model
}
\description{
Densenet161
}
\details{
"Densely Connected Convolutional Networks" <https://arxiv.org/pdf/1608.06993.pdf>
}
