% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{DenseResBlock}
\alias{DenseResBlock}
\title{Dense Res Block}
\usage{
DenseResBlock(
  nf,
  norm_type = 1,
  ks = 3,
  stride = 1,
  padding = NULL,
  bias = NULL,
  ndim = 2,
  bn_1st = TRUE,
  act_cls = nn$ReLU,
  transpose = FALSE,
  init = "auto",
  xtra = NULL,
  bias_std = 0.01,
  dilation = 1,
  groups = 1,
  padding_mode = "zeros"
)
}
\arguments{
\item{nf}{number of features}

\item{norm_type}{normalization type}

\item{ks}{kernel size}

\item{stride}{stride}

\item{padding}{padding}

\item{bias}{bias}

\item{ndim}{number of dimensions}

\item{bn_1st}{batch normalization 1st}

\item{act_cls}{activation}

\item{transpose}{transpose}

\item{init}{initizalier}

\item{xtra}{xtra}

\item{bias_std}{bias standard deviation}

\item{dilation}{dilation number}

\item{groups}{groups number}

\item{padding_mode}{padding mode}
}
\value{
block
}
\description{
Resnet block of `nf` features. `conv_kwargs` are passed to `conv_layer`.
}
