% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{div}
\alias{div}
\alias{/.fastai.torch_core.TensorMask}
\alias{*.fastai.torch_core.TensorMask}
\alias{/.torch.Tensor}
\alias{*.torch.Tensor}
\title{Div}
\usage{
\method{/}{fastai.torch_core.TensorMask}(a, b)

\method{*}{fastai.torch_core.TensorMask}(a, b)

\method{/}{torch.Tensor}(a, b)

\method{*}{torch.Tensor}(a, b)
}
\description{
Div

Multiply

Div

Multiply
}
