% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{error_rate}
\alias{error_rate}
\title{Error_rate}
\usage{
error_rate(inp, targ, axis = -1)
}
\arguments{
\item{inp}{The predictions of the model}

\item{targ}{The corresponding labels}

\item{axis}{Axis}
}
\description{
1 - `accuracy`
}
