% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{ExplainedVariance}
\alias{ExplainedVariance}
\title{ExplainedVariance}
\usage{
ExplainedVariance(sample_weight = NULL)
}
\arguments{
\item{sample_weight}{sample_weight}
}
\description{
Explained variance between predictions and targets
}
