% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{FetchPredsCallback}
\alias{FetchPredsCallback}
\title{FetchPredsCallback}
\usage{
FetchPredsCallback(
  ds_idx = 1,
  dl = NULL,
  with_input = FALSE,
  with_decoded = FALSE,
  cbs = NULL,
  reorder = TRUE
)
}
\arguments{
\item{ds_idx}{dataset index}

\item{dl}{DL application}

\item{with_input}{with input or not}

\item{with_decoded}{with decoded or not}

\item{cbs}{callbacks}

\item{reorder}{reorder or not}
}
\value{
None
}
\description{
A callback to fetch predictions during the training loop
}
