% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{FileSplitter}
\alias{FileSplitter}
\title{FileSplitter}
\usage{
FileSplitter(fname)
}
\arguments{
\item{fname}{fname}
}
\description{
Split `items` by providing file `fname` (contains names of valid items separated by newline).
}
