% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_learner.R
\name{fit.fastai.learner.Learner}
\alias{fit.fastai.learner.Learner}
\title{Fit}
\usage{
\method{fit}{fastai.learner.Learner}(object, ...)
}
\arguments{
\item{object}{a learner object}

\item{...}{parameters to pass}
}
\value{
train history
}
\description{
Fit the model on this learner with `lr` learning rate,
`wd` weight decay for `epochs` with `callbacks` as cbs argument.
}
