% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{fit.fastai.tabular.learner.TabularLearner}
\alias{fit.fastai.tabular.learner.TabularLearner}
\title{Fit}
\usage{
\method{fit}{fastai.tabular.learner.TabularLearner}(object, ...)
}
\arguments{
\item{epochs}{epochs}

\item{lr}{learning rate}

\item{wd}{weight decay}

\item{callbacks}{callbacks}
}
\value{
data frame
}
\description{
Fit the model on this learner with `lr` learning rate, `wd` weight decay for `epochs` with `callbacks`.
}
