% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loaders.R
\name{fit_one_cycle}
\alias{fit_one_cycle}
\title{Fit one cycle}
\usage{
fit_one_cycle(object, n_epoch, ...)
}
\arguments{
\item{n_epoch}{n_epoch}

\item{...}{parameters to pass}
}
\description{
Fit one cycle
}
