% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{FixedGANSwitcher}
\alias{FixedGANSwitcher}
\title{Fixed GAN Switcher}
\usage{
FixedGANSwitcher(n_crit = 1, n_gen = 1)
}
\arguments{
\item{n_crit}{number of discriminator}

\item{n_gen}{number of generator}
}
\value{
None
}
\description{
Switcher to do `n_crit` iterations of the critic then `n_gen` iterations of the generator.
}
