% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R
\name{Flip}
\alias{Flip}
\title{Flip}
\usage{
Flip(
  p = 0.5,
  draw = NULL,
  size = NULL,
  mode = "bilinear",
  pad_mode = "reflection",
  align_corners = TRUE,
  batch = FALSE
)
}
\arguments{
\item{p}{p}

\item{draw}{draw}

\item{size}{size}

\item{mode}{mode}

\item{pad_mode}{pad_mode}

\item{align_corners}{align_corners}

\item{batch}{batch}
}
\description{
Randomly flip a batch of images with a probability `p`
}
