% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{flip_mat}
\alias{flip_mat}
\title{Flip_mat}
\usage{
flip_mat(x, p = 0.5, draw = NULL, batch = FALSE)
}
\arguments{
\item{x}{tensor}

\item{p}{probability}

\item{draw}{draw}

\item{batch}{batch}
}
\value{
None
}
\description{
Return a random flip matrix
}
