% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{floor_div}
\alias{floor_div}
\alias{\%/\%.torch.Tensor}
\title{Floor divide}
\usage{
\method{\%/\%}{torch.Tensor}(x, y)
}
\arguments{
\item{x}{tensor}

\item{y}{tensor}
}
\value{
tensor
}
\description{
Floor divide
}
