% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{gan_critic}
\alias{gan_critic}
\title{Gan critic}
\usage{
gan_critic(n_channels = 3, nf = 128, n_blocks = 3, p = 0.15)
}
\arguments{
\item{n_channels}{number of channels}

\item{nf}{number of features}

\item{n_blocks}{number of blocks}

\item{p}{probability}
}
\value{
GAN object
}
\description{
Critic to train a `GAN`.
}
