% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{GANDiscriminativeLR}
\alias{GANDiscriminativeLR}
\title{GANDiscriminativeLR}
\usage{
GANDiscriminativeLR(mult_lr = 5)
}
\arguments{
\item{mult_lr}{mult_lr}
}
\description{
`Callback` that handles multiplying the learning rate by `mult_lr` for the critic.
}
