% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{GANLearner_wgan}
\alias{GANLearner_wgan}
\title{Wgan}
\usage{
GANLearner_wgan(
  dls,
  generator,
  critic,
  switcher = NULL,
  clip = 0.01,
  switch_eval = FALSE,
  gen_first = FALSE,
  show_img = TRUE,
  cbs = NULL,
  metrics = NULL,
  opt_func = Adam(),
  lr = 0.001,
  splitter = trainable_params,
  path = NULL,
  model_dir = "models",
  wd = NULL,
  wd_bn_bias = FALSE,
  train_bn = TRUE,
  moms = list(0.95, 0.85, 0.95)
)
}
\arguments{
\item{dls}{dls}

\item{generator}{generator}

\item{critic}{critic}

\item{switcher}{switcher}

\item{clip}{clip}

\item{switch_eval}{switch_eval}

\item{gen_first}{gen_first}

\item{show_img}{show_img}

\item{cbs}{cbs}

\item{metrics}{metrics}

\item{opt_func}{opt_func}

\item{lr}{lr}

\item{splitter}{splitter}

\item{path}{path}

\item{model_dir}{model_dir}

\item{wd}{wd}

\item{wd_bn_bias}{wd_bn_bias}

\item{train_bn}{train_bn}

\item{moms}{moms}
}
\description{
Create a WGAN from `data`, `generator` and `critic`.
}
