% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{GANLoss}
\alias{GANLoss}
\title{GANLoss}
\usage{
GANLoss(gen_loss_func, crit_loss_func, gan_model)
}
\arguments{
\item{gen_loss_func}{gen_loss_func}

\item{crit_loss_func}{crit_loss_func}

\item{gan_model}{gan_model}
}
\description{
Wrapper around `crit_loss_func` and `gen_loss_func`
}
