% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{GANTrainer}
\alias{GANTrainer}
\title{GAN Trainer}
\usage{
GANTrainer(
  switch_eval = FALSE,
  clip = NULL,
  beta = 0.98,
  gen_first = FALSE,
  show_img = TRUE
)
}
\arguments{
\item{switch_eval}{switch_eval}

\item{clip}{clip}

\item{beta}{beta}

\item{gen_first}{gen_first}

\item{show_img}{show_img}
}
\value{
None
}
\description{
Handles GAN Training.
}
