% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{GatherPredsCallback}
\alias{GatherPredsCallback}
\title{GatherPredsCallback}
\usage{
GatherPredsCallback(
  with_input = FALSE,
  with_loss = FALSE,
  save_preds = NULL,
  save_targs = NULL,
  concat_dim = 0
)
}
\arguments{
\item{with_input}{include inputs or not}

\item{with_loss}{include loss or not}

\item{save_preds}{save predictions}

\item{save_targs}{save targets/actuals}

\item{concat_dim}{concatenate dimensions}
}
\value{
None
}
\description{
`Callback` that saves the predictions and targets, optionally `with_loss`
}
