% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_modules.R
\name{get_data_loaders}
\alias{get_data_loaders}
\title{get_data_loaders}
\usage{
get_data_loaders(train_batch_size, val_batch_size)
}
\arguments{
\item{train_batch_size}{train_batch_size}

\item{val_batch_size}{val_batch_size}
}
\description{
get_data_loaders
}
