% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medical.R
\name{get_dicom_files}
\alias{get_dicom_files}
\title{get_dicom_files}
\usage{
get_dicom_files(path, recurse = TRUE, folders = NULL)
}
\arguments{
\item{path}{path}

\item{recurse}{recurse}

\item{folders}{folders}
}
\description{
Get dicom files in `path` recursively, only in `folders`, if specified.
}
