% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_data.R
\name{get_grid}
\alias{get_grid}
\title{get_grid}
\usage{
get_grid(
  n,
  nrows = NULL,
  ncols = NULL,
  add_vert = 0,
  figsize = NULL,
  double = FALSE,
  title = NULL,
  return_fig = FALSE,
  imsize = 3
)
}
\arguments{
\item{n}{n}

\item{nrows}{nrows}

\item{ncols}{ncols}

\item{add_vert}{add_vert}

\item{figsize}{figsize}

\item{double}{double}

\item{title}{title}

\item{return_fig}{return_fig}

\item{imsize}{imsize}
}
\description{
Return a grid of `n` axes, `rows` by `cols`
}
\details{

}
