% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loaders.R
\name{get_image_files}
\alias{get_image_files}
\title{Get image files}
\usage{
get_image_files(path, recurse = TRUE, folders = NULL)
}
\arguments{
\item{path}{The folder where to work}

\item{recurse}{recursive path}

\item{folders}{folder names}
}
\value{
None
}
\description{
Get image files in `path` recursively, only in `folders`, if specified.
}
\examples{

\dontrun{

URLs_PETS()

path = 'oxford-iiit-pet'

path_img = 'oxford-iiit-pet/images'
fnames = get_image_files(path_img)

}

}
