% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab.R
\name{get_weights}
\alias{get_weights}
\title{Get weights}
\usage{
get_weights(object, arr, is_item = TRUE, convert = FALSE)
}
\arguments{
\item{object}{extract weights}

\item{arr}{R data frame}

\item{is_item}{logical, is item}

\item{convert}{to R matrix}
}
\value{
tensor
}
\description{
Weight for item or user (based on `is_item`) for all in `arr`
}
\examples{

\dontrun{

movie_w = learn \%>\% get_weights(top_movies, is_item = TRUE, convert = TRUE)

}

}
