% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{greater_or_equal}
\alias{greater_or_equal}
\alias{>=.torch.Tensor}
\title{Greater or equal}
\usage{
\method{>=}{torch.Tensor}(a, b)
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}
}
\value{
tensor
}
\description{
Greater or equal
}
