% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loss.R, R/metric.R
\name{HammingLoss}
\alias{HammingLoss}
\title{HammingLoss}
\usage{
HammingLoss(axis = -1, sample_weight = NULL)

HammingLoss(axis = -1, sample_weight = NULL)
}
\arguments{
\item{axis}{axis}

\item{sample_weight}{sample_weight}

\item{...}{parameters to pass}
}
\value{
Loss object

None
}
\description{
Hamming loss for single-label classification problems

Hamming loss for single-label classification problems
}
