% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loss.R, R/metric.R
\name{HammingLossMulti}
\alias{HammingLossMulti}
\title{HammingLossMulti}
\usage{
HammingLossMulti(
  thresh = 0.5,
  sigmoid = TRUE,
  labels = NULL,
  sample_weight = NULL
)

HammingLossMulti(
  thresh = 0.5,
  sigmoid = TRUE,
  labels = NULL,
  sample_weight = NULL
)
}
\arguments{
\item{thresh}{thresh}

\item{sigmoid}{sigmoid}

\item{labels}{labels}

\item{sample_weight}{sample_weight}
}
\description{
Hamming loss for multi-label classification problems

Hamming loss for multi-label classification problems
}
