% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurr_hugging_face.R
\name{HF_TextBlock}
\alias{HF_TextBlock}
\title{HF_TextBlock}
\usage{
HF_TextBlock(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
A basic wrapper that links defaults transforms for the data block API
}
