% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R, R/callbacks_hook.R
\name{HookCallback}
\alias{HookCallback}
\title{HookCallback}
\usage{
HookCallback(
  modules = NULL,
  every = NULL,
  remove_end = TRUE,
  is_forward = TRUE,
  detach = TRUE,
  cpu = TRUE
)

HookCallback(
  modules = NULL,
  every = NULL,
  remove_end = TRUE,
  is_forward = TRUE,
  detach = TRUE,
  cpu = TRUE
)
}
\arguments{
\item{modules}{modules}

\item{every}{every}

\item{remove_end}{remove_end or not}

\item{is_forward}{is_forward or not}

\item{detach}{detach or not}

\item{cpu}{cpu or not}
}
\value{
None

None
}
\description{
`Callback` that can be used to register hooks on `modules`

`Callback` that can be used to register hooks on `modules`
}
