% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{HookCallback}
\alias{HookCallback}
\title{HookCallback}
\usage{
HookCallback(
  modules = NULL,
  every = NULL,
  remove_end = TRUE,
  is_forward = TRUE,
  detach = TRUE,
  cpu = TRUE
)
}
\arguments{
\item{modules}{the modules}

\item{every}{int, every epoch}

\item{remove_end}{logical, remove_end}

\item{is_forward}{logical, is_forward}

\item{detach}{detach}

\item{cpu}{to cpu or not}
}
\value{
None
}
\description{
`Callback` that can be used to register hooks on `modules`
}
