% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loaders.R
\name{ImageDataLoaders_from_folder}
\alias{ImageDataLoaders_from_folder}
\title{ImageDataLoaders from folder}
\usage{
ImageDataLoaders_from_folder(
  path,
  train = "train",
  valid = "valid",
  valid_pct = NULL,
  seed = NULL,
  vocab = NULL,
  item_tfms = NULL,
  batch_tfms = NULL,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  device = NULL,
  size = NULL,
  ...
)
}
\arguments{
\item{path}{The folder where to work}

\item{train}{train data}

\item{valid}{validation data}

\item{valid_pct}{valid_pct}

\item{seed}{random seed}

\item{vocab}{vocabulary}

\item{item_tfms}{One or several transforms applied to the items before batching them}

\item{batch_tfms}{One or several transforms applied to the batches once they are formed}

\item{bs}{batch size}

\item{val_bs}{The batch size for the validation DataLoader (defaults to bs)}

\item{shuffle_train}{If we shuffle the training DataLoader or not}

\item{device}{device}

\item{size}{image size}

\item{...}{additional parameters to pass}
}
\description{
Create from imagenet style dataset in `path` with `train` and `valid` subfolders (or provide `valid_pct`)
}
