% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loaders.R
\name{ImageDataLoaders_from_lists}
\alias{ImageDataLoaders_from_lists}
\title{ImageDataLoaders_from_lists}
\usage{
ImageDataLoaders_from_lists(
  path,
  fnames,
  labels,
  valid_pct = 0.2,
  seed = NULL,
  y_block = NULL,
  item_tfms = NULL,
  batch_tfms = NULL,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  device = NULL
)
}
\arguments{
\item{path}{The folder where to work}

\item{fnames}{fnames}

\item{labels}{labels}

\item{valid_pct}{valid_pct}

\item{seed}{seed}

\item{y_block}{y_block}

\item{item_tfms}{One or several transforms applied to the items before batching them}

\item{batch_tfms}{One or several transforms applied to the batches once they are formed}

\item{bs}{The batch size}

\item{val_bs}{The batch size for the validation DataLoader (defaults to bs)}

\item{shuffle_train}{If we shuffle the training DataLoader or not}

\item{device}{device}
}
\description{
Create from list of `fnames` and `labels` in `path`
}
