% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loaders.R
\name{ImageDataLoaders_from_path_re}
\alias{ImageDataLoaders_from_path_re}
\title{ImageDataLoaders_from_path_re}
\usage{
ImageDataLoaders_from_path_re(
  path,
  fnames,
  pat,
  valid_pct = 0.2,
  seed = NULL,
  item_tfms = NULL,
  batch_tfms = NULL,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  device = NULL,
  ...
)
}
\arguments{
\item{path}{The folder where to work}

\item{fnames}{file names}

\item{pat}{an argument that requires regex}

\item{valid_pct}{The random percentage of the dataset to set aside for validation (with an optional seed)}

\item{seed}{random seed}

\item{item_tfms}{One or several transforms applied to the items before batching them}

\item{batch_tfms}{One or several transforms applied to the batches once they are formed}

\item{bs}{batch size}

\item{val_bs}{The batch size for the validation DataLoader (defaults to bs)}

\item{shuffle_train}{If we shuffle the training DataLoader or not}

\item{device}{device}

\item{...}{additional parameters to pass}
}
\value{
None
}
\description{
Create from list of `fnames` in `path`s with re expression `pat`
}
