% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R
\name{ImagePoints}
\alias{ImagePoints}
\title{ImagePoints}
\usage{
ImagePoints(flow, scale = TRUE, y_first = TRUE)
}
\arguments{
\item{flow}{flow}

\item{scale}{scale}

\item{y_first}{y_first}
}
\description{
Support applying transforms to a `flow` of points.
}
\details{

}
