% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_fastai.R
\name{install_fastai}
\alias{install_fastai}
\title{Install fastai}
\usage{
install_fastai(version, gpu = FALSE, cuda_version = "10.1", overwrite = FALSE)
}
\arguments{
\item{version}{specify version}

\item{gpu}{installation of gpu}

\item{cuda_version}{if gpu true, then cuda version is required. By default it is 10.1}

\item{overwrite}{will install all the dependencies}
}
\description{
Install fastai
}
