% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{IntToFloatTensor}
\alias{IntToFloatTensor}
\title{IntToFloatTensor}
\usage{
IntToFloatTensor(div = 255, div_mask = 1)
}
\arguments{
\item{div}{divide value}

\item{div_mask}{divide mask}
}
\value{
None
}
\description{
Transform image to float tensor, optionally dividing by 255 (e.g. for images).
}
