% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{l2_reg}
\alias{l2_reg}
\title{l2_reg}
\usage{
l2_reg(p, lr, wd, do_wd = TRUE, ...)
}
\arguments{
\item{p}{p}

\item{lr}{lr}

\item{wd}{wd}

\item{do_wd}{do_wd}

\item{...}{additional args to pass}
}
\description{
L2 regularization as adding `wd*p` to `p.grad`
}
