% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loss.R
\name{LabelSmoothingCrossEntropyFlat}
\alias{LabelSmoothingCrossEntropyFlat}
\title{LabelSmoothingCrossEntropyFlat}
\usage{
LabelSmoothingCrossEntropyFlat(...)
}
\arguments{
\item{...}{additional parameters to pass}
}
\description{
Same as `nn.Module`, but no need for subclasses to call `super().__init__`
}
