% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{larc_step}
\alias{larc_step}
\title{larc_step}
\usage{
larc_step(p, local_lr, grad_avg = NULL, ...)
}
\arguments{
\item{p}{p}

\item{local_lr}{local_lr}

\item{grad_avg}{grad_avg}

\item{...}{additional args to pass}
}
\description{
Step for LARC `local_lr` on `p`
}
