% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{less_or_equal}
\alias{less_or_equal}
\alias{<=.torch.Tensor}
\title{Less or equal}
\usage{
\method{<=}{torch.Tensor}(a, b)
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}
}
\value{
tensor
}
\description{
Less or equal
}
