% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{log}
\alias{log}
\alias{log.torch.Tensor}
\title{Log}
\usage{
\method{log}{torch.Tensor}(x, base = exp(1))
}
\arguments{
\item{x}{tensor}

\item{base}{base parameter}
}
\value{
tensor
}
\description{
Log
}
