% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{logical_or}
\alias{logical_or}
\alias{|.torch.Tensor}
\alias{!.torch.Tensor}
\title{Logical_or}
\usage{
\method{|}{torch.Tensor}(a, b)

\method{!}{torch.Tensor}(a)
}
\description{
Logical_or

Logical_not
}
