% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{logical_or}
\alias{logical_or}
\alias{|.torch.Tensor}
\title{Logical_or}
\usage{
\method{|}{torch.Tensor}(a, b)
}
\arguments{
\item{x}{tensor}

\item{y}{tensor}
}
\value{
tensor
}
\description{
Logical_or
}
