% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medical.R
\name{mask_from_blur}
\alias{mask_from_blur}
\title{mask_from_blur}
\usage{
mask_from_blur(img, window, sigma = 0.3, thresh = 0.05, remove_max = TRUE)
}
\arguments{
\item{window}{window}

\item{sigma}{sigma}

\item{thresh}{thresh}

\item{remove_max}{remove_max}

\item{x}{x}
}
\description{
mask_from_blur
}
