% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{max}
\alias{max}
\alias{max.torch.Tensor}
\title{Max}
\usage{
\method{max}{torch.Tensor}(a, ..., na.rm = FALSE)
}
\arguments{
\item{a}{tensor}

\item{...}{additional parameters}

\item{na.rm}{remove NAs}
}
\value{
tensor
}
\description{
Max
}
