% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{momentum_step}
\alias{momentum_step}
\title{momentum_step}
\usage{
momentum_step(p, lr, grad_avg, ...)
}
\arguments{
\item{p}{p}

\item{lr}{lr}

\item{grad_avg}{grad_avg}

\item{...}{additional args to pass}
}
\description{
Step for SGD with momentum with `lr`
}
