% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R, R/test.R
\name{most_confused}
\alias{most_confused}
\title{Most_confused}
\usage{
most_confused(interp, min_val = 1)

most_confused(interp, min_val = 1)
}
\arguments{
\item{min_val}{min_val}

\item{object}{interpret object}
}
\value{
data.frame
}
\description{
Sorted descending list of largest non-diagonal entries of confusion matrix,
presented as actual, predicted, number of occurrences.

Sorted descending list of largest non-diagonal entries of confusion matrix, presented as actual, predicted, number of occurrences.
}
