% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{MultiCategorize}
\alias{MultiCategorize}
\title{MultiCategorize}
\usage{
MultiCategorize(vocab = NULL, add_na = FALSE)
}
\arguments{
\item{vocab}{vocabulary}

\item{add_na}{add NA}
}
\value{
None
}
\description{
Reversible transform of multi-category strings to `vocab` id
}
