% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{nn_loss}
\alias{nn_loss}
\title{Fastai custom loss}
\usage{
nn_loss(model_fn, name = "CustomLoss")
}
\arguments{
\item{model_fn}{pass custom model function}

\item{name}{set name for nn_module}
}
\value{
None
}
\description{
Fastai custom loss
}
