% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{norm_apply_denorm}
\alias{norm_apply_denorm}
\title{Norm_apply_denorm}
\usage{
norm_apply_denorm(x, f, nrm)
}
\arguments{
\item{x}{tensor}

\item{f}{function}

\item{nrm}{nrm}
}
\value{
None
}
\description{
Normalize `x` with `nrm`, then apply `f`, then denormalize
}
