% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_tensor.R
\name{not_equal_to}
\alias{not_equal_to}
\alias{!=.torch.Tensor}
\title{Not equal}
\usage{
\method{!=}{torch.Tensor}(a, b)
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}
}
\value{
tensor
}
\description{
Not equal
}
