% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{OldRandomCrop}
\alias{OldRandomCrop}
\title{OldRandomCrop}
\usage{
OldRandomCrop(size, pad_mode = "zeros", ...)
}
\arguments{
\item{size}{size}

\item{pad_mode}{padding mode}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Randomly crop an image to `size`
}
