% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_readers.R
\name{open_mask}
\alias{open_mask}
\title{Open mask}
\usage{
open_mask(fn, div = FALSE, convert_mode = "L", after_open = NULL)
}
\arguments{
\item{fn}{path}

\item{div}{divide or not}

\item{convert_mode}{convert_mode}

\item{after_open}{after open}
}
\value{
ImageSegment object create from mask in file `fn`. If `div`, divides pixel values by 255.
}
\description{
Open mask
}
