% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_readers.R
\name{open_mask_rle}
\alias{open_mask_rle}
\title{open_mask_rle}
\usage{
open_mask_rle(mask_rle, shape)
}
\arguments{
\item{mask_rle}{mask_rle}

\item{shape}{shape}
}
\description{
Return `ImageSegment` object create from run-length encoded string in `mask_lre` with size in `shape`.
}
