% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{parallel}
\alias{parallel}
\title{Parallel}
\usage{
parallel(f, items, ...)
}
\arguments{
\item{f}{file names}

\item{items}{items}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Applies `func` in parallel to `items`, using `n_workers`
}
