% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{partial}
\alias{partial}
\title{Partial}
\usage{
partial(...)
}
\arguments{
\item{...}{additional arguments}
}
\value{
None
}
\description{
partial(func, *args, **keywords) - new function with partial application
}
