% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{Pipeline}
\alias{Pipeline}
\title{Pipeline}
\usage{
Pipeline(funcs = NULL, split_idx = NULL)
}
\arguments{
\item{funcs}{funcs}

\item{split_idx}{split_idx}
}
\description{
A pipeline of composed (for encode/decode) transforms, setup with types
}
