% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{plot}
\alias{plot}
\title{Plot dicom}
\usage{
plot(object, ..., dpi = 100)
}
\arguments{
\item{object}{model}

\item{...}{parameters to pass}

\item{dpi}{dots per inch}
}
\value{
None
}
\description{
Plot dicom
}
