% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot confusion matrix}
\usage{
plot_confusion_matrix(object, dataloader)
}
\arguments{
\item{object}{model}

\item{dataloader}{dataloaders object}
}
\description{
Plot confusion matrix
}
