% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot_confusion_matrix}
\usage{
plot_confusion_matrix(
  interp,
  normalize = FALSE,
  title = "Confusion matrix",
  cmap = "Blues",
  norm_dec = 2,
  plot_txt = TRUE,
  figsize = c(4, 4),
  ...,
  dpi = 120
)
}
\arguments{
\item{interp}{interpretation object}

\item{normalize}{normalize}

\item{title}{title}

\item{cmap}{color map}

\item{norm_dec}{norm dec}

\item{plot_txt}{plot text}

\item{figsize}{plot size}

\item{...}{additional parameters to pass}

\item{dpi}{dots per inch}
}
\value{
None
}
\description{
Plot the confusion matrix, with `title` and using `cmap`.
}
\examples{

\dontrun{

interp = ClassificationInterpretation_from_learner(model)
interp \%>\% plot_confusion_matrix(dpi = 90,figsize = c(6,6))

}

}
