% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{plot_loss}
\alias{plot_loss}
\title{plot_loss}
\usage{
plot_loss(object, skip_start = 5, with_valid = TRUE, dpi = 100)
}
\arguments{
\item{skip_start}{skip_start}

\item{with_valid}{with_valid}
}
\description{
Plot the losses from `skip_start` and onward
}
