% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{plot_loss}
\alias{plot_loss}
\title{Plot_loss}
\usage{
plot_loss(object, skip_start = 5, with_valid = TRUE, dpi = 200)
}
\arguments{
\item{object}{model}

\item{skip_start}{n points to skip the start}

\item{with_valid}{with validation}

\item{dpi}{dots per inch}
}
\value{
None
}
\description{
Plot the losses from `skip_start` and onward
}
