% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{plot_top_losses}
\alias{plot_top_losses}
\title{Plot_top_losses}
\usage{
plot_top_losses(
  interp,
  k,
  largest = TRUE,
  figsize = c(19.2, 10.8),
  ...,
  dpi = 90
)
}
\arguments{
\item{interp}{interpretation object}

\item{k}{number of images}

\item{largest}{largest}

\item{figsize}{plot size}

\item{...}{additional parameters to pass}

\item{dpi}{dots per inch}
}
\value{
None
}
\description{
Plot_top_losses
}
\examples{

\dontrun{

# get interperetation from learn object, the model.
interp = ClassificationInterpretation_from_learner(learn)
interp \%>\% plot_top_losses(k = 9, figsize = c(15,11))

}

}
