% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{print.fastai.tabular.learner.TabularLearner}
\alias{print.fastai.tabular.learner.TabularLearner}
\title{Print model}
\usage{
\method{print}{fastai.tabular.learner.TabularLearner}(x, ...)
}
\description{
Print model
}
