% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medical.R
\name{print.pydicom.dataset.FileDataset}
\alias{print.pydicom.dataset.FileDataset}
\title{Dicom}
\usage{
\method{print}{pydicom.dataset.FileDataset}(img)
}
\arguments{
\item{img}{dicom file}
}
\description{
Dicom
}
