% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{qhadam_step}
\alias{qhadam_step}
\title{qhadam_step}
\usage{
qhadam_step(p, lr, mom, sqr_mom, sqr_avg, nu_1, nu_2, step, grad_avg, eps, ...)
}
\arguments{
\item{p}{p}

\item{lr}{lr}

\item{mom}{mom}

\item{sqr_mom}{sqr_mom}

\item{sqr_avg}{sqr_avg}

\item{nu_1}{nu_1}

\item{nu_2}{nu_2}

\item{step}{step}

\item{grad_avg}{grad_avg}

\item{eps}{eps}

\item{...}{additional args to pass}
}
\description{
qhadam_step
}
