% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{radam_step}
\alias{radam_step}
\title{radam_step}
\usage{
radam_step(p, lr, mom, step, sqr_mom, grad_avg, sqr_avg, eps, beta, ...)
}
\arguments{
\item{p}{p}

\item{lr}{lr}

\item{mom}{mom}

\item{step}{step}

\item{sqr_mom}{sqr_mom}

\item{grad_avg}{grad_avg}

\item{sqr_avg}{sqr_avg}

\item{eps}{eps}

\item{beta}{beta}

\item{...}{additional args to pass}
}
\description{
Step for RAdam with `lr` on `p`
}
