% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loaders.R
\name{random_batch}
\alias{random_batch}
\title{Random image batch}
\usage{
random_batch(object, regex = "[A-z]+_")
}
\arguments{
\item{object}{dataloaders object}

\item{regex}{for img titles}
}
\description{
for visualization
}
