% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loaders.R
\name{random_batch}
\alias{random_batch}
\title{Random image batch}
\usage{
random_batch(
  object,
  regex = "[0-9]+",
  label = "label",
  folder_name = NULL,
  img_patter = "jpg|jpeg|png"
)
}
\arguments{
\item{object}{dataloaders object}

\item{regex}{for img titles}

\item{label}{if dataloader from csv, then colname should be provided for viz-n}

\item{folder_name}{if dataloader from csv, then colname should be provided for viz-n}
}
\description{
for visualization
}
