% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{RandomErasing}
\alias{RandomErasing}
\title{RandomErasing}
\usage{
RandomErasing(p = 0.5, sl = 0, sh = 0.3, min_aspect = 0.3, max_count = 1)
}
\arguments{
\item{p}{p}

\item{sl}{sl}

\item{sh}{sh}

\item{min_aspect}{min_aspect}

\item{max_count}{max_count}
}
\value{
None
}
\description{
Randomly selects a rectangle region in an image and randomizes its pixels.
}
