% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\name{RandomSplitter}
\alias{RandomSplitter}
\title{RandomSplitter}
\usage{
RandomSplitter(valid_pct = 0.2, seed = NULL)
}
\arguments{
\item{valid_pct}{valid_pct}

\item{seed}{seed}
}
\description{
Create function that splits `items` between train/val with `valid_pct` randomly.
}
