% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{ranger}
\alias{ranger}
\title{ranger}
\usage{
ranger(
  p,
  lr,
  mom = 0.95,
  wd = 0.01,
  eps = 1e-06,
  sqr_mom = 0.99,
  beta = 0,
  decouple_wd = TRUE
)
}
\arguments{
\item{p}{p}

\item{lr}{lr}

\item{mom}{mom}

\item{wd}{wd}

\item{eps}{eps}

\item{sqr_mom}{sqr_mom}

\item{beta}{beta}

\item{decouple_wd}{decouple_wd}
}
\description{
Convenience method for `Lookahead` with `RAdam`
}
