% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{ResBlock}
\alias{ResBlock}
\title{ResBlock}
\usage{
ResBlock(
  expansion,
  ni,
  nf,
  stride = 1,
  groups = 1,
  reduction = NULL,
  nh1 = NULL,
  nh2 = NULL,
  dw = FALSE,
  g2 = 1,
  sa = FALSE,
  sym = FALSE,
  norm_type = 1,
  act_cls = nn$ReLU,
  ndim = 2,
  ks = 3,
  pool = AvgPool(),
  pool_first = TRUE,
  padding = NULL,
  bias = NULL,
  bn_1st = TRUE,
  transpose = FALSE,
  init = "auto",
  xtra = NULL,
  bias_std = 0.01,
  dilation = 1,
  padding_mode = "zeros"
)
}
\arguments{
\item{expansion}{expansion}

\item{ni}{ni}

\item{nf}{nf}

\item{stride}{stride}

\item{groups}{groups}

\item{reduction}{reduction}

\item{nh1}{nh1}

\item{nh2}{nh2}

\item{dw}{dw}

\item{g2}{g2}

\item{sa}{sa}

\item{sym}{sym}

\item{norm_type}{norm_type}

\item{act_cls}{act_cls}

\item{ndim}{ndim}

\item{ks}{ks}

\item{pool}{pool}

\item{pool_first}{pool_first}

\item{padding}{padding}

\item{bias}{bias}

\item{bn_1st}{bn_1st}

\item{transpose}{transpose}

\item{init}{init}

\item{xtra}{xtra}

\item{bias_std}{bias_std}

\item{dilation}{dilation}

\item{padding_mode}{padding_mode}
}
\value{
Block object
}
\description{
Resnet block from `ni` to `nh` with `stride`
}
