% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{reshape}
\alias{reshape}
\title{reshape}
\usage{
reshape(x, h, w, resample = 0)
}
\arguments{
\item{x}{x}

\item{h}{h}

\item{w}{w}

\item{resample}{resample}
}
\description{
`resize` `x` to `(w,h)`
}
\details{

}
