% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R
\name{Resize}
\alias{Resize}
\title{Resize}
\usage{
Resize(size, method = "crop", pad_mode = "reflection", resamples = list(2, 0))
}
\arguments{
\item{size}{size}

\item{method}{method}

\item{pad_mode}{pad_mode}

\item{resamples}{resamples}
}
\description{
A transform that before_call its state at each `__call__`
}
