% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{resize}
\alias{resize}
\title{resize}
\usage{
resize(img, size, resample = 3, box = NULL, reducing_gap = NULL)
}
\arguments{
\item{img}{image}

\item{size}{size}

\item{resample}{resample}

\item{box}{box}

\item{reducing_gap}{reducing_gap}
}
\description{
Returns a resized copy of this image.
}
