% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{resnet101}
\alias{resnet101}
\title{Resnet101}
\usage{
resnet101(pretrained = FALSE, progress)
}
\arguments{
\item{pretrained}{pretrained or not}

\item{progress}{to see progress bar or not}
}
\value{
model
}
\description{
ResNet-101 model from
}
\details{
"Deep Residual Learning for Image Recognition" <https://arxiv.org/pdf/1512.03385.pdf>
}
