% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{resnet152}
\alias{resnet152}
\title{resnet152}
\usage{
resnet152(pretrained = FALSE, progress = TRUE)
}
\arguments{
\item{pretrained}{pretrained}

\item{progress}{progress}
}
\description{
ResNet-152 model from
}
\details{
`"Deep Residual Learning for Image Recognition" <https://arxiv.org/pdf/1512.03385.pdf>`_
}
