% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastaibuilt.R
\name{RetinaNet}
\alias{RetinaNet}
\title{RetinaNet}
\usage{
RetinaNet(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
model
}
\description{
Implements RetinaNet from https://arxiv.org/abs/1708.02002
}
