% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{rms_prop_step}
\alias{rms_prop_step}
\title{rms_prop_step}
\usage{
rms_prop_step(p, lr, sqr_avg, eps, grad_avg = NULL, ...)
}
\arguments{
\item{p}{p}

\item{lr}{lr}

\item{sqr_avg}{sqr_avg}

\item{eps}{eps}

\item{grad_avg}{grad_avg}

\item{...}{additional args to pass}
}
\description{
Step for SGD with momentum with `lr`
}
