% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R
\name{Rotate}
\alias{Rotate}
\title{Rotate}
\usage{
Rotate(
  max_deg = 10,
  p = 0.5,
  draw = NULL,
  size = NULL,
  mode = "bilinear",
  pad_mode = "reflection",
  align_corners = TRUE,
  batch = FALSE
)
}
\arguments{
\item{max_deg}{maximum degrees}

\item{p}{probability}

\item{draw}{draw}

\item{size}{size of image}

\item{mode}{mode}

\item{pad_mode}{reflection, zeros, border as string parameter}

\item{align_corners}{align corners or not}

\item{batch}{batch or not}
}
\value{
None
}
\description{
Apply a random rotation of at most `max_deg` with probability `p` to a batch of images
}
