% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_mask.R, R/generics_tensor.R
\name{round}
\alias{round}
\alias{round.fastai.torch_core.TensorMask}
\alias{round.torch.Tensor}
\title{Round}
\usage{
\method{round}{fastai.torch_core.TensorMask}(a)

\method{round}{torch.Tensor}(a)
}
\description{
Round

Round
}
