% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{SEBlock}
\alias{SEBlock}
\title{SEBlock}
\usage{
SEBlock(expansion, ni, nf, groups = 1, reduction = 16, stride = 1)
}
\arguments{
\item{expansion}{decoder}

\item{ni}{ni}

\item{nf}{nf}

\item{groups}{number of groups}

\item{reduction}{number of reduction}

\item{stride}{number of stride}
}
\value{
Block object
}
\description{
SEBlock
}
