% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_data.R
\name{SegmentationDataLoaders_from_label_func}
\alias{SegmentationDataLoaders_from_label_func}
\title{SegmentationDataLoaders_from_label_func}
\usage{
SegmentationDataLoaders_from_label_func(
  path,
  fnames,
  label_func,
  valid_pct = 0.2,
  seed = NULL,
  codes = NULL,
  item_tfms = NULL,
  batch_tfms = NULL,
  bs = 64,
  val_bs = NULL,
  shuffle_train = TRUE,
  device = NULL
)
}
\arguments{
\item{path}{path}

\item{fnames}{file names}

\item{label_func}{label function}

\item{valid_pct}{validation percentage}

\item{seed}{seed}

\item{codes}{codes}

\item{item_tfms}{item transformations}

\item{batch_tfms}{batch transformations}

\item{bs}{batch size}

\item{val_bs}{validation batch size}

\item{shuffle_train}{shuffle train}

\item{device}{device name}
}
\value{
None
}
\description{
Create from list of `fnames` in `path`s with `label_func`.
}
