% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{SeparableBlock}
\alias{SeparableBlock}
\title{SeparableBlock}
\usage{
SeparableBlock(expansion, ni, nf, reduction = 16, stride = 1, base_width = 4)
}
\arguments{
\item{expansion}{expansion}

\item{ni}{ni}

\item{nf}{nf}

\item{reduction}{reduction}

\item{stride}{stride}

\item{base_width}{base_width}
}
\description{
SeparableBlock
}
