% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{SeparableBlock}
\alias{SeparableBlock}
\title{SeparableBlock}
\usage{
SeparableBlock(expansion, ni, nf, reduction = 16, stride = 1, base_width = 4)

SeparableBlock(expansion, ni, nf, reduction = 16, stride = 1, base_width = 4)
}
\arguments{
\item{expansion}{decoder}

\item{ni}{number of linear inputs}

\item{nf}{number of features}

\item{reduction}{reduction number}

\item{stride}{stride number}

\item{base_width}{int, base width}
}
\value{
Block object

Block object
}
\description{
SeparableBlock

SeparableBlock
}
