% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{sgd_step}
\alias{sgd_step}
\title{Sgd_step}
\usage{
sgd_step(p, lr, ...)
}
\arguments{
\item{p}{p}

\item{lr}{learning rate}

\item{...}{additional arguments to pass}
}
\value{
None
}
\description{
Sgd_step
}
