% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{show}
\alias{show}
\title{show}
\usage{
show(img, frames = 1, scale = TRUE, ...)
}
\arguments{
\item{frames}{frames}

\item{scale}{scale}
}
\description{
Adds functionality to view dicom images where each file may have more than 1 frame
}
\details{

}
