% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{show_batch}
\alias{show_batch}
\title{show_batch}
\usage{
show_batch(
  dls,
  b = NULL,
  max_n = 9,
  ctxs = NULL,
  show = TRUE,
  unique = FALSE,
  dpi = 90
)
}
\arguments{
\item{dls}{dataloader object}

\item{b}{b}

\item{max_n}{max_n}

\item{ctxs}{ctxs}

\item{show}{show}

\item{unique}{unique}
}
\description{

}
