% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_augment.R
\name{show_image}
\alias{show_image}
\title{show_image}
\usage{
show_image(
  im,
  ax = NULL,
  figsize = NULL,
  title = NULL,
  ctx = NULL,
  cmap = NULL,
  norm = NULL,
  aspect = NULL,
  interpolation = NULL,
  alpha = NULL,
  vmin = NULL,
  vmax = NULL,
  origin = NULL,
  extent = NULL
)
}
\arguments{
\item{im}{im}

\item{ax}{ax}

\item{figsize}{figsize}

\item{title}{title}

\item{ctx}{ctx}

\item{cmap}{cmap}

\item{norm}{norm}

\item{aspect}{aspect}

\item{interpolation}{interpolation}

\item{alpha}{alpha}

\item{vmin}{vmin}

\item{vmax}{vmax}

\item{origin}{origin}

\item{extent}{extent}
}
\description{
Show a PIL or PyTorch image on `ax`.
}
