% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{show_images}
\alias{show_images}
\title{Show_images}
\usage{
show_images(
  ims,
  nrows = 1,
  ncols = NULL,
  titles = NULL,
  figsize = NULL,
  imsize = 3,
  add_vert = 0
)
}
\arguments{
\item{ims}{images}

\item{nrows}{number of rows}

\item{ncols}{number of columns}

\item{titles}{titles}

\item{figsize}{figure size}

\item{imsize}{image size}

\item{add_vert}{add vertical}
}
\value{
None
}
\description{
Show all images `ims` as subplots with `rows` using `titles`
}
