% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{skm_to_fastai}
\alias{skm_to_fastai}
\title{Skm to fastai}
\usage{
skm_to_fastai(
  func,
  is_class = TRUE,
  thresh = NULL,
  axis = -1,
  activation = NULL,
  ...
)
}
\arguments{
\item{func}{function}

\item{is_class}{is classification or not}

\item{thresh}{threshold point}

\item{axis}{axis}

\item{activation}{activation}
}
\value{
None
}
\description{
Convert `func` from sklearn$metrics to a fastai metric
}
