% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_core.R
\name{SpacyTokenizer}
\alias{SpacyTokenizer}
\title{SpacyTokenizer}
\usage{
SpacyTokenizer(lang = "en", special_toks = NULL, buf_sz = 5000)
}
\arguments{
\item{lang}{lang}

\item{special_toks}{special_toks}

\item{buf_sz}{buf_sz}
}
\description{
Spacy tokenizer for `lang`
}
