% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{squeezenet1_0}
\alias{squeezenet1_0}
\title{squeezenet1_0}
\usage{
squeezenet1_0(pretrained = FALSE, progress = TRUE)
}
\arguments{
\item{pretrained}{pretrained}

\item{progress}{progress}
}
\description{
SqueezeNet model architecture from the `"SqueezeNet: AlexNet-level
}
\details{
accuracy with 50x fewer parameters and <0.5MB model size"
<https://arxiv.org/abs/1602.07360>`_ paper.
}
