% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_model.R
\name{tabular_config}
\alias{tabular_config}
\title{tabular_config}
\usage{
tabular_config(
  ps = NULL,
  embed_p = 0,
  y_range = NULL,
  use_bn = TRUE,
  bn_final = FALSE,
  bn_cont = TRUE,
  act_cls = nn$ReLU(inplace = TRUE)
)
}
\arguments{
\item{ps}{ps}

\item{embed_p}{embed_p}

\item{y_range}{y_range}

\item{use_bn}{use_bn}

\item{bn_final}{bn_final}

\item{bn_cont}{bn_cont}

\item{act_cls}{act_cls}
}
\description{
Convenience function to easily create a config for `TabularModel`
}
\details{

}
