% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{tabular_learner}
\alias{tabular_learner}
\title{Tabular learner}
\usage{
tabular_learner(
  data,
  layers,
  emb_szs = NULL,
  metrics = NULL,
  path = NULL,
  callback_fns = NULL,
  ps = NULL,
  emb_drop = 0,
  y_range = NULL,
  use_bn = TRUE,
  ...
)
}
\arguments{
\item{data}{data}

\item{layers}{layers}

\item{emb_szs}{emb_szs}

\item{metrics}{metrics}

\item{ps}{ps}

\item{emb_drop}{emb_drop}

\item{y_range}{y_range}

\item{use_bn}{use_bn}
}
\description{
Get a `Learner` using `data`, with `metrics`,
including a `TabularModel` created using the remaining params.
}
