% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{TabularDataTable}
\alias{TabularDataTable}
\title{TabularDataTable}
\usage{
TabularDataTable(
  df,
  procs = NULL,
  cat_names = NULL,
  cont_names = NULL,
  y_names = NULL,
  y_block = NULL,
  splits = NULL,
  do_setup = TRUE,
  device = NULL,
  inplace = FALSE,
  reduce_memory = TRUE
)
}
\arguments{
\item{df}{df}

\item{procs}{procs}

\item{cat_names}{cat_names}

\item{cont_names}{cont_names}

\item{y_names}{y_names}

\item{y_block}{y_block}

\item{splits}{splits}

\item{do_setup}{do_setup}

\item{device}{device}

\item{inplace}{inplace}

\item{reduce_memory}{reduce_memory}
}
\description{
A `Tabular` object with transforms
}
