% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_model.R
\name{TabularModel}
\alias{TabularModel}
\title{TabularModel}
\usage{
TabularModel(
  emb_szs,
  n_cont,
  out_sz,
  layers,
  ps = NULL,
  embed_p = 0,
  y_range = NULL,
  use_bn = TRUE,
  bn_final = FALSE,
  bn_cont = TRUE,
  act_cls = nn$ReLU(inplace = TRUE)
)
}
\arguments{
\item{emb_szs}{embedding size}

\item{n_cont}{number of cont}

\item{out_sz}{output size}

\item{layers}{layers}

\item{ps}{ps}

\item{embed_p}{embed proportion}

\item{y_range}{y range}

\item{use_bn}{use batch normalization}

\item{bn_final}{batch normalization final}

\item{bn_cont}{batch normalization cont}

\item{act_cls}{activation}
}
\value{
None
}
\description{
Basic model for tabular data.
}
