% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_core.R
\name{TensorBBox_create}
\alias{TensorBBox_create}
\title{create}
\usage{
TensorBBox_create(x, img_size = NULL)
}
\arguments{
\item{x}{x}

\item{img_size}{img_size}
}
\description{
Delegates (`__call__`,`decode`,`setup`) to (`encodes`,`decodes`,`setups`) if `split_idx` matches
}
