% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_learner.R
\name{TextLearner_save_encoder}
\alias{TextLearner_save_encoder}
\title{save_encoder}
\usage{
TextLearner_save_encoder(file)
}
\arguments{
\item{file}{file}
}
\description{
Save the encoder to `file` in the model directory
}
