% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_core.R
\name{tokenize_df}
\alias{tokenize_df}
\title{tokenize_df}
\usage{
tokenize_df(
  df,
  text_cols,
  n_workers = 6,
  rules = NULL,
  mark_fields = NULL,
  tok = NULL,
  res_col_name = "text"
)
}
\arguments{
\item{df}{df}

\item{text_cols}{text_cols}

\item{n_workers}{n_workers}

\item{rules}{rules}

\item{mark_fields}{mark_fields}

\item{tok}{tok}

\item{res_col_name}{res_col_name}
}
\description{
Tokenize texts in `df[text_cols]` in parallel using `n_workers`
}
\details{

}
