% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{top_k_accuracy}
\alias{top_k_accuracy}
\title{top_k_accuracy}
\usage{
top_k_accuracy(inp, targ, k = 5L, axis = -1)
}
\arguments{
\item{inp}{inp}

\item{targ}{targ}

\item{k}{k}

\item{axis}{axis}
}
\description{
Computes the Top-k accuracy (`targ` is in the top `k` predictions of `inp`)
}
\details{

}
