% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{top_k_accuracy}
\alias{top_k_accuracy}
\title{Top_k_accuracy}
\usage{
top_k_accuracy(inp, targ, k = 5, axis = -1)
}
\arguments{
\item{inp}{predictions}

\item{targ}{targets}

\item{k}{k}

\item{axis}{axis}
}
\value{
None
}
\description{
Computes the Top-k accuracy (`targ` is in the top `k` predictions of `inp`)
}
\examples{

\dontrun{

loaders = loaders()

data = Data_Loaders(loaders['train'], loaders['valid'])$cuda()

model = nn$Sequential() +
  nn$Flatten() +
  nn$Linear(28L * 28L, 10L)
metrics = list(accuracy,top_k_accuracy)
learn = Learner(data, model, loss_func = F$cross_entropy, opt_func = Adam,
                metrics = metrics)

}

}
