% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collab.R
\name{trainable_params}
\alias{trainable_params}
\title{Trainable_params}
\usage{
trainable_params(m)
}
\arguments{
\item{m}{trainable parameters}
}
\value{
None
}
\description{
Return all trainable parameters of `m`
}
