% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAN.R
\name{TransformBlock}
\alias{TransformBlock}
\title{TransformBlock}
\usage{
TransformBlock(
  type_tfms = NULL,
  item_tfms = NULL,
  batch_tfms = NULL,
  dl_type = NULL,
  dls_kwargs = NULL
)
}
\arguments{
\item{type_tfms}{type_tfms}

\item{item_tfms}{item_tfms}

\item{batch_tfms}{batch_tfms}

\item{dl_type}{dl_type}

\item{dls_kwargs}{dls_kwargs}
}
\description{
A basic wrapper that links defaults transforms for the data block API
}
