% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet.R
\name{unet_config}
\alias{unet_config}
\title{Unet_config}
\usage{
unet_config(
  blur = FALSE,
  blur_final = TRUE,
  self_attention = FALSE,
  y_range = NULL,
  last_cross = TRUE,
  bottle = FALSE,
  act_cls = nn$ReLU,
  init = nn$init$kaiming_normal_,
  norm_type = NULL
)
}
\arguments{
\item{blur}{blur}

\item{blur_final}{blur_final}

\item{self_attention}{self_attention}

\item{y_range}{y_range}

\item{last_cross}{last_cross}

\item{bottle}{bottle}

\item{act_cls}{act_cls}

\item{init}{init}

\item{norm_type}{norm_type}
}
\description{
Convenience function to easily create a config for `DynamicUnet`
}
