% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet.R, R/vision_models.R
\name{UnetBlock}
\alias{UnetBlock}
\title{UnetBlock}
\usage{
UnetBlock(...)

UnetBlock(...)

UnetBlock(...)
}
\arguments{
\item{up_in_c}{up_in_c}

\item{x_in_c}{x_in_c}

\item{hook}{hook}

\item{final_div}{final_div}

\item{blur}{blur}

\item{act_cls}{act_cls}

\item{self_attention}{self_attention}

\item{init}{init}

\item{norm_type}{norm_type}

\item{ks}{ks}

\item{stride}{stride}

\item{padding}{padding}

\item{bias}{bias}

\item{ndim}{ndim}

\item{bn_1st}{bn_1st}

\item{transpose}{transpose}

\item{xtra}{xtra}

\item{bias_std}{bias_std}

\item{dilation}{dilation}

\item{groups}{groups}

\item{padding_mode}{padding_mode}
}
\description{
A quasi-UNet block, using `PixelShuffle_ICNR upsampling`.

A quasi-UNet block, using `PixelShuffle_ICNR upsampling`.

Load model architecture
}
