% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tar.R
\name{URLs_IMAGEWOOF}
\alias{URLs_IMAGEWOOF}
\title{IMAGEWOOF dataset}
\usage{
URLs_IMAGEWOOF(filename = "IMAGEWOOF", untar = TRUE)
}
\arguments{
\item{filename}{the name of the file}

\item{untar}{logical, whether to untar the '.tgz' file}
}
\description{
download IMAGEWOOF dataset
}
