% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_core_config.R
\name{Voice}
\alias{Voice}
\title{Voice}
\usage{
Voice(
  sample_rate = 16000,
  n_fft = 1024,
  win_length = NULL,
  hop_length = 128,
  f_min = 50,
  f_max = 8000,
  pad = 0,
  n_mels = 128,
  window_fn = torch()$hann_window,
  power = 2,
  normalized = FALSE,
  wkwargs = NULL,
  mel = TRUE,
  to_db = TRUE
)
}
\arguments{
\item{sample_rate}{sample rate}

\item{n_fft}{number of fast fourier transforms}

\item{win_length}{windowing length}

\item{hop_length}{hopping length}

\item{f_min}{minimum frequency}

\item{f_max}{maximum frequency}

\item{pad}{padding mode}

\item{n_mels}{number of mel-spectrograms}

\item{window_fn}{window function}

\item{power}{power}

\item{normalized}{normalized or not}

\item{wkwargs}{additional arguments}

\item{mel}{mel-spectrogram or not}

\item{to_db}{to decibels}
}
\value{
None
}
\description{
Voice
}
