% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{weight_decay}
\alias{weight_decay}
\title{weight_decay}
\usage{
weight_decay(p, lr, wd, do_wd = TRUE, ...)
}
\arguments{
\item{p}{p}

\item{lr}{lr}

\item{wd}{wd}

\item{do_wd}{do_wd}

\item{...}{additional args to pass}
}
\description{
Weight decay as decaying `p` with `lr*wd`
}
