% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_models.R
\name{XResNet}
\alias{XResNet}
\title{XResNet}
\usage{
XResNet(block, expansion, layers, c_in = 3, c_out = 1000, ...)
}
\arguments{
\item{block}{the blocks to pass to XResNet}

\item{expansion}{argument for inputs and filters}

\item{layers}{the layers to pass to XResNet}

\item{c_in}{number of inputs}

\item{c_out}{number of outputs}

\item{...}{additional arguments}
}
\description{
A sequential container.
}
